### 2022-01-08 #4042 @whedon
[PRE REVIEW]: HylleraasMD: Massively parallel hybrid particle-field molecular dynamics in python
**Submitting author:** @mortele (<a href="http://orcid.org/0000-0003-4244-4876">Morten Ledum</a>)
**Repository:** <a href="https://github.com/Cascella-Group-UiO/HyMD" target ="_blank">https://github.com/Cascella-Group-UiO/HyMD</a>
**Version:** v1.0.3
