### 2022-01-08 #4042 @danielskatz
👋 @jgostick - would you be able to edit this submission for JOSS?
