### 2022-01-08 120111c
Change python to Python in JOSS paper
