### 2022-01-10 1baa5c1
Fix crash caused by missing thermostat specification
Properly tests provided start_temperature and target_temperature values
ensuring they are numbers before checking validity. Fails in a sane way
if this is not true.
