### 2022-01-19 8e3b152
adding GPE code
PS: Remaining code to be added in inputparser
PS PS: Run and check everything ok.
+ need to add documentation
