+++ hymd/main.py
@@ -9,12 +9,13 @@ import pmesh.pm as pmesh
 import warnings
 from .configure_runtime import configure_runtime
 from .hamiltonian import DefaultNoChi, DefaultWithChi, SquaredPhi
-from .input_parser import check_config
+from .input_parser import check_config, sort_dielectric_by_type_id,
 from .logger import Logger, format_timedelta
 from .file_io import distribute_input, OutDataset, store_static, store_data
 from .field import (compute_field_force, update_field,
                     compute_field_and_kinetic_energy, domain_decomposition,
-                    update_field_force_q, compute_field_energy_q,)
+                    update_field_force_q, compute_field_energy_q,
+                    update_field_force_q_GPE, compute_field_energy_q_GPE)
 from .thermostat import (csvr_thermostat, cancel_com_momentum,
                          generate_initial_velocities)
 from .force import dipole_forces_redistribution, prepare_bonds
@@ -96,6 +97,38 @@ def main():
             charges_flag = False
 
     config = check_config(config, indices, names, types, comm=comm)
+
+    ## dielectric from toml
+    if config.coulombtype == 'PIC_Spectral_GPE':
+        dielectric_sorted = sort_dielectric_by_type_id(config,charges,types)
+        ###^----- only needed if dielectric given from toml.
+        dielectric_flag = True
+
+        if config.convergence_type is not None:
+            if config.convergence_type == 'csum':
+                def conv_fun(comm,diff_mesh):
+                    return diff_mesh.csum()
+
+            elif config.convergence_type == 'euclidean_norm':
+                def conv_fun(comm,diff_mesh):
+                    return diff_mesh.cnorm()
+
+            elif config.convergence_type == 'max_diff':
+                def conv_fun(comm,diffmesh):
+                    msg = np.max(diffmesh)
+                    res = comm.allreduce(sendobj=msg, op=MPI.MAX)
+                    return res
+
+        else: ## default choice is the max difference if not specified
+            config.convergence_type = 'max_diff'
+            def conv_fun(comm,diffmesh):
+                msg = np.max(diffmesh)
+                res = comm.allreduce(sendobj=msg, op=MPI.MAX)
+                return res
+
+    else:
+        dielectric_flag = False
+
     if config.n_print:
         if config.n_flush is None:
             config.n_flush = 10000 // config.n_print
@@ -191,6 +224,30 @@ def main():
             "complex", value=0.0
         )
 
+    if dielectric_flag  and config.coulombtype == 'PIC_Spectral_GPE': ## initializing the density mesh
+        phi_q = pm.create("real", value=0.0)
+        phi_q_fourier = pm.create("complex", value=0.0)
+        elec_field_fourier= [pm.create("complex", value=0.0) for _ in range(_SPACE_DIM)] #for force calculation
+        elec_field = [pm.create("real", value=0.0) for _ in range(_SPACE_DIM)] #for force calculation
+        elec_energy_field = pm.create("complex", value=0.0) # for energy calculation --> complex form needed as its converted from complex field; Imaginary part as zero;
+
+        ## GPE relevant
+        phi_q_eps = pm.create("real", value = 0.0) ## real contrib of non-polarization part of GPE
+        phi_q_eps_fourier = pm.create("complex", value = 0.0) # complex contrib of phi q eps
+        phi_q_effective_fourier = pm.create("complex", value = 0.0) ## fourier of non-polarization part of GPE
+        phi_eps = pm.create("real", value = 0.0) ## real contrib of the epsilon dielectric painted to grid
+        phi_eps_fourier = pm.create("complex", value = 0.0) # complex contrib of phi eps
+        phi_eta = [pm.create("real", value = 0.0)for _ in range(_SPACE_DIM)] ## real contrib of factor in polarization charge density
[…]
