+++ hymd/field.py
@@ -612,6 +612,255 @@ def compute_field_and_kinetic_energy(
     kinetic_energy = comm.allreduce(0.5 * config.mass * np.sum(velocity ** 2))
     return field_energy, kinetic_energy
 
+def compute_field_energy_q_GPE(
+    config,phi_eps_fourier,elec_field_contrib,phi_q_effective_fourier,
+    phi_q_fourier,elec_energy_field, field_q_energy, comm=MPI.COMM_WORLD,
+):
+    """
+    - added for general poisson equation (GPE)
+    - used phi_q_effective_fourier (from iterative method) to calculate E_field, thus needed
+    - multiply with actual charge density
+    - PS: One can use the potential directly --> U = g*elec_potential
+    """
+
+    #COULK_GMX = 138.935458 # 1/(4pi eps0) config.dielectric_const
+
+    intermediate = np.abs(phi_q_fourier)*np.abs(phi_q_effective_fourier)
+    ## ^ ----- For some reason it won't do this inside the transfer function
+    ### even though they have the same shape
+
+    def transfer_energy(k,v):  ### potential field is electric field / (-ik)  --> potential field * q -->
+        return 4.0 * np.pi * config.coulomb_constant * v / k.normp(p=2,zeromode=1)
+
+    intermediate.apply(transfer_energy,  kind='wavenumber', out=elec_energy_field)
+
+    V = np.prod(config.box_size)
+
+    field_q_energy = 0.5 * V * comm.allreduce(np.sum(elec_energy_field.value))
+
+    ### Add contribution from polarization
+    #pol_energy = -0.5*(dielectric-phi_eps.readout(positions, layout=layout_q))*(elec_field_contrib.readout(positions, layout=layout_q))
+    #pol_energy = comm.allreduce(np.sum(pol_energy))
+    ### ^---- Alternative 1: From V_ext
+    #print(pol_energy)
+
+    pol_energy = 0.0
+    eps_check = 1e-5 # to avoid invalid value in divide
+    if eps_check < np.abs(np.min(phi_eps_fourier)):
+        def transfer_pol_energy(k,v,epsk = phi_eps_fourier):
+            return np.abs(v)**2  / (k.normp(p=2,zeromode=1))
+        ##  ^----- Alternative 2: From its functional form W
+        phi_q_fourier.apply(transfer_pol_energy,out = elec_energy_field)
+        elec_energy_field = elec_energy_field/phi_eps_fourier
+        elec_energy_field.c2r
+        pol_energy = 0.5 * V * comm.allreduce(np.sum(elec_energy_field.value))
+    ### ^---- Alternative 2: From W_{ext}
+
+    return field_q_energy.real + pol_energy.real
+
+def update_field_force_q_GPE(conv_fun,phi, types, charges, dielectric, phi_q,
+    phi_q_fourier,phi_eps, phi_eps_fourier,phi_q_eps, phi_q_eps_fourier,
+    phi_q_effective_fourier,phi_eta,  phi_eta_fourier, phi_pol_prev,
+    phi_pol, phi_pol_fourier,sum_fourier,phi_pol_temp,elec_field_fourier,
+    elec_field,elec_forces,elec_field_contrib, layout_q, layout,pm,positions,
+    config,comm = MPI.COMM_WORLD, compute_potential = False,
+    E_field_from_potential = False,
+):
+    """
+    - added for the general poisson equation (GPE) eletrostatics (follow PIC_Spectral_GPE)
+    - this function get the electrostatic forces
+    - refering to the test-pure-sphere-new.py, this inlcudes:
+        [O] hpf_init_simple_gmx_units(grid_num,box_size,coords,charges,masses)
+        ## ^----- define the pm and layout, already out outside this fucntion
[…]
