+++ hymd/main.py
@@ -9,7 +9,7 @@ import pmesh.pm as pmesh
 import warnings
 from .configure_runtime import configure_runtime
 from .hamiltonian import DefaultNoChi, DefaultWithChi, SquaredPhi
-from .input_parser import check_config, sort_dielectric_by_type_id,
+from .input_parser import check_config, sort_dielectric_by_type_id
 from .logger import Logger, format_timedelta
 from .file_io import distribute_input, OutDataset, store_static, store_data
 from .field import (compute_field_force, update_field,
