### 2022-02-02 32e9728
Set default verbose to 1 (#154)
