### 2022-02-08 92174e3
Update main.py
added missing param and append args_in for variable dielectric. Now runs
