### 2022-02-10 #4042 @yhtang
> 👋 @yhtang, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

@rkurchin Happy to help!
