### 2022-02-10 #4149 @whedon
[REVIEW]: HylleraasMD: Massively parallel hybrid particle-field molecular dynamics in python
**Submitting author:** @mortele (<a href="http://orcid.org/0000-0003-4244-4876">Morten Ledum</a>)
**Repository:** https://github.com/Cascella-Group-UiO/HyMD
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.9

- [x] #5308
