### 2022-02-10 #4149 @yhtang
I found another article on ChemRxiv titled [HylleraasMD: A Domain Decomposition-Based Hybrid Particle-Field Software for Multi-Scale Simulations of Soft Matter](https://doi.org/10.26434/chemrxiv-2021-796ql) from the same group of authors.

It seems that both papers focus on the HylleraasMD software and the ChenRxiv one is more elaborated. Is it meant for a more traditional platform?

Could the authors (@mortele) please clarify this, especially if the ChemRxiv version is still under review somewhere?
