### 2022-02-10 Cascella-Group-UiO/HyMD#157 @yhtang
Install non-Python dependencies with `dnf`
Hi there,

This is related to the 'installation' checklist item of the JOSS review:
https://github.com/openjournals/joss-reviews/issues/4149.

Could you please also provide the external package names under the DNF/RPM system? This could be very helpful for users of Linux distros such as [Fedora](https://getfedora.org/), [CentOS](https://www.centos.org/), and [Scientific Linux](https://scientificlinux.org/), etc.

Many thanks and please let me know your thoughts!
