### 2022-02-10 Cascella-Group-UiO/HyMD#158 @yhtang
Documentation badge
Another question related to the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4149):

The documentation badge currently points to this repo's Actions page. Perhaps changing it to the actual online docs page would be more useful for potential users?
