### 2022-02-10 Cascella-Group-UiO/HyMD#158 @mortele
Hi @yhtang, thank you for pointing this out. I agree that it would be much more useful for this link to point to the actual documentation.
