### 2022-02-10 d249209
Make the README.md docs badge link to the docs
As opposed to linking to the docs github actions page.
