### 2022-02-10 Cascella-Group-UiO/HyMD#157 @mortele
Hi @yhtang, thank you for the suggestion. I have briefly looked into making the package installable (wholesale with all the dependencies packaged) with some of the standard package managers like anaconda and spack. This is on the todo-list but has been pushed back in favor of more pressing matters at the moment.

In all honesty, In my own natural Debian bias I had not even considered making a `yum` package, but I will investigate how to make it RedHat installable when I get some spare time soon.
