### 2022-02-10 Cascella-Group-UiO/HyMD#157 @yhtang
Oh, I guess there is some misunderstanding here. I am not asking for packaging the software as an rpm package. Rather, I was just wondering what would be the DNF counterparts for the following installation steps from READMD:
```
sudo apt-get update -y
sudo apt-get install -y pkg-config libhdf5-mpi-dev
python3 -m pip uninstall h5py  # Remove any serial h5py installation present
CC="mpicc" HDF5_MPI="ON" python3 -m pip install --no-binary=h5py h5py
```
