### 2022-02-10 Cascella-Group-UiO/HyMD#157 @mortele
@yhtang Aha, ok, my mistake! The following builds the non-Python dependencies in a `fedora:35` docker container and so essentially answers the question:

```bash
dnf install openmpi-devel hdf5-openmpi-devel python3-devel python3-pip
export CC=/usr/lib64/openmpi/bin/mpicc
python3 -m pip install cython numpy mpi4py
```

I will have to work out tomorrow how to properly build `h5py-mpi`, which seems to be a small issue here, and add it to the docs as a separate installation guide section.
