### 2022-02-11 Cascella-Group-UiO/HyMD#157 @yhtang
> Seems like the default Python 3.10.2 in fedora:latest does not play well with h5py, so falling back to Python 3.8.12.

I wonder if a virtual environment-based recipe could make life slightly easier? :grin:
