### 2022-02-11 #4149 @yhtang
I noticed that the automated test coverage is around 50%. Should this be considered something that needs improvement? @rkurchin @blakeaw
