### 2022-02-11 Cascella-Group-UiO/HyMD#159 @yhtang
Community guidelines
Hi Morten,

Another question here regarding the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4149). It seems like the online documentation page does not include a page on guidelines for community contributions. Specifically, this pertains to the question:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Could you please point me to this page if I happened to miss it, or maybe add something along the line if it does not exist yet? Thanks!
