### 2022-02-11 #4149 @yhtang
I am now stuck at the functionality and performance check due to a difficulty in installing `h5py` in Fedora as already documented in https://github.com/Cascella-Group-UiO/HyMD/issues/157. I will resume as soon as a workaround has been found.
