### 2022-02-14 #4149 @rkurchin
I would agree with @blakeaw's perspective – JOSS doesn't include specific coverage requirements for this reason. That being said, @yhtang, if you see lack of coverage in areas crucial to functionality, that could certainly be an argument for suggesting additional tests!
