### 2022-02-14 Cascella-Group-UiO/HyMD#160 @blakeaw
Error when running tests
Hi,

This issue is part of the JOSS Review at https://github.com/openjournals/joss-reviews/issues/4149

When I run the standard automated tests with pytest I get some errors:
     test_distribute_input.py EEEE

When I run `pytest test_distribute_input.py` it gives more details. Here is the summary:
```
=========================================================================== short test summary info ===========================================================================
ERROR test_distribute_input.py::test_file_io_distribute_input_size_1
ERROR test_distribute_input.py::test_file_io_distribute_input_size_2
ERROR test_distribute_input.py::test_file_io_distribute_input_size_3
ERROR test_distribute_input.py::test_file_io_distribute_input_various_sizes
```

And the recurring error at all 4 locations is this:
```
  @pytest.fixture()
  def h5py_molecules_file(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.
```

Did I miss something? Or should I just be running the pytest-mpi script instead?
