### 2022-02-14 #4149 @blakeaw
Ran into a problem/errors when running automated tests: https://github.com/Cascella-Group-UiO/HyMD/issues/160
