### 2022-02-14 Cascella-Group-UiO/HyMD#160 @blakeaw
It looks like I'm also getting some errors when I run the MPI enabled tests `./pytest-mpi -oo -n 2 -ns`:

```
./pytest-mpi: line 67: [: : integer expression expected
./pytest-mpi: line 71: [: : integer expression expected
ERROR: usage: pytest [options] [file_or_dir] [file_or_dir] [...]
pytest: error: unrecognized arguments: --only-mpi
  inifile: None
  rootdir: joss_review/hylleraasmd/hymd

ERROR: usage: pytest [options] [file_or_dir] [file_or_dir] [...]
pytest: error: unrecognized arguments: --only-mpi
  inifile: None
  rootdir: joss_review/hylleraasmd/hymd

--------------------------------------------------------------------------
Primary job  terminated normally, but 1 process returned
a non-zero exit code. Per user-direction, the job has been aborted.
--------------------------------------------------------------------------
--------------------------------------------------------------------------
mpirun detected that one or more processes exited with non-zero status, thus causing
the job to be terminated. The first process to do so was:

  Process name: [[50091,1],1]
  Exit code:    4
--------------------------------------------------------------------------

################################################################################
################################# RANK 0 / 2 ###################################
################################################################################

################################################################################
################################# RANK 1 / 2 ###################################
################################################################################
```
