### 2022-02-14 Cascella-Group-UiO/HyMD#161 @blakeaw
Supplemental instructions to install MPI compiler
Hi,

This issue is part of the JOSS Review at https://github.com/openjournals/joss-reviews/issues/4149

Since a MPI compiler is required for installation it might be good to include supplemental instructions for installing the compiler in case users don't already have one installed on their system.  For example, I installed hymd in a WSL2 Ubuntu 20.04 instance which didn't have MPI compilers already. I ended up using the following for mpich:
```
sudo apt install mpich libmpich-dev
```
And then installing mpi4py:
```
CC=mpicc pip install mpi4py
```

This comment may have some overlap with Issue https://github.com/Cascella-Group-UiO/HyMD/issues/157.
