### 2022-02-15 Cascella-Group-UiO/HyMD#160 @blakeaw
Hi,

I went ahead and tried reinstalling everything in a fresh conda environment. I uninstalled mpich and instead used openmpi `apt install openmpi-bin libopenmpi-dev` this time so that mpi4py and h5py were compiled with openmpi (setting `CC=mpicc.openmpi`).

Now when I run standard test set I still get many errors (`E`) but I also get one failure in test_force. It looks like many of the Es are the same `E       fixture 'mpi_file_name' not found` as before. It looks like the failure in test_force is from `test/test_force.py:99: AssertionError`. The full pytest output is below:

```
========================================================================= test session starts ==========================================================================
rootdir: joss_review/hylleraasmd/hymd
collected 42 items

test/test_distribute_input.py EEEE                                                                                                                               [  9%]
test/test_field.py .                                                                                                                                             [ 11%]
test/test_force.py .F....                                                                                                                                        [ 26%]
test/test_hamiltonian.py EEE......                                                                                                                               [ 47%]
test/test_input_parser.py EEEEEEEEEEEE                                                                                                                           [ 76%]
test/test_integrator.py .........                                                                                                                                [ 97%]
test/test_thermostat.py .                                                                                                                                        [100%]

================================================================================ ERRORS ================================================================================________________________________________________________ ERROR at setup of test_file_io_distribute_input_size_1 ________________________________________________________file joss_review/hylleraasmd/hymd/test/test_distribute_input.py, line 17
  def test_file_io_distribute_input_size_1(h5py_molecules_file):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 202
  @pytest.fixture()
  def h5py_molecules_file(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:202
________________________________________________________ ERROR at setup of test_file_io_distribute_input_size_2 ________________________________________________________file joss_review/hylleraasmd/hymd/test/test_distribute_input.py, line 28
  def test_file_io_distribute_input_size_2(h5py_molecules_file):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 202
  @pytest.fixture()
  def h5py_molecules_file(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:202
________________________________________________________ ERROR at setup of test_file_io_distribute_input_size_3 ________________________________________________________file joss_review/hylleraasmd/hymd/test/test_distribute_input.py, line 49
  def test_file_io_distribute_input_size_3(h5py_molecules_file):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 202
  @pytest.fixture()
  def h5py_molecules_file(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:202
____________________________________________________ ERROR at setup of test_file_io_distribute_input_various_sizes _____________________________________________________file joss_review/hylleraasmd/hymd/test/test_distribute_input.py, line 74
  def test_file_io_distribute_input_various_sizes(h5py_molecules_file):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 202
  @pytest.fixture()
  def h5py_molecules_file(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:202
_______________________________________________ ERROR at setup of test_DefaultNoChi_window_function[0.2988365823859701] ________________________________________________file joss_review/hylleraasmd/hymd/test/test_hamiltonian.py, line 17
  @pytest.mark.parametrize(
      "sigma",
      [
          0.2988365823859701,
          1.2585762493242553,
          9.2159828579248931,
      ],
  )
  def test_DefaultNoChi_window_function(
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
_______________________________________________ ERROR at setup of test_DefaultNoChi_window_function[1.2585762493242554] ________________________________________________file joss_review/hylleraasmd/hymd/test/test_hamiltonian.py, line 17
  @pytest.mark.parametrize(
      "sigma",
      [
          0.2988365823859701,
          1.2585762493242553,
          9.2159828579248931,
      ],
  )
  def test_DefaultNoChi_window_function(
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
________________________________________________ ERROR at setup of test_DefaultNoChi_window_function[9.215982857924892] ________________________________________________file joss_review/hylleraasmd/hymd/test/test_hamiltonian.py, line 17
  @pytest.mark.parametrize(
      "sigma",
      [
          0.2988365823859701,
          1.2585762493242553,
          9.2159828579248931,
      ],
  )
  def test_DefaultNoChi_window_function(
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
_________________________________________________________ ERROR at setup of test_input_parser_read_config_toml _________________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 25
  def test_input_parser_read_config_toml(config_toml):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
_______________________________________________________________ ERROR at setup of test_input_parser_file _______________________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 31
  def test_input_parser_file(config_toml):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
______________________________________________________ ERROR at setup of test_input_parser_file_check_n_particles ______________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 38
  @pytest.mark.mpi()
  def test_input_parser_file_check_n_particles(config_toml, caplog):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
_________________________________________________________ ERROR at setup of test_input_parser_check_optionals __________________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 73
  @pytest.mark.mpi()
  def test_input_parser_check_optionals(config_toml, caplog):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
___________________________________________________________ ERROR at setup of test_input_parser_check_bonds ____________________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 157
  @pytest.mark.mpi()
  def test_input_parser_check_bonds(config_toml, dppc_single, caplog):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
___________________________________________________________ ERROR at setup of test_input_parser_check_angles ___________________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 203
  @pytest.mark.mpi()
  def test_input_parser_check_angles(config_toml, dppc_single, caplog):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
____________________________________________________________ ERROR at setup of test_input_parser_check_chi _____________________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 251
  @pytest.mark.mpi()
  def test_input_parser_check_chi(config_toml, dppc_single, caplog):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
__________________________________________________________ ERROR at setup of test_input_parser_check_box_size __________________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 334
  def test_input_parser_check_box_size(config_toml, caplog):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
_________________________________________________________ ERROR at setup of test_input_parser_check_integrator _________________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 354
  def test_input_parser_check_integrator(config_toml, caplog):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
____________________________________________________ ERROR at setup of test_input_parser_thermostat_coupling_groups ____________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 376
  def test_input_parser_thermostat_coupling_groups(config_toml, caplog):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

joss_review/hylleraasmd/hymd/test/conftest.py:237
____________________________________________________ ERROR at setup of test_input_parser_check_cancel_com_momentum _____________________________________________________file joss_review/hylleraasmd/hymd/test/test_input_parser.py, line 459
  def test_input_parser_check_cancel_com_momentum(config_toml, caplog):
file joss_review/hylleraasmd/hymd/test/conftest.py, line 237
  @pytest.fixture()
  def config_toml(mpi_file_name):
E       fixture 'mpi_file_name' not found
>       available fixtures: alanine_octapeptide, cache, capfd, capfdbinary, caplog, capsys, capsysbinary, config_CONF, config_toml, doctest_namespace, dppc_single, filter, h5py_molecules_file, molecules_with_solvent, monkeypatch, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, three_atoms, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory, v_ext
>       use 'pytest --fixtures [testpath]' for help on them.

[…]
