### 2022-02-16 Cascella-Group-UiO/HyMD#160 @mortele
Hi @blakeaw, I believe the main issue the parallel tests tagged with
```python
@pytest.mark.mpi()
```
fail outright if `pytest-mpi` is not installed. With this package available, the MPI tests are automatically skipped unless you flag `--only-mpi` to pytest. My bad for not adding this pytest extra package as a test dependency. I will add it to the test docs as soon as I get a chance.

As for the `test_force` failure, I am not sure what is going on there. Both locally on my machine and in the `mortele/hymd:latest` docker image, the force tests succeed:
```bash
docker pull mortele/hymd
docker run -it mortele/hymd:latest
root@c7de44a7d5d1:/app# apt-get update -y
root@c7de44a7d5d1:/app# apt-get install git -y
root@c7de44a7d5d1:/app# pip3 install hymd
root@c7de44a7d5d1:/app# git clone https://github.com/Cascella-Group-UiO/HyMD.git
root@c7de44a7d5d1:/app# cd HyMD
root@c7de44a7d5d1:/app# pip3 install pytest
root@c7de44a7d5d1:/app# pytest
==================================================================================================== test session starts =====================================================================================================
rootdir: /app
collected 42 items

HyMD/test/test_distribute_input.py EEEE                                                                                                                                                                                [  9%]
HyMD/test/test_field.py .                                                                                                                                                                                              [ 11%]
HyMD/test/test_force.py ......                                                                                                                                                                                         [ 26%]
HyMD/test/test_hamiltonian.py EEE......                                                                                                                                                                                [ 47%]
HyMD/test/test_input_parser.py EEEEEEEEEEEE                                                                                                                                                                            [ 76%]
HyMD/test/test_integrator.py .........                                                                                                                                                                                 [ 97%]
HyMD/test/test_thermostat.py .                                                                                                                                                                                         [100%]

=========================================================================================================== ERRORS ===========================================================================================================

   (...)

========================================================================================= 23 passed, 7 warnings, 19 errors in 7.15s ==========================================================================================
root@c7de44a7d5d1:/app# pip3 install pytest-mpi
root@c7de44a7d5d1:/app# pytest
==================================================================================================== test session starts =====================================================================================================
rootdir: /app
plugins: mpi-0.6
collected 42 items

HyMD/test/test_distribute_input.py ....                                                                                                                                                                                [  9%]
HyMD/test/test_field.py s                                                                                                                                                                                              [ 11%]
HyMD/test/test_force.py ......                                                                                                                                                                                         [ 26%]
HyMD/test/test_hamiltonian.py .........                                                                                                                                                                                [ 47%]
HyMD/test/test_input_parser.py ..sssss.....                                                                                                                                                                            [ 76%]
HyMD/test/test_integrator.py .........                                                                                                                                                                                 [ 97%]
HyMD/test/test_thermostat.py s                                                                                                                                                                                         [100%]

=============================================================================================== 35 passed, 7 skipped in 6.39s ================================================================================================
root@c7de44a7d5d1:/app# mpirun --allow-run-as-root -n 6 pytest --only-mpi

   (...)

================== 7 passed, 35 skipped, 22 warnings in 0.86s ==================
```

I will have to look into the force issue as soon as I get a chance.
