### 2022-02-16 Cascella-Group-UiO/HyMD#159 @mortele
Hi @yhtang, thanks for pointing this out. Right now we have not defined any such guidelines, but we enthusiastically encourage and welcome any contributions from external contributors as will get to this ASAP.
