### 2022-02-16 9df6a53
Update main.py
took away unnecessary if test
