### 2022-02-17 860a0c6
Add CONTRIBUTING.md draft
