### 2022-02-17 cab92ed
main + field update
Main:
-comment out prints again
field:
- changes to energy calculations
- readout changed s.t. sum of forces for each dim component closer to zero
