### 2022-02-18 28f3d25
main + field
safe divide seemingly fixed in elec energies
+ took away uncessary objects after trying out energy terms.
