+++ hymd/main.py
@@ -230,6 +230,7 @@ def main():
         elec_field_fourier= [pm.create("complex", value=0.0) for _ in range(_SPACE_DIM)] #for force calculation
         elec_field = [pm.create("real", value=0.0) for _ in range(_SPACE_DIM)] #for force calculation
         elec_energy_field = pm.create("complex", value=0.0) # for energy calculation --> complex form needed as its converted from complex field; Imaginary part as zero;
+        elec_energy_field_real = pm.create("real", value=0.0)
 
         ## GPE relevant
         phi_q_eps = pm.create("real", value = 0.0) ## real contrib of non-polarization part of GPE
@@ -322,7 +323,7 @@ def main():
     if charges_flag:
         layout_q = pm.decompose(positions)
         if config.coulombtype == 'PIC_Spectral_GPE': #dielectric_flag
-            elec_field_contrib, phi_eps_fourier = update_field_force_q_GPE(
+            phi_eps_fourier, phi_q_fourier = update_field_force_q_GPE(
                 conv_fun, phi, types, charges,dielectric_sorted,
                 phi_q, phi_q_fourier, phi_eps, phi_eps_fourier,phi_q_eps,
                 phi_q_eps_fourier,phi_q_effective_fourier, phi_eta,
@@ -333,8 +334,7 @@ def main():
             )
 
             field_q_energy =compute_field_energy_q_GPE(
-                config, phi_eps_fourier, elec_field_contrib,
-                phi_q_effective_fourier, phi_q_fourier,elec_energy_field,
+                config, phi_eps_fourier, phi_q_fourier,elec_energy_field,
                 field_q_energy,
                 comm=comm
             )
@@ -665,7 +665,7 @@ def main():
             if charges_flag:
                 layout_q = pm.decompose(positions)
                 if config.coulombtype == "PIC_Spectral_GPE": # dielectric_flag and
-                    elec_field_contrib, phi_eps_fourier = update_field_force_q_GPE(
+                    phi_eps_fourier, phi_q_fourier = update_field_force_q_GPE(
                         conv_fun, phi, types, charges,dielectric_sorted,
                         phi_q, phi_q_fourier, phi_eps, phi_eps_fourier,phi_q_eps,
                         phi_q_eps_fourier,phi_q_effective_fourier, phi_eta,
@@ -676,8 +676,7 @@ def main():
                     )
 
                     field_q_energy = compute_field_energy_q_GPE(
-                        config, phi_eps_fourier, elec_field_contrib,
-                        phi_q_effective_fourier, phi_q_fourier,elec_energy_field,
+                        config, phi_eps_fourier, phi_q_fourier,elec_energy_field,
                         field_q_energy,
                         comm=comm
                     )
@@ -875,8 +874,7 @@ def main():
                     if charges_flag:
                         if config.coulombtype == "PIC_Spectral_GPE":
                             field_q_energy = compute_field_energy_q_GPE(
-                                config, phi_eps_fourier, elec_field_contrib,
-                                phi_q_effective_fourier, phi_q_fourier,elec_energy_field,
+                                config, phi_eps_fourier, phi_q_fourier,elec_energy_field,
                                 field_q_energy,
                                 comm=comm
                             )
@@ -943,7 +941,7 @@ def main():
             if charges_flag:
                 layout_q = pm.decompose(positions)
                 if config.coulombtype == "PIC_Spectral_GPE":
-                    elec_field_contrib, phi_eps_fourier = update_field_force_q_GPE(
+                    phi_eps_fourier, phi_q_fourier = update_field_force_q_GPE(
                         conv_fun, phi, types, charges,dielectric_sorted,
[…]
