### 2022-02-18 afe4c77
energy elec GPE now only pos
field + main changed
- now only positive energy terms from electrostatics
