### 2022-02-21 Cascella-Group-UiO/HyMD#160 @blakeaw
Ah, ok. Yeah, the mpi_file_name fixture errors were resolved when I installed pytest-mpi. Thanks! As you already deduced I didn't realize pytest-mpi was a separate package needed for the tests.

I'll try installing and running again on a different machine to see if I still get the failure in test_force there as well.
