### 2022-02-21 Cascella-Group-UiO/HyMD#161 @blakeaw
Ok, I see. Thanks!
I do think maybe explicit is still better here since the mpi-enabled hdf5 dependency is listed as optional. If you follow the pip install instructions that are listed first without doing the mpi-enabled hdf5 step (as I did initially) but you don't already have the mpi compiler the install will fail when it tries to build mpi4py.

I'd also maybe recommend listing the dependencies first by moving the __Install dependencies__ subsection to the front of the **Installation** section of the README (and similarly in the documentation). That way users see the dependencies (in particular, the non-python ones) and can make sure they are installed first before they try to install hymd itself.  Also, if you rearrange you could probably move this line `python3 -m pip install --upgrade numpy mpi4py cython` into the dependencies section (and note that those are core dependencies).
