+++ hymd/field.py
@@ -669,8 +669,8 @@ def update_field_force_q_GPE(conv_fun,phi, types, charges, dielectric, phi_q,
     ## ^------ tbr; phi_transfer_funciton by hamiltonian.H ??
     phi_q_fourier.c2r(out=phi_q) ## this phi_q is after applying the smearing function
 
-    denom_phi_tot =  pm.create("real", value=0.0)
-    num_types =  pm.create("real", value=0.0)
+    denom_phi_tot =  pm.create("real", value= 0.0)
+    num_types =  pm.create("real", value = 0.0)
 
     #print("rank {:d} and phi shape {:}".format(comm.Get_rank(), np.shape(phi)))
     #for t_ in range(config.n_types):
@@ -680,15 +680,22 @@ def update_field_force_q_GPE(conv_fun,phi, types, charges, dielectric, phi_q,
     ### ^ ----- Calculate the relative dielectric (permittivity) to field
     ### ------- from a mean contribution of particle number densities
 
+    eps0_inv =  4.0 * np.pi * config.coulomb_constant
+
     for t_ in range(config.n_types): # do this with painting? (possible?)
         num_types = num_types  + (config.dielectric_type[t_])*phi[t_] #+ phi[t_].apply(dielectric_transfer_function, out = phi_eps)
         denom_phi_tot = denom_phi_tot + phi[t_]
-    phi_eps = num_types/denom_phi_tot
-
-    #print("shapes: denom, phi", np.shape(denom_phi_tot), np.shape(phi))
+    phi_eps = np.divide(num_types*eps0_inv,denom_phi_tot,
+                        where = np.abs(denom_phi_tot > 1e-15),
+                        out = np.zeros_like(phi_q))
+    # print("names ",config.name_to_type_map)
+    #print(dielectric)
+    # print("shapes: denom, phi", np.shape(denom_phi_tot), np.shape(phi))
     #print("rank {:d} : max eps {:2f} , min eps {:2f}".format(comm.Get_rank(), np.max(phi_eps), np.min(phi_eps)))
     #print("mean", np.mean(phi_eps), "rank ", comm.Get_rank())
     phi_eps.r2c(out=phi_eps_fourier)
+    #phi_eps_fourier.apply(phi_transfer_function, out=phi_eps_fourier)
+    #phi_eps_fourier.c2r(out=phi_eps)
     ###  ^ ------  Alternatively with dielectric given by id list
 
     phi_q_eps = (phi_q/phi_eps)
@@ -713,7 +720,7 @@ def update_field_force_q_GPE(conv_fun,phi, types, charges, dielectric, phi_q,
     def iterate_apply_k_vec(k,additive_terms,d = _d):
         return additive_terms * (- 1j * k[_d]) / k.normp(p=2, zeromode=1)
 
-    max_iter = 100; i = 0; delta = 1.0
+    max_iter = 200; i = 0; delta = 1.0
     #phi_pol_prev = pm.create("real", value = 0.0)
     ### ^------ set to zero before each iterative procedure or soft start
     conv_criteria = config.conv_crit # conv. criteria (default 1e-6)
@@ -733,12 +740,13 @@ def update_field_force_q_GPE(conv_fun,phi, types, charges, dielectric, phi_q,
         delta = conv_fun(comm,diff) # decided from toml input
         phi_pol_prev = phi_pol.copy()
         i = i + 1
-    #print("Stopping after iteration {:d} with stop crit {:.2e}, delta {:.2e}".format(i,conv_criteria,delta))
+
+    print("Stopping after iteration {:d} with stop crit {:.2e}, delta {:.2e}, rank: ".format(i,conv_criteria,delta, comm.Get_rank()))
 
[…]
