+++ hymd/main.py
@@ -340,25 +340,27 @@ def main():
                 comm=comm
             )
 
-            #rank = comm.Get_rank()
+            rank = comm.Get_rank()
+            #print(charges)
             #print("main",elec_forces[:,2])
-            #sum_elec = np.sum(elec_forces,axis = 0)
-            #if rank == 0:
-            #    sums = np.zeros_like(sum_elec)
-            #else:
-            #    sums = None
+            sum_elec = np.sum(elec_forces,axis = 0)
+            if rank == 0:
+                sums = np.zeros_like(sum_elec)
+            else:
+                sums = None
             #comm.Barrier()
-            #comm.Reduce(sum_elec, sums,
-            #op=MPI.SUM, root=0)
+            comm.Reduce(sum_elec, sums,
+            op=MPI.SUM, root=0)
             #print("rank ", comm.Get_rank())
             #print("sum elec_forces", np.sum(elec_forces,axis = 0))
-            #if rank == 0:
-            #    f = open("./sum_forces.txt", "w")
-            #    f.write("time step, sum forces x,y,z \n")
-            #    f.write("{:.2f}, {:.2f}, {:.2f}, {:.2f} \n".format(0, sums[0], sums[1], sums[2]))
-            #    print("here ",sums)
-            #    #print(sum_elec)
-            #    print("sum elec_forces", np.sum(elec_forces,axis = 0))
+            if rank == 0:
+                #    f = open("./sum_forces.txt", "w")
+                #    f.write("time step, sum forces x,y,z \n")
+                #    f.write("{:.2f}, {:.2f}, {:.2f}, {:.2f} \n".format(0, sums[0], sums[1], sums[2]))
+                print("here ",sums)
+                #    #print(sum_elec)
+                #    print("sum elec_forces", np.sum(elec_forces,axis = 0))
+
         if config.coulombtype == "PIC_Spectral":
             update_field_force_q(
                 charges, phi_q, phi_q_fourier, elec_field_fourier, elec_field,
@@ -369,6 +371,24 @@ def main():
                 config, phi_q_fourier, elec_energy_field, field_q_energy,
                 comm=comm,
             )
+
+            sum_elec = np.sum(elec_forces,axis = 0)
+            rank = comm.Get_rank()
+            if rank == 0:
+                sums = np.zeros_like(sum_elec)
+            else:
+                sums = None
+            #comm.Barrier()
+            comm.Reduce(sum_elec, sums,
+            op=MPI.SUM, root=0)
+            #print("rank ", comm.Get_rank())
+            #print("sum elec_forces", np.sum(elec_forces,axis = 0))
+            #print("here", sums)
+
+            if rank == 0:
+                print("sum elec forces", sums)
+            #print(config.type_to_name_map)
+            #print(charges[types == 4])
 
     if molecules_flag:
         if not (args.disable_bonds
@@ -651,6 +671,7 @@ def main():
 
         # Update slow forces
         if not args.disable_field:
+            #print("update phi w rank", comm.Get_rank())
             layouts = [
                 pm.decompose(positions[types == t]) for t in range(config.n_types)  # noqa: E501
             ]
@@ -666,6 +687,7 @@ def main():
             if charges_flag:
                 layout_q = pm.decompose(positions)
                 if config.coulombtype == "PIC_Spectral_GPE": # dielectric_flag and
+                    #print("update elec forces  w rank", comm.Get_rank())
                     phi_eps_fourier,elec_field_contrib_fourier = update_field_force_q_GPE(
                         conv_fun, phi, types, charges,dielectric_sorted,
                         phi_q, phi_q_fourier, phi_eps, phi_eps_fourier,phi_q_eps,
@@ -682,15 +704,16 @@ def main():
                         comm=comm
                     )
 
[…]
