+++ hymd/field.py
@@ -257,6 +257,7 @@ def update_field_force_q(
             elec_field[_d].readout(positions, layout=layout_q)
         )
 
+    print("elec_field PE", elec_field[2][0,0,0])
 
 def update_field_force_energy_q(
     charges, phi_q, phi_q_fourier, elec_field_fourier, elec_field, elec_forces,
@@ -622,7 +623,9 @@ def compute_field_energy_q_GPE(
 
     V = np.prod(config.box_size)
 
-    field_q_energy = 0.5 * V * comm.allreduce(np.sum(phi_eps_fourier.value*elec_field_contrib_fourier.value))
+    eps_0 = 1.0/(config.coulomb_constant*4*np.pi)
+
+    field_q_energy = (0.5 * eps_0) * V * comm.allreduce(np.sum(phi_eps_fourier*elec_field_contrib_fourier))
 
     return field_q_energy.real
 
@@ -667,8 +670,8 @@ def update_field_force_q_GPE(conv_fun,phi, types, charges, dielectric, phi_q,
     ## ^------ tbr; phi_transfer_funciton by hamiltonian.H ??
     phi_q_fourier.c2r(out=phi_q) ## this phi_q is after applying the smearing function
 
-    denom_phi_tot =  pm.create("real", value= 0.0)
-    num_types =  pm.create("real", value = 0.0)
+    denom_phi_tot =  pm.create("real", value=0.0)
+    num_types =  pm.create("real", value=0.0)
 
     #print("rank {:d} and phi shape {:}".format(comm.Get_rank(), np.shape(phi)))
     #for t_ in range(config.n_types):
@@ -678,25 +681,25 @@ def update_field_force_q_GPE(conv_fun,phi, types, charges, dielectric, phi_q,
     ### ^ ----- Calculate the relative dielectric (permittivity) to field
     ### ------- from a mean contribution of particle number densities
 
-    eps0_inv =  4.0 * np.pi * config.coulomb_constant
-
     for t_ in range(config.n_types): # do this with painting? (possible?)
         num_types = num_types  + (config.dielectric_type[t_])*phi[t_] #+ phi[t_].apply(dielectric_transfer_function, out = phi_eps)
         denom_phi_tot = denom_phi_tot + phi[t_]
-    phi_eps = np.divide(num_types*eps0_inv,denom_phi_tot,
-                        where = np.abs(denom_phi_tot > 1e-15),
-                        out = np.zeros_like(phi_q))
-    # print("names ",config.name_to_type_map)
-    #print(dielectric)
-    # print("shapes: denom, phi", np.shape(denom_phi_tot), np.shape(phi))
+    #phi_eps = num_types/denom_phi_tot
+    np.divide(num_types,denom_phi_tot,
+    where = np.abs(denom_phi_tot > 1e-6),
+    out = phi_eps)
+
+    #print("shapes: denom, phi", np.shape(denom_phi_tot), np.shape(phi))
     #print("rank {:d} : max eps {:2f} , min eps {:2f}".format(comm.Get_rank(), np.max(phi_eps), np.min(phi_eps)))
-    #print("mean", np.mean(phi_eps), "rank ", comm.Get_rank())
+    print("mean", np.mean(phi_eps), "rank ", comm.Get_rank())
     phi_eps.r2c(out=phi_eps_fourier)
-    #phi_eps_fourier.apply(phi_transfer_function, out=phi_eps_fourier)
-    #phi_eps_fourier.c2r(out=phi_eps)
     ###  ^ ------  Alternatively with dielectric given by id list
 
-    phi_q_eps = (phi_q/phi_eps)
+    #phi_q_eps = (phi_q/phi_eps)
+    np.divide(phi_q,phi_eps,
+    where = np.abs(phi_eps > 1e-6),
+    out = phi_q_eps)
+
[…]
