+++ utils/get_next_point.py
@@ -1,0 +1,161 @@
+import os
+import numpy as np
+import argparse
+from read_parameter_file import read_bounds_file
+from bayes_opt import BayesianOptimization
+from bayes_opt import UtilityFunction
+from sklearn.gaussian_process.kernels import Matern, WhiteKernel
+from sklearn.gaussian_process import GaussianProcessRegressor
+
+
+def get_next_point_BO(bounds_file_path, opt_file_path, kappa):
+    if os.path.exists(opt_file_path):
+        opt_data = np.loadtxt(opt_file_path, dtype=str)
+    else:
+        with open(opt_file_path, "w") as _:
+            ...
+        opt_data = np.empty(shape=(1, 1,), dtype=np.float64)
+
+    parameter_names_correct_order, lower_bounds, upper_bounds = (
+        read_bounds_file(bounds_file_path)
+    )
+    bounds = {}
+    for param, lower, upper in zip(
+        parameter_names_correct_order, lower_bounds, upper_bounds
+    ):
+        bounds[param] = (lower, upper)
+
+    optimizer = BayesianOptimization(f=None, pbounds=bounds, verbose=2)
+    white_kernel = WhiteKernel(noise_level_bounds=(1e-5, 1e+3))
+    matern_kernel = Matern(nu=2.5, length_scale_bounds=(1e-3, 1e3))
+    kernel = matern_kernel + white_kernel
+    optimizer._gp = GaussianProcessRegressor(
+        kernel=kernel, normalize_y=True, n_restarts_optimizer=100, alpha=0.0
+    )
+    utility = UtilityFunction(kind="ucb", kappa=args.kappa, xi=None)
+
+    print("Registering optimization results for previous runs:")
+    print(f" {'':4} |------------------------------------------------------")
+    print(f" {'':<4} | ", end="")
+    for parameter in parameter_names_correct_order:
+        print(f"{parameter:>20} ", end="")
+    print(f"{args.fitness:>20}")
+    print(f" {'':4} |------------------------------------------------------")
+
+    for i in range(opt_data.shape[1]):
+        if opt_data[0, i] == args.fitness:
+            target_ind = i
+            break
+
+    param = []
+    target = []
+
+    for iteration in range(1, opt_data.shape[0]):
+        param.append({})
+        for i in range(len(parameter_names_correct_order)):
+            param[-1][opt_data[0, i]] = float(opt_data[iteration, i])
+        target.append(float(opt_data[iteration, target_ind]))
+
+    target = np.array(target, dtype=np.float64)
+    target_flip = np.array(target, dtype=np.float64)
+    max_target = np.max(target)
+
+    for i, t in enumerate(target):
+        if args.fitness in ("MSE", "RMSE", "MAE", "MAPE", "SMAPE"):
+            target_flip[i] = max_target - t
+        else:
+            target_flip[i] = t
+
+    iteration = 0
+    for p, t, t_print in zip(param, target_flip, target):
+        iteration += 1
+        added = False
+        tries = 0
+        random_key = None
+        while not added:
+            try:
+                if tries > 0:
+                    random_key = np.random.choice(list(p.keys()), 1)[0]
+                    p[random_key] += np.random.normal(loc=0.0, scale=1e-14)
+                optimizer.register(params=p, target=t)
+                added = True
+            except KeyError:
+                tries += 1
+
+        print(f" {iteration:>4} | ", end="")
+        for parameter in parameter_names_correct_order:
+            print(f"{p[parameter]:20.15f} ", end="")
+        print(f"{t_print:20.15f} ",
+              end=f" ({tries}, {random_key})\n" if tries != 0 else "\n")
+    print(f" {'':4} |------------------------------------------------------")
+
+    print("Evaluating acquisition function for next parameter space point:")
[…]
