### 2022-03-07 b487467
update fortran code
correct merge with pressure
