### 2022-03-08 46ab9ca
correct naming of fortran functions
cbf --> cbf_d
Now pip installs w/o error
