+++ hymd/file_io.py
@@ -481,11 +481,14 @@ def store_data(
     kinetic_energy,
     bond2_energy,
     bond3_energy,
+    bond4_energy,
     field_energy,
+    field_q_energy,
     time_step,
     config,
     velocity_out=False,
     force_out=False,
+    charge_out=False,
     dump_per_particle=False,
     comm=MPI.COMM_WORLD,
 ):
