### 2022-03-08 c844c69
GPE-pressure test system
DPPC config with charges, pressure and variable dielectric.
