### 2022-03-09 5e4c035
small update main + field
elec energy sum in real space
- Decided for mathematical correctness
- before sum in fourier space
