+++ utils/make_input.py
@@ -1,150 +1,0 @@
-import h5py
-import sys
-import numpy as np
-import json
-
-CONF = {}
-exec(open(sys.argv[1]).read(), CONF)
-np.random.seed(0)
-# Initialization of simulation variables
-kb = 2.479 / 298
-if "T0" in CONF:
-    CONF["kbT0"] = kb * CONF["T0"]
-    E_kin0 = kb * 3 * CONF["Np"] * CONF["T0"] / 2.0
-
-if "T_start" in CONF:
-    CONF["kbT_start"] = kb * CONF["T_start"]
-
-Ncells = CONF["Nv"] ** 3
-CONF["L"] = np.array(CONF["L"])
-CONF["V"] = CONF["L"][0] * CONF["L"][1] * CONF["L"][2]
-CONF["dV"] = CONF["V"] / (CONF["Nv"] ** 3)
-# CONF['n_frames'] = CONF['NSTEPS']//CONF['nprint']
-
-if "phi0" not in CONF:
-    CONF["phi0"] = CONF["Np"] / CONF["V"]
-
-f_hd5 = h5py.File("input.hdf5", "w")
-
-# dset_pos      = f_hd5.create_dataset("coordinates", (1,CONF['Np'],3), dtype="Float32")
-# dset_vel      = f_hd5.create_dataset("velocities",  (1,CONF['Np'],3), dtype="Float32")
-dset_pos = f_hd5.create_dataset("coordinates", (1, CONF["Np"], 3), dtype="float32")
-dset_vel = f_hd5.create_dataset("velocities", (1, CONF["Np"], 3), dtype="float32")
-dset_types = f_hd5.create_dataset("types", (CONF["Np"],), dtype="i")
-dset_indices = f_hd5.create_dataset("indices", (CONF["Np"],), dtype="i")
-
-dt = h5py.string_dtype(encoding="ascii")
-dset_names = f_hd5.create_dataset("names", (CONF["Np"],), dtype="S10")
-# dset_mass = f_hd5.create_dataset("mass",  (CONF['Np'],), dtype="Float32")
-dset_mass = f_hd5.create_dataset("mass", (CONF["Np"],), dtype="float32")
-
-dset_pos.attrs["units"] = "nanometers"
-dset_pos.attrs["units"] = "nanometers/picoseconds"
-
-indices = []
-parts = 1000
-n = int(CONF["Np"] / parts)
-# print(n,CONF['Np'])
-for i in range(parts - 1):
-    indices.append(list(range(i * n, (i + 1) * n)))
-
-if not (n * i == CONF["Np"]):
-    indices.append(list(range((i + 1) * n, CONF["Np"])))
-
-
-def cube(r):
-    radius = np.max(np.abs(r - CONF["L"][None, :] * 0.5), axis=1)
-    ind = np.argsort(radius)[::-1]
-    r = r.copy()[ind, :]
-
-    return r
-
-
-def GEN_START_UNIFORM():
-
-    n = int(np.ceil(CONF["Np"] ** (1.0 / 3.0)))
-    print(n)
-    l = CONF["L"][0] / n
-    x = np.linspace(0.5 * l, (n - 0.5) * l, n)
-
-    j = 0
-    for ix in range(n):
-        for iy in range(n):
-            for iz in range(n):
-                if j < CONF["Np"]:
-                    print(x[ix])
-                    dset_pos[0, j, 0] = x[ix]
-                    dset_pos[0, j, 1] = x[iy]
-                    dset_pos[0, j, 2] = x[iz]
-    if j < CONF["Np"]:
-        dset_pos[0, j:, :] = CONF["L"] * np.random.random((CONF["Np"] - j, 3))
[…]
