### 2022-03-09 d6d6084
Fix error in sympy.diff introduced with sympy 1.10
After sympy 1.9, differentiating a function with respect to a string
representing a sympy Symbol appears to not work. Changed in
hymd/hamiltonian.py to instead differentiate explicitly with regards to
the field Symbol, self.phi[i].
