### 2022-03-09 769440a
Fix sympy differentiate bug in test/conftest.py
Also fix a bug in pytest-mpi where coverage output was not handled
correctly with the --coverage and --cov-append flags provided.
