+++ pyproject.toml
@@ -1,2 +1,4 @@
 [build-system]
 requires = ["setuptools", "wheel", "numpy", "cython", "mpi4py"]
+[tool.black]
+exclude = "\\(./)"
