### 2022-03-11 a8c9da0
Remove legacy scripts (#169)
