+++ utils/gro2fort5.py
@@ -1,645 +1,0 @@
-import os
-import subprocess
-import argparse
-import numpy as np
-
-
-def _format_fort5_line(index, atom_type, atom_index, total_bonds,
-                       x, y, z, vx, vy, vz, bonds, velocity=True,
-                       zero_velocity=False):
-    return_str = f'{index:>5d} {atom_type:<3s} {atom_index} {total_bonds}\t'
-    return_str += ''.join(f'{p:<20.15f}' for p in (x, y, z))
-    if velocity:
-        if zero_velocity:
-            return_str += ''.join(f'{v:<20.15f}' for v in (0.0, 0.0, 0.0))
-        else:
-            return_str += ''.join(f'{v:<20.15f}' for v in (vx, vy, vz))
-    return_str += '\t' + ' '.join(f'{b}' for b in bonds)
-    return return_str, [vx, vy, vz]
-
-
-def _convert_DPPC(molecule_index, name, index,
-                  x, y, z, vx, vy, vz, split_carbons=False, velocity=True,
-                  zero_velocity=False, water_index=5):
-    index, x, y, z, vx, vy, vz = (int(index), float(x), float(y), float(z),
-                                  float(vx), float(vy), float(vz))
-    bonds = [0, 0, 0, 0, 0, 0]
-    if name == 'NC3':  # index 1, bonds 2
-        atom_type = 'N'
-        atom_index = 1
-        bonds[0] = index + 1
-        total_bonds = 1
-    elif name == 'PO4':  # index 2, bonds 1 3
-        atom_type = 'P'
-        atom_index = 2
-        bonds[:2] = index - 1, index + 1
-        total_bonds = 2
-    elif name == 'GL1':  # index 3, bonds 2 4 5
-        atom_type = 'G'
-        atom_index = 3
-        bonds[:3] = index - 1, index + 1, index + 2
-        total_bonds = 3
-    elif name == 'GL2':  # index 4, bonds 3 9
-        atom_type = 'G'
-        atom_index = 3
-        bonds[:2] = index - 1, index + 5
-        total_bonds = 2
-    elif name == 'C1A':  # index 5, bonds 3 6
-        atom_type = 'C'
-        atom_index = 4
-        bonds[:2] = index - 2, index + 1
-        total_bonds = 2
-    elif name == 'C2A':  # index 6, 5 7
-        atom_type = 'C'
-        atom_index = 4
-        bonds[:2] = index - 1, index + 1
-        total_bonds = 2
-    elif name == 'C3A':  # index 7, 6 8
-        atom_type = 'C'
-        atom_index = 4
-        bonds[:2] = index - 1, index + 1
-        total_bonds = 2
-    elif name == 'C4A':  # index 8, 7
-        atom_type = 'CE' if split_carbons else 'C'
-        atom_index = 6 if split_carbons else 4
-        bonds[0] = index - 1
-        total_bonds = 1
-    elif name == 'C1B':  # index 9, 4 10
-        atom_type = 'C'
-        atom_index = 4
-        bonds[:2] = index - 5, index + 1
-        total_bonds = 2
-    elif name == 'C2B':  # index 10, 9 11
-        atom_type = 'C'
-        atom_index = 4
-        bonds[:2] = index - 1, index + 1
-        total_bonds = 2
-    elif name == 'C3B':  # index 11, 10 12
-        atom_type = 'C'
-        atom_index = 4
-        bonds[:2] = index - 1, index + 1
-        total_bonds = 2
-    elif name == 'C4B':  # index 12, 11
-        atom_type = 'CE' if split_carbons else 'C'
-        atom_index = 6 if split_carbons else 4
-        bonds[0] = index - 1
-        total_bonds = 1
-    else:
-        raise ValueError(f'Unrecognized name in DPPC: {name}')
[…]
