+++ hymd/version.py
@@ -1,1 +1,1 @@
-__version__ = "1.0.6"
+__version__ = "1.0.6-post1"
