+++ hymd/main.py
@@ -176,6 +176,7 @@ def main():
 
     # Ignore numpy numpy.VisibleDeprecationWarning: Creating an ndarray from
     # ragged nested sequences until it is fixed in pmesh
+    # to be deleted 
     with warnings.catch_warnings():
         warnings.filterwarnings(
             action="ignore",
