### 2022-03-12 13c53e2
simple clean pmesh + filtered forces
- removed two unused pmeshes in elecGPE
- forces now filtered, sums to zero, but needs further check to be sure/safe
