+++ test-system/run-dppc-pressure/run-pressure-test.sh
@@ -1,3 +1,3 @@
 ## test for Lasse GPE
 
-mpirun -n 1 python3 -m  hymd configtest_HyMD.tomli dppc-with-charge-dielectric.h5  --double-precision --verbose  --velocity-output  --logfile GPE-log.txt
+mpirun -n 2 python3 -m  hymd configtest_HyMD.tomli dppc-with-charge-dielectric.h5  --double-precision --verbose  --velocity-output  --logfile GPE-log.txt
