+++ hymd/field.py
@@ -67,17 +67,13 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         list_coulomb = [phi_q, phi_q_fourier, elec_field_fourier, elec_field, elec_energy_field]
 
     if config.coulombtype == 'PIC_Spectral_GPE': ## initializing the density mesh #dielectric_flag
-        
+
         #samiran to lasse: can we write this block of creations just once
         #if config.coulombtype: create these
         phi_q = pm.create("real", value=0.0)
         phi_q_fourier = pm.create("complex", value=0.0)
         elec_field_fourier= [pm.create("complex", value=0.0) for _ in range(_SPACE_DIM)] #for force calculation
         elec_field = [pm.create("real", value=0.0) for _ in range(_SPACE_DIM)] #for force calculation
-        elec_energy_field = pm.create("complex", value=0.0) # for energy calculation --> complex form needed as its converted from complex field; Imaginary part as zero;
-        
-        #samiran to lasse: do you use this:
-        elec_energy_field_real = pm.create("real", value=0.0)
 
         ## GPE relevant
         phi_q_eps = pm.create("real", value = 0.0) ## real contrib of non-polarization part of GPE
@@ -94,13 +90,18 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         phi_pol_prev = pm.create("real", value = 0.0)
         elec_dot = pm.create("real", value = 0.0)
 
+        # Vbar elec needed before obtaining external potential
+        Vbar_elec = phi = [pm.create("real", value=0.0) for _ in range(config.n_types)]
+        Vbar_elec_fourier = [
+            pm.create("complex", value=0.0) for _ in range(config.n_types)
+        ]
+
         ## Polarization force contribution
         elec_field_contrib = pm.create("real", value = 0.0) # needed for pol energies later
-        #elec_energy_field_real = pm.create("real", value=0.0)
-        list_coulomb = [phi_q, phi_q_fourier, elec_field_fourier, elec_field, elec_energy_field,
+        list_coulomb = [phi_q, phi_q_fourier, elec_field_fourier, elec_field,
                 phi_q_eps, phi_q_eps_fourier, phi_q_effective_fourier, phi_eps, phi_eps_fourier,
                 phi_eta, phi_eta_fourier, phi_pol, phi_pol_fourier, phi_pol_temp, sum_fourier,
-                phi_pol_prev, elec_dot
+                phi_pol_prev, elec_dot, elec_field_contrib, Vbar_elec, Vbar_elec_fourier
                 ]
 
     field_list = [phi, phi_fourier, force_on_grid, v_ext_fourier, v_ext, phi_transfer,
@@ -860,12 +861,12 @@ def compute_field_energy_q_GPE(
 
     return field_q_energy
 
-def update_field_force_q_GPE(conv_fun,phi, types, charges, dielectric, phi_q,
+def update_field_force_q_GPE(conv_fun,phi, types, charges, config_charges, dielectric, phi_q,
     phi_q_fourier,phi_eps, phi_eps_fourier,phi_q_eps, phi_q_eps_fourier,
     phi_q_effective_fourier,phi_eta,  phi_eta_fourier, phi_pol_prev,
     phi_pol, phi_pol_fourier,sum_fourier,phi_pol_temp,elec_field_fourier,
-    elec_field,elec_forces,elec_field_contrib, layout_q, layout,pm,positions,
-    config,comm = MPI.COMM_WORLD, compute_potential = False,
+    elec_field,elec_forces,elec_field_contrib, Vbar_elec, Vbar_elec_fourier, hamiltonian,
[…]
