### 2022-03-12 ceb3f5a
GPE: small cleanup + reuse of pmeshes
