+++ hymd/field.py
@@ -72,10 +72,8 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         #if config.coulombtype: create these
         phi_q = pm.create("real", value=0.0)
         phi_q_fourier = pm.create("complex", value=0.0)
-        elec_field_fourier= [pm.create("complex", value=0.0) for _ in range(_SPACE_DIM)] #for force calculation
         elec_field = [pm.create("real", value=0.0) for _ in range(_SPACE_DIM)] #for force calculation
 
-        ## GPE relevant
         phi_q_eps = pm.create("real", value = 0.0) ## real contrib of non-polarization part of GPE
         phi_q_eps_fourier = pm.create("complex", value = 0.0) # complex contrib of phi q eps
         phi_q_effective_fourier = pm.create("complex", value = 0.0) ## fourier of non-polarization part of GPE
@@ -85,23 +83,36 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         phi_eta_fourier = [pm.create("complex", value = 0.0)for _ in range(_SPACE_DIM)] ## fourier of factor in polarization charge density
         phi_pol = pm.create("real", value = 0.0) ## real contrib of the polarization charge
         phi_pol_fourier = [pm.create("complex", value = 0.0) for _ in range(_SPACE_DIM)] # complex contrib of the polarization charge
-        phi_pol_temp = [pm.create("real", value = 0.0) for _ in range (_SPACE_DIM)] # complex contrib of the polarization charge
-        sum_fourier = pm.create("complex", value = 0.0)
         phi_pol_prev = pm.create("real", value = 0.0)
         elec_dot = pm.create("real", value = 0.0)
+        elec_potential = pm.create("real", value = 0.0)
 
-        # Vbar elec needed before obtaining external potential
-        Vbar_elec = phi = [pm.create("real", value=0.0) for _ in range(config.n_types)]
+        # External potential and force meshes
+        Vbar_elec = phi = [
+                    pm.create("real", value=0.0) for _ in range(config.n_types)
+        ]
+
         Vbar_elec_fourier = [
             pm.create("complex", value=0.0) for _ in range(config.n_types)
         ]
 
+        force_mesh_elec = [
+                    [pm.create("real", value=0.0) for d in range(3)
+                    ] for _ in range(config.n_types)
+        ]
+
+        force_mesh_elec_fourier = [
+                    [pm.create("complex", value=0.0) for d in range(3)
+                    ] for _ in range(config.n_types)
+        ]
+
         ## Polarization force contribution
         elec_field_contrib = pm.create("real", value = 0.0) # needed for pol energies later
-        list_coulomb = [phi_q, phi_q_fourier, elec_field_fourier, elec_field,
+        list_coulomb = [phi_q, phi_q_fourier, elec_field,
                 phi_q_eps, phi_q_eps_fourier, phi_q_effective_fourier, phi_eps, phi_eps_fourier,
-                phi_eta, phi_eta_fourier, phi_pol, phi_pol_fourier, phi_pol_temp, sum_fourier,
-                phi_pol_prev, elec_dot, elec_field_contrib, Vbar_elec, Vbar_elec_fourier
+                phi_eta, phi_eta_fourier, phi_pol, phi_pol_fourier,
+                phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec, Vbar_elec_fourier,
+                force_mesh_elec, force_mesh_elec_fourier
                 ]
 
[…]
