### 2022-03-13 ac9c4bb
config_charges
added np any test in config_charges loop to avoid looping over unecessarily
