### 2022-03-14 67275d9
concatenate fix for Vbar
Vbar + Vbar-elec now zipped to avoid concatenation of two tuples
