### 2022-03-14 05846c2
vbar + vbar-elec comment
add comment to shape Vbar after zip sum
