### 2022-03-14 45e4757
PE can now also be run w pressure
V_bar added to pm initializer, but remains a mesh of zeroes bc PE not defined from w
