### 2022-03-15 8a02986
gradient transfer to field transfer
changed names and usage of function to obtain electric field
