### 2022-03-15 ca6c80c
cleanup elec forces
Elec forces now computed in update field
