+++ hymd/field.py
@@ -100,27 +100,34 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
                     pm.create("real", value=0.0) for _ in range(config.n_types)
         ]
 
-        Vbar_elec_fourier = [
-            pm.create("complex", value=0.0) for _ in range(config.n_types)
-        ]
+        #Vbar_elec_fourier = [
+        #    pm.create("complex", value=0.0) for _ in range(config.n_types)
+        #]
 
-        force_mesh_elec = [
-                    [pm.create("real", value=0.0) for d in range(3)
-                    ] for _ in range(config.n_types)
-        ]
+        #force_mesh_elec = [
+        #            [pm.create("real", value=0.0) for d in range(3)
+        #            ] for _ in range(config.n_types)
+        #]
 
-        force_mesh_elec_fourier = [
-                    [pm.create("complex", value=0.0) for d in range(3)
-                    ] for _ in range(config.n_types)
-        ]
+        #force_mesh_elec_fourier = [
+        #            [pm.create("complex", value=0.0) for d in range(3)
+        #            ] for _ in range(config.n_types)
+        #]
 
         ## Polarization force contribution
         elec_field_contrib = pm.create("real", value = 0.0) # needed for pol energies later
+
+        #list_coulomb = [phi_q, phi_q_fourier, elec_field,
+        #        phi_q_eps, phi_q_eps_fourier, phi_q_effective_fourier, phi_eps, phi_eps_fourier,
+        #        phi_eta, phi_eta_fourier, phi_pol, phi_pol_fourier,
+        #        phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec, Vbar_elec_fourier,
+        #        force_mesh_elec, force_mesh_elec_fourier
+        #        ]
+
         list_coulomb = [phi_q, phi_q_fourier, elec_field,
                 phi_q_eps, phi_q_eps_fourier, phi_q_effective_fourier, phi_eps, phi_eps_fourier,
                 phi_eta, phi_eta_fourier, phi_pol, phi_pol_fourier,
-                phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec, Vbar_elec_fourier,
-                force_mesh_elec, force_mesh_elec_fourier
+                phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec
                 ]
 
     field_list = [phi, phi_fourier, force_on_grid, v_ext_fourier, v_ext, phi_transfer,
@@ -606,6 +613,7 @@ def update_field(
     v_ext_fourier,
     phi_lap_filtered,
     v_ext1,
+    v_ext_elec,
     m,
     compute_potential=False,
 ):
@@ -750,10 +758,17 @@ def update_field(
         hamiltonian.v_ext1(phi_lap_filtered, v_ext1)
 
     for t in range(config.n_types):
-        if config.squaregradient:
-            v = hamiltonian.v_ext[t](phi) + v_ext1[t]
+        if  config.squaregradient:
+            if config.coulombtype == 'PIC_Spectral_GPE':
+                v = hamiltonian.v_ext[t](phi) + v_ext1[t] + v_ext_elec[t]
+            else:
+                v = hamiltonian.v_ext[t](phi) + v_ext1[t]
         else:
-            v = hamiltonian.v_ext[t](phi)
+            if config.coulombtype == 'PIC_Spectral_GPE':
[…]
