### 2022-03-15 6b661ee
reduced nr of pmesh in GPE
from 17 to 13
