+++ hymd/field.py
@@ -82,15 +82,15 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         phi_q_fourier = pm.create("complex", value=0.0)
         elec_field = [pm.create("real", value=0.0) for _ in range(_SPACE_DIM)] #for force calculation
 
-        phi_q_eps = pm.create("real", value = 0.0) ## real contrib of non-polarization part of GPE
-        phi_q_eps_fourier = pm.create("complex", value = 0.0) # complex contrib of phi q eps
-        phi_q_effective_fourier = pm.create("complex", value = 0.0) ## fourier of non-polarization part of GPE
+        #phi_q_eps = pm.create("real", value = 0.0) ## real contrib of non-polarization part of GPE
+        #phi_q_eps_fourier = pm.create("complex", value = 0.0) # complex contrib of phi q eps
+        #phi_q_effective_fourier = pm.create("complex", value = 0.0) ## fourier of non-polarization part of GPE
         phi_eps = pm.create("real", value = 0.0) ## real contrib of the epsilon dielectric painted to grid
         phi_eps_fourier = pm.create("complex", value = 0.0) # complex contrib of phi eps
         phi_eta = [pm.create("real", value = 0.0)for _ in range(_SPACE_DIM)] ## real contrib of factor in polarization charge density
         phi_eta_fourier = [pm.create("complex", value = 0.0)for _ in range(_SPACE_DIM)] ## fourier of factor in polarization charge density
         phi_pol = pm.create("real", value = 0.0) ## real contrib of the polarization charge
-        phi_pol_fourier = [pm.create("complex", value = 0.0) for _ in range(_SPACE_DIM)] # complex contrib of the polarization charge
+        #phi_pol_fourier = [pm.create("complex", value = 0.0) for _ in range(_SPACE_DIM)] # complex contrib of the polarization charge
         phi_pol_prev = pm.create("real", value = 0.0)
         elec_dot = pm.create("real", value = 0.0)
         elec_potential = pm.create("real", value = 0.0)
@@ -125,8 +125,8 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         #        ]
 
         list_coulomb = [phi_q, phi_q_fourier, elec_field,
-                phi_q_eps, phi_q_eps_fourier, phi_q_effective_fourier, phi_eps, phi_eps_fourier,
-                phi_eta, phi_eta_fourier, phi_pol, phi_pol_fourier,
+                phi_eps, phi_eps_fourier,
+                phi_eta, phi_eta_fourier, phi_pol,
                 phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec
                 ]
 
@@ -897,9 +897,8 @@ def compute_field_energy_q_GPE(
     return field_q_energy
 
 def update_field_force_q_GPE(conv_fun,phi, types, charges, config_charges, phi_q,
-    phi_q_fourier,phi_eps, phi_eps_fourier,phi_q_eps, phi_q_eps_fourier,
-    phi_q_effective_fourier,phi_eta, phi_eta_fourier, phi_pol_prev,
-    phi_pol, phi_pol_fourier, elec_field,elec_forces, elec_field_contrib, elec_potential,
+    phi_q_fourier,phi_eps, phi_eps_fourier,phi_eta, phi_eta_fourier, phi_pol_prev,
+    phi_pol, elec_field,elec_forces, elec_field_contrib, elec_potential,
[…]
