### 2022-03-15 da1535e
pmesh initializer update
added elec list common for pmeshes shared between PE and GPE methods
