+++ hymd/field.py
@@ -116,33 +116,33 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
                     pm.create("real", value=0.0) for _ in range(config.n_types)
         ]
 
-        #Vbar_elec_fourier = [
-        #    pm.create("complex", value=0.0) for _ in range(config.n_types)
-        #]
+        Vbar_elec_fourier = [
+            pm.create("complex", value=0.0) for _ in range(config.n_types)
+        ]
 
-        #force_mesh_elec = [
-        #            [pm.create("real", value=0.0) for d in range(3)
-        #            ] for _ in range(config.n_types)
-        #]
+        force_mesh_elec = [
+                    [pm.create("real", value=0.0) for d in range(3)
+                    ] for _ in range(config.n_types)
+        ]
 
-        #force_mesh_elec_fourier = [
-        #            [pm.create("complex", value=0.0) for d in range(3)
-        #            ] for _ in range(config.n_types)
-        #]
+        force_mesh_elec_fourier = [
+                    [pm.create("complex", value=0.0) for d in range(3)
+                    ] for _ in range(config.n_types)
+        ]
 
-
-        #list_coulomb = [phi_q, phi_q_fourier, elec_field,
-        #        phi_q_eps, phi_q_eps_fourier, phi_q_effective_fourier, phi_eps, phi_eps_fourier,
-        #        phi_eta, phi_eta_fourier, phi_pol, phi_pol_fourier,
-        #        phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec, Vbar_elec_fourier,
-        #        force_mesh_elec, force_mesh_elec_fourier
-        #        ]
 
         list_coulomb = [phi_q, phi_q_fourier, elec_field,
                 phi_eps, phi_eps_fourier,
                 phi_eta, phi_eta_fourier, phi_pol,
-                phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec
+                phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec, Vbar_elec_fourier,
+                force_mesh_elec, force_mesh_elec_fourier
                 ]
+
+        #list_coulomb = [phi_q, phi_q_fourier, elec_field,
+        #        phi_eps, phi_eps_fourier,
+        #        phi_eta, phi_eta_fourier, phi_pol,
+        #        phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec
+        #        ]
 
     field_list = [phi, phi_fourier, force_on_grid, v_ext_fourier, v_ext, phi_transfer,
             phi_gradient, phi_laplacian, phi_lap_filtered_fourier, phi_lap_filtered,
@@ -150,7 +150,6 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
             ]
 
     return (pm, field_list, list_coulomb,list_elec_common)
-
 def compute_field_force(layouts, r, force_mesh, force, types, n_types):
     """Interpolate particle-field forces from the grid onto particle positions
 
@@ -628,7 +627,6 @@ def update_field(
     v_ext_fourier,
     phi_lap_filtered,
     v_ext1,
-    v_ext_elec,
     m,
     compute_potential=False,
 ):
@@ -773,17 +771,10 @@ def update_field(
         hamiltonian.v_ext1(phi_lap_filtered, v_ext1)
 
     for t in range(config.n_types):
-        if  config.squaregradient:
[…]
