### 2022-03-17 0f12e85
elec_common_list + length
Avoid subscribing twice
