+++ hymd/field.py
@@ -131,7 +131,7 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         ]
 
 
-        list_coulomb = [phi_q, phi_q_fourier, elec_field,
+        list_coulomb = [
                 phi_eps, phi_eps_fourier,
                 phi_eta, phi_eta_fourier, phi_pol,
                 phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec, Vbar_elec_fourier,
