### 2022-03-21 7f94b33
Remove hanging if statement
