### 2022-03-22 #4149 @rkurchin
Hi all, checking in on this! Happy for any general updates, but here are some questions related to specific issues:
* @blakeaw, were you able to resolve the issues with testing from https://github.com/Cascella-Group-UiO/HyMD/issues/160 ?
* @mortele (and/or @Lun4m and @samiransen23), what's the status of the community guidelines as discussed at https://github.com/Cascella-Group-UiO/HyMD/issues/159 ?
* It seems like both reviewers also filed some issues related to ease of installation of the package, e.g. https://github.com/Cascella-Group-UiO/HyMD/issues/157 and https://github.com/Cascella-Group-UiO/HyMD/issues/161, so would also be good to get updates on those...
