### 2022-03-22 ec59554
Flesh out the installation guide
Add dnf-based Fedora and brew-based macOS build guides, in addition
to the apt-based Ubuntu guide. Reorder to put all dependencies before
the HyMD install step to make it more.. sane.
