+++ docs/doc_pages/installation.rst
@@ -2,41 +2,156 @@
 
 Installation
 ############
+
+Dependencies
+============
+
+.. note::
+    HyMD requires Python ≥3.6. Python version 3.9 is most likely fine
+    to use, but HyMD is only regularly tested with Python3.8 at the moment.
+
+.. warning::
+    MPI-enabled h5py may fail to build under Python3.10. In this case, a
+    fall-back to Python3.9 or lower may be necessary.
+
+
+**Installing non-Python dependencies** may be done by
+
+.. tabs::
+
+   .. group-tab:: Ubuntu (apt)
+
+       .. code-block:: bash
+
+           sudo apt-get update -y
+           sudo apt-get install -y python3 python3-pip  # Install python3 and pip
+           sudo apt-get install -y libopenmpi-dev       # Install MPI development headers
+           sudo apt-get install -y libhdf5-openmpi-dev  # Install MPI-enabled HDF5
+           sudo apt-get install -y pkg-config           # Install pkg-config, required for h5py install
+           sudo apt-get install -y curl wget
+
+   .. group-tab:: Fedora (dnf)
+
+       .. code-block:: bash
+
+           sudo dnf update -y
+           sudo dnf install -y python3.9 python3-devel  # Install python3 and pip
+           sudo dnf install -y openmpi-devel            # Install MPI development headers
+           sudo dnf install -y hdf5-openmpi-devel       # Install MPI-enabled HDF5
+           sudo dnf install -y curl wget
+
+
+      The automatic download and building of PFFT has some issues, so we
+      manually build FFTW and PFFT before calling :code:`pip3 install pfft-python`.
+
+      .. code-block:: bash
+
+           sudo dnf install -y git
+           sudo dnf install -y libtool
+           sudo dnf install -y fftw-openmpi-devel fftw-openmpi-libs
+           export PATH=$PATH:/usr/lib64/openmpi/bin/                 # Ensure MPICC is in path
+           git clone https://github.com/mpip/pfft.git
+           cd pfft/
+           ./bootstrap.sh
+           ./configure
+           make
+           make install
+
+
+   .. group-tab:: Mac OSX (brew)
+
+       .. code-block:: bash
+
+           brew update
+           brew install python      # Install python3 and pip
+           brew install open-mpi    # Install MPI development headers
+           brew install hdf5-mpi    # Install MPI-enabled HDF5
+           brew install pkg-config  # Install pkg-config, required for h5py install
+           brew install curl wget
+
+**Installing Python dependencies** may be done by
+
+.. tabs::
+
+   .. group-tab:: Ubuntu (apt)
+
+       .. code-block:: bash
+
+           python3 -m pip install --upgrade pip
+           CC="mpicc" HDF5_MPI="ON" python3 -m pip install --no-binary=h5py h5py
+           python3 -m pip install mpi4py numpy cython
+
+
+   .. group-tab:: Fedora (dnf)
+
+       .. code-block:: bash
+
+          python3.9 -m ensurepip --upgrade
+          PATH=$PATH:/usr/lib64/openmpi/bin/:/usr/lib64/openmpi/lib/
+          python3.9 -m pip install mpi4py numpy cython
+          export HDF5_DIR="/usr/lib64/openmpi/"
+          CC="mpicc" HDF5_MPI="ON" python3.9 -m pip install --no-binary=h5py h5py
+
+   .. group-tab:: Mac OSX (brew)
+
+       Find the location of the installed :code:`hdf5-mpi` package by
+
+       .. code-block:: bash
+
+           find /usr -iname "*hdf5.h"
+
+       or
+
+       .. code-block:: bash
+
+           brew info hdf5-mpi
+
[…]
