### 2022-03-22 Cascella-Group-UiO/HyMD#161 @mortele
Hi again @blakeaw, #172 updates and rearranges the installation instructions to explicitly include installing an MPI-enabled C compiler. Also, the dependencies are moved further up, making the whole presentation more sane.

Think this issue can be safely closed now.
