### 2022-03-22 Cascella-Group-UiO/HyMD#157 @mortele
Hi again @yhtang, the documentation should now be updated with instructions for installing HyMD using apt, dnf, or homebrew, see #172 .

I did not add virtual environment-based installation steps, because I mostly expect users to know themselves when/if that is a good idea for their specific use case. I did however note that on Fedora, the default Python3.10 has some issues compiling MPI-enabled h5py, so maybe it would be good to add a small pyenv guide.
