### 2022-03-22 Cascella-Group-UiO/HyMD#160 @mortele
Hi again, @blakeaw . I have tried to reproduce this behaviour across three different operating systems now, but I have not seen anything that looks like these `test_force` failures. Can you give me some more information about the conda environment you are using to run these tests?
