### 2022-03-22 Cascella-Group-UiO/HyMD#159 @mortele
Just linking #163 here for reference.
