### 2022-03-23 #4149 @mortele
Hi again @rkurchin. The docs have been updated to reflect new and improved build directives for the package. Hopefully that should be sufficient to address the last point. Regarding the finalization of the community contributor guidelines, I will have an update within a day or two. As for the first point—the mysterious `test_force` failures reported by @blakeaw, I am  so far stumped. But hopefully I will be able to work out what is going on with a little more information about the specific environment used.
