### 2022-03-23 Cascella-Group-UiO/HyMD#161 @blakeaw
Hi @mortele, it looks good. Thanks. I'll go ahead and close.

Just as a side note,  as the warning in installation docs suggested I couldn't get mpi-enabled h5py to install with python 3.10. But there weren't any problems installing in a new conda environment with python 3.8.
