### 2022-03-23 Cascella-Group-UiO/HyMD#160 @blakeaw
Hi @mortele. Yeah, it is odd.
I haven't gotten around to trying it on a different machine myself yet, but I did reinstall everything in a fresh conda environment with python 3.8, and I pulled in the most recent changes to the HyMD repo.

I still get a similar `test_force` failure in `test_comp_bonds` when I run pytest:

```
====================================================================== FAILURES ======================================================================
__________________________________________________________________ test_comp_bonds ___________________________________________________________________

dppc_single = (array([ 0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11]), array([[ 1, -1, -1],
       [ 0,  2, -1],
       [ 1,  3,  4... equilibrium=180.0, strength=25.0), Angle(atom_1='C', atom_2='C', atom_3='C', equilibrium=180.0, strength=25.0)), ...})

    def test_comp_bonds(dppc_single):
        indices, bonds, names, molecules, r, CONF = dppc_single
        config = Config(n_steps=1, time_step=0.03, mesh_size=[30, 30, 30],
                        box_size=np.array([13.0, 13.0, 14.0]), sigma=0.5, kappa=1)
        config.bonds = CONF['bond_2']
        bonds_2, _, _, _ = prepare_bonds(molecules, names, bonds, indices, config)

        expected_energies = np.array([0.24545803261508981,
                                      0.76287125411373635,
                                      2.1147847786016976,
                                      9.3338621118512890,
                                      5.2422610583092810,
                                      1.5660141344942931,
                                      0.32418863877682835,
                                      0.15582792534587750,
                                      6.2278499169351944,
                                      0.65253272104760673,
                                      0.031806604466348753], dtype=np.float64)
        expected_forces_i = np.array([
            [20.998021457611852,    9.0071937483622282,  -9.5707176942411820],
            [34.384875147498597,   12.778365176329645,  -23.697507881539778],
            [17.917253461973676,  -20.426928097726201,  -67.443862458892141],
            [-52.053439897669733,  83.19032187729123,   117.06607117521537],
            [101.22671118485104,   33.178722608565110,   41.928247692168640],
            [-3.8634845559150439, -22.155214019527303,  -58.388828683840543],
            [8.9963139950908992,   12.594088552272277,   23.894075939054996],
            [-2.9353597041563950,   2.4521443094825361, -19.363379484412864],
            [95.85080439063942,    33.391869600494857,   72.575692449944356],
            [23.050482374363451,    8.4211009839435036,  32.079465755314608],
            [5.5323489341501677,    5.4217094311092637,   4.4175438063815795]],
            dtype=np.float64
        )

        expected_forces_j = np.array([
            [-20.998021457611852,  -9.0071937483622282,   9.5707176942411820],
            [-34.384875147498597, -12.778365176329645,   23.697507881539778],
            [-17.917253461973676,  20.426928097726201,   67.443862458892141],
            [52.053439897669733,  -83.19032187729123,  -117.06607117521537],
            [-101.22671118485104, -33.178722608565110,  -41.928247692168640],
            [3.8634845559150439,   22.155214019527303,   58.388828683840543],
            [-8.9963139950908992, -12.594088552272277,  -23.894075939054996],
            [2.9353597041563950,   -2.4521443094825361,  19.363379484412864],
            [-95.85080439063942,  -33.391869600494857,  -72.575692449944356],
            [-23.050482374363451,  -8.4211009839435036, -32.079465755314608],
            [-5.5323489341501677,  -5.4217094311092637,  -4.4175438063815795]],
            dtype=np.float64
        )
        for i, b in enumerate(bonds_2):
            f_bonds = np.zeros(shape=r.shape, dtype=np.float64)
            energy = 0.0
            energy = compute_bond_forces(f_bonds, r, (b,), CONF['L'])
            assert energy == pytest.approx(expected_energies[i], abs=1e-13)
>           assert f_bonds[b[0], :] == pytest.approx(expected_forces_i[i], abs=1e-13)  # noqa: E501
E           assert array([23.050... 32.07946576]) == approx([23.05...61 ± 1.0e-13])
E             comparison failed. Mismatched elements: 1 / 3:
E             Max absolute difference: 1.0658141036401503e-13
E             Max relative difference: 3.3224184958989637e-15
E             Index | Obtained           | Expected
E             (2,)  | 32.079465755314715 | 32.07946575531461 ± 1.0e-13

test/test_force.py:99: AssertionError
============================================================== short test summary info ===============================================================
FAILED test/test_force.py::test_comp_bonds - assert array([23.050... 32.07946576]) == approx([23.05...61 ± 1.0e-13])
====================================================== 1 failed, 34 passed, 7 skipped in 3.43s =======================================================
```

It looks like that particular computation falls just a little outside the comparison tolerance.

Running the mpi-enabled pytest script seems to pass but it skips several things:

```
./pytest-mpi -oo -n 2 -ns

################################################################################
################################# RANK 0 / 2 ###################################
################################################################################
plugins: mpi-0.6
collected 42 items

test/test_distribute_input.py ssss                                       [  9%]
test/test_field.py .                                                     [ 11%]
test/test_force.py ssssss                                                [ 26%]
test/test_hamiltonian.py sssssssss                                       [ 47%]
test/test_input_parser.py ss.....sssss                                   [ 76%]
test/test_integrator.py sssssssss                                        [ 97%]
test/test_thermostat.py .                                                [100%]

=================== 7 passed, 35 skipped, 1 warning in 0.22s ===================

################################################################################
################################# RANK 1 / 2 ###################################
################################################################################
plugins: mpi-0.6
collected 42 items

test/test_distribute_input.py ssss                                       [  9%]
test/test_field.py .                                                     [ 11%]
test/test_force.py ssssss                                                [ 26%]
test/test_hamiltonian.py sssssssss                                       [ 47%]
test/test_input_parser.py ss.....sssss                                   [ 76%]
test/test_integrator.py sssssssss                                        [ 97%]
test/test_thermostat.py .                                                [100%]

================== 7 passed, 35 skipped, 22 warnings in 0.23s ==================
```

Here are my conda environment details:
```
channels:
  - defaults
dependencies:
  - _libgcc_mutex=0.1=main
  - _openmp_mutex=4.5=1_gnu
  - ca-certificates=2022.3.18=h06a4308_0
  - certifi=2021.10.8=py38h06a4308_2
  - ld_impl_linux-64=2.35.1=h7274673_9
  - libffi=3.3=he6710b0_2
  - libgcc-ng=9.3.0=h5101ec6_17
  - libgomp=9.3.0=h5101ec6_17
  - libstdcxx-ng=9.3.0=hd4cf53a_17
  - ncurses=6.3=h7f8727e_2
  - openssl=1.1.1n=h7f8727e_0
  - pip=21.2.4=py38h06a4308_0
  - python=3.8.12=h12debd9_0
  - readline=8.1.2=h7f8727e_1
  - setuptools=58.0.4=py38h06a4308_0
  - sqlite=3.38.0=hc218d9a_0
  - tk=8.6.11=h1ccaba5_0
  - wheel=0.37.1=pyhd3eb1b0_0
  - xz=5.2.5=h7b6447c_0
  - zlib=1.2.11=h7f8727e_4
  - pip:
    - attrs==21.4.0
    - cython==0.29.28
    - h5py==3.6.0
    - hymd==1.0.6.post1
    - iniconfig==1.1.1
    - mpi4py==3.1.3
    - mpmath==1.2.1
    - mpsort==0.1.17
    - networkx==2.7.1
    - numpy==1.22.3
    - packaging==21.3
    - pfft-python==0.1.21
    - pluggy==1.0.0
    - pmesh==0.1.56
    - py==1.11.0
    - pyparsing==3.0.7
    - pytest==7.1.1
    - pytest-mpi==0.6
    - sympy==1.10.1
    - tomli==2.0.1
```

I used openmpi to install/build h5py and mpi4py.
