### 2022-03-23 #4149 @blakeaw
Hi @rkurchin. I went ahead closed https://github.com/Cascella-Group-UiO/HyMD/issues/161 as it has been addressed. I'll work with @mortele to see if we can figure what is going on with https://github.com/Cascella-Group-UiO/HyMD/issues/160.
