+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
 ## Contributor's Guidelines
-Reporting bugs and any problem you encounter with HyMD is the most important contribution for us.  
-You can do that by submitting them (along with feature requests or proposals) on the
+Reporting bugs and any problem you encounter with HyMD is the most important contribution to us.  
+You can do that by submitting them (along with feature requests) on the
 [issue tracker](https://github.com/Cascella-Group-UiO/HyMD/issues).  
 
 If you want to contribute your own code, this short guide takes you through the
@@ -13,25 +13,28 @@ instructions given in the [documentation](https://cascella-group-uio.github.io/H
 
 Finally install the package in development mode using
 ```bash
-cd HyMD
+cd HyMD_fork
 pip install -e .
 ```
 
 2. Contribute  
 Write your code, test it (we welcome setting up tests with [pytest](https://docs.pytest.org/en/7.0.x/)) and follow the instructions
 [here](https://cascella-group-uio.github.io/HyMD/doc_pages/overview.html#running-parallel-simulations) to make sure it runs correctly.
-When you feel ready you can open a pull request to merge your contributions in the main HyMD branch. 
 
-If you are contributing new FORTRAN force calculation routines, you need to include the files in
-the `Makefile` and inside the `force_kernel` extension in `setup.py` so they can be compiled with
+If you are contributing new Fortran routines, you need to include the files in
+the `force_kernel` extension in `setup.py` so they can be compiled with
 [f2py](https://numpy.org/doc/stable/f2py/) when running `pip install`.
-We also require you provide both a single and double precision version of the FORTRAN routines.
+We also require both a single and double precision version of the routines.
 
-To import them you simply call
+To import them inside Python you simply call
 ```python3
 from force_kernel import <routine_name>
 ```
-and use the imported module to manipulate NumPy arrays.
+
+3. Open a pull request (PR)  
+When you feel ready you can open a PR to merge your contributions in the main HyMD branch. 
+Add a simple title and give a short description of the PR, explaining what problem is trying to solve or what feature is being implemented.
+Once the PR is submitted we will review it and provide feedback.  
 
 ## Code style
 For python files we follow the [PEP8](https://www.python.org/dev/peps/pep-0008/) guidelines, so please adapt to them when implementing new code. 
@@ -43,10 +46,10 @@ The most important stylistic requirements are:
     name variables and functions all lowercase with underscore as a separator (e.g. some_variable)
     name classes with starting letters capitalized and no separator (e.g. SomeClass)
 
-While for FORTRAN the basic requirements are:
+While for Fortran the basic requirements are:
 
     use 2 spaces for indentation
     wrap lines to max 80 characters whenever possible
-    do not use GOTO statements
+    do not use 'goto' statements
     use lowercase for commands and functions
 
