+++ CONTRIBUTING.md
@@ -12,25 +12,25 @@ Clone the fork on your local machine and install the dependencies following the
 instructions given in the [documentation](https://cascella-group-uio.github.io/HyMD/doc_pages/installation.html#dependencies).
 
   Finally install the package in development mode using
-    ```bash
-    cd HyMD_fork
-    pip install -e .
-    ```
+  ```bash
+  cd HyMD_fork
+  pip install -e .
+  ```
 
 
 2. Contribute  
 Write your code, test it (we welcome setting up tests with [pytest](https://docs.pytest.org/en/7.0.x/)) and follow the instructions
 [here](https://cascella-group-uio.github.io/HyMD/doc_pages/overview.html#running-parallel-simulations) to make sure it runs correctly.
 
-If you are contributing new Fortran routines, you need to include the files in
-the `force_kernel` extension in `setup.py` so they can be compiled with
-[f2py](https://numpy.org/doc/stable/f2py/) when running `pip install`.
-We also require both a single and double precision version of the routines.
+  If you are contributing new Fortran routines, you need to include the files in
+  the `force_kernel` extension in `setup.py` so they can be compiled with
+  [f2py](https://numpy.org/doc/stable/f2py/) when running `pip install`.
+  We also require both a single and double precision version of the routines.
 
-To import them inside Python you simply call
-    ```python3
-    from force_kernel import <routine_name>
-    ```
+  To import them inside Python you simply call
+  ```python3
+  from force_kernel import <routine_name>
+  ```
 
 3. Open a pull request (PR)  
 When you feel ready you can open a PR to merge your contributions in the main HyMD branch. 
