### 2022-03-24 Cascella-Group-UiO/HyMD#160 @mortele
In a clean `ubuntu:20.04` docker container, the following runs and does *not* reproduce this error:
```bash
apt-get update -y
apt-get install -y python3 python3-pip  # Install python3 and pip
apt-get install -y libopenmpi-dev       # Install MPI development headers
apt-get install -y libhdf5-openmpi-dev  # Install MPI-enabled HDF5
apt-get install -y pkg-config           # Install pkg-config, required for h5py install
apt-get install -y curl wget
apt-get install -y git

cd /tmp
curl -O https://repo.anaconda.com/archive/Anaconda3-2021.11-Linux-x86_64.sh
bash Anaconda3-2021.11-Linux-x86_64.sh
source ~/.bashrc

conda create --name hymd python=3.8
conda activate hymd

python3 -m ensurepip --upgrade
CC="mpicc" HDF5_MPI="ON" python3 -m pip install --no-binary=h5py h5py
python3 -m pip install mpi4py numpy cython
python3 -m pip install pmesh mpsort tomli sympy
python3 -m pip install hymd

cd /tmp
git clone https://github.com/Cascella-Group-UiO/HyMD.git hymd
cd hymd
python3 -m pip install pytest pytest-mpi
pytest
```

------
At this point I am stumped and starting to think that the tolerance we defined is just slightly too strict for the level of numerical precision we can expect to get with double-type force calculations across different machines. 🤔

<summary>CONDA ENVIRONMENT SUMMARY</summary>

This is using the conda environment
```bash
# packages in environment at /root/anaconda3/envs/hymd:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main
_openmp_mutex             4.5                       1_gnu
attrs                     21.4.0                   pypi_0    pypi
ca-certificates           2022.3.18            h06a4308_0
certifi                   2021.10.8        py38h06a4308_2
cython                    0.29.28                  pypi_0    pypi
h5py                      3.6.0                    pypi_0    pypi
hymd                      1.0.6.post1              pypi_0    pypi
iniconfig                 1.1.1                    pypi_0    pypi
ld_impl_linux-64          2.35.1               h7274673_9
libffi                    3.3                  he6710b0_2
libgcc-ng                 9.3.0               h5101ec6_17
libgomp                   9.3.0               h5101ec6_17
libstdcxx-ng              9.3.0               hd4cf53a_17
mpi4py                    3.1.3                    pypi_0    pypi
mpmath                    1.2.1                    pypi_0    pypi
mpsort                    0.1.17                   pypi_0    pypi
ncurses                   6.3                  h7f8727e_2
networkx                  2.7.1                    pypi_0    pypi
numpy                     1.22.3                   pypi_0    pypi
openssl                   1.1.1n               h7f8727e_0
packaging                 21.3                     pypi_0    pypi
pfft-python               0.1.21                   pypi_0    pypi
pip                       21.2.4           py38h06a4308_0
pluggy                    1.0.0                    pypi_0    pypi
pmesh                     0.1.56                   pypi_0    pypi
py                        1.11.0                   pypi_0    pypi
pyparsing                 3.0.7                    pypi_0    pypi
pytest                    7.1.1                    pypi_0    pypi
pytest-mpi                0.6                      pypi_0    pypi
python                    3.8.12               h12debd9_0
readline                  8.1.2                h7f8727e_1
setuptools                58.0.4           py38h06a4308_0
sqlite                    3.38.0               hc218d9a_0
sympy                     1.10.1                   pypi_0    pypi
tk                        8.6.11               h1ccaba5_0
tomli                     2.0.1                    pypi_0    pypi
wheel                     0.37.1             pyhd3eb1b0_0
xz                        5.2.5                h7b6447c_0
zlib                      1.2.11               h7f8727e_4
```

<summary>INSTALLED APT PACKAGES</summary>

This is using the `apt` environment
```bash
(hymd) root@bfca519f2696:/tmp/hymd# apt list --installed
Listing... Done
adduser/focal,now 3.118ubuntu2 all [installed]
apt/focal-updates,now 2.0.6 amd64 [installed]
autoconf/focal,now 2.69-11.1 all [installed,automatic]
automake/focal,now 1:1.16.1-4ubuntu6 all [installed,automatic]
autotools-dev/focal,now 20180224.1 all [installed,automatic]
base-files/now 11ubuntu5.4 amd64 [installed,upgradable to: 11ubuntu5.5]
base-passwd/focal,now 3.5.47 amd64 [installed]
bash/focal-updates,now 5.0-6ubuntu1.1 amd64 [installed]
binutils-common/focal-updates,focal-security,now 2.34-6ubuntu1.3 amd64 [installed,automatic]
binutils-x86-64-linux-gnu/focal-updates,focal-security,now 2.34-6ubuntu1.3 amd64 [installed,automatic]
binutils/focal-updates,focal-security,now 2.34-6ubuntu1.3 amd64 [installed,automatic]
bsdutils/now 1:2.34-0.1ubuntu9.1 amd64 [installed,upgradable to: 1:2.34-0.1ubuntu9.3]
build-essential/focal-updates,now 12.8ubuntu1.1 amd64 [installed,automatic]
bzip2/focal,now 1.0.8-2 amd64 [installed]
ca-certificates/focal-updates,focal-security,now 20210119~20.04.2 all [installed,automatic]
coreutils/focal,now 8.30-3ubuntu2 amd64 [installed]
cpp-8/focal,now 8.4.0-3ubuntu2 amd64 [installed,automatic]
cpp-9/focal-updates,focal-security,now 9.4.0-1ubuntu1~20.04.1 amd64 [installed,automatic]
cpp/focal,now 4:9.3.0-1ubuntu2 amd64 [installed,automatic]
curl/focal-updates,focal-security,now 7.68.0-1ubuntu2.7 amd64 [installed]
dash/focal,now 0.5.10.2-6 amd64 [installed]
debconf/focal,now 1.5.73 all [installed]
debianutils/focal,now 4.9.1 amd64 [installed]
diffutils/focal,now 1:3.7-3 amd64 [installed]
dirmngr/focal-updates,now 2.2.19-3ubuntu2.1 amd64 [installed,automatic]
dpkg-dev/focal,now 1.19.7ubuntu3 all [installed,automatic]
dpkg/focal,now 1.19.7ubuntu3 amd64 [installed]
e2fsprogs/focal,now 1.45.5-2ubuntu1 amd64 [installed]
fakeroot/focal,now 1.24-1 amd64 [installed,automatic]
fdisk/now 2.34-0.1ubuntu9.1 amd64 [installed,upgradable to: 2.34-0.1ubuntu9.3]
file/focal,now 1:5.38-4 amd64 [installed,automatic]
findutils/focal,now 4.7.0-1ubuntu1 amd64 [installed]
g++-9/focal-updates,focal-security,now 9.4.0-1ubuntu1~20.04.1 amd64 [installed,automatic]
g++/focal,now 4:9.3.0-1ubuntu2 amd64 [installed,automatic]
gcc-10-base/focal-updates,focal-security,now 10.3.0-1ubuntu1~20.04 amd64 [installed]
gcc-8-base/focal,now 8.4.0-3ubuntu2 amd64 [installed,automatic]
gcc-8/focal,now 8.4.0-3ubuntu2 amd64 [installed,automatic]
gcc-9-base/focal-updates,focal-security,now 9.4.0-1ubuntu1~20.04.1 amd64 [installed,automatic]
gcc-9/focal-updates,focal-security,now 9.4.0-1ubuntu1~20.04.1 amd64 [installed,automatic]
gcc/focal,now 4:9.3.0-1ubuntu2 amd64 [installed,automatic]
gfortran-8/focal,now 8.4.0-3ubuntu2 amd64 [installed,automatic]
gfortran-9/focal-updates,focal-security,now 9.4.0-1ubuntu1~20.04.1 amd64 [installed,automatic]
gfortran/focal,now 4:9.3.0-1ubuntu2 amd64 [installed,automatic]
git-man/focal-updates,focal-security,now 1:2.25.1-1ubuntu3.2 all [installed,automatic]
git/focal-updates,focal-security,now 1:2.25.1-1ubuntu3.2 amd64 [installed]
gnupg-l10n/focal-updates,now 2.2.19-3ubuntu2.1 all [installed,automatic]
gnupg-utils/focal-updates,now 2.2.19-3ubuntu2.1 amd64 [installed,automatic]
gnupg/focal-updates,now 2.2.19-3ubuntu2.1 all [installed,automatic]
gpg-agent/focal-updates,now 2.2.19-3ubuntu2.1 amd64 [installed,automatic]
gpg-wks-client/focal-updates,now 2.2.19-3ubuntu2.1 amd64 [installed,automatic]
gpg-wks-server/focal-updates,now 2.2.19-3ubuntu2.1 amd64 [installed,automatic]
gpg/focal-updates,now 2.2.19-3ubuntu2.1 amd64 [installed,automatic]
gpgconf/focal-updates,now 2.2.19-3ubuntu2.1 amd64 [installed,automatic]
gpgsm/focal-updates,now 2.2.19-3ubuntu2.1 amd64 [installed,automatic]
gpgv/focal-updates,now 2.2.19-3ubuntu2.1 amd64 [installed]
grep/focal,now 3.4-1 amd64 [installed]
gzip/focal,now 1.10-0ubuntu4 amd64 [installed]
hostname/focal,now 3.23 amd64 [installed]
ibverbs-providers/focal,now 28.0-1ubuntu1 amd64 [installed,automatic]
init-system-helpers/focal,now 1.57 all [installed]
krb5-locales/focal-updates,focal-security,now 1.17-6ubuntu4.1 all [installed,automatic]
less/focal-updates,now 551-1ubuntu0.1 amd64 [installed,automatic]
libacl1/focal,now 2.2.53-6 amd64 [installed]
libaec-dev/focal,now 1.0.4-1 amd64 [installed,automatic]
libaec0/focal,now 1.0.4-1 amd64 [installed,automatic]
libalgorithm-diff-perl/focal,now 1.19.03-2 all [installed,automatic]
libalgorithm-diff-xs-perl/focal,now 0.04-6 amd64 [installed,automatic]
libalgorithm-merge-perl/focal,now 0.08-3 all [installed,automatic]
libapt-pkg6.0/focal-updates,now 2.0.6 amd64 [installed]
libasan5/focal-updates,focal-security,now 9.4.0-1ubuntu1~20.04.1 amd64 [installed,automatic]
libasn1-8-heimdal/focal,now 7.7.0+dfsg-1ubuntu1 amd64 [installed,automatic]
libassuan0/focal,now 2.5.3-7ubuntu2 amd64 [installed,automatic]
libatomic1/focal-updates,focal-security,now 10.3.0-1ubuntu1~20.04 amd64 [installed,automatic]
libattr1/focal,now 1:2.4.48-5 amd64 [installed]
libaudit-common/focal,now 1:2.8.5-2ubuntu6 all [installed]
libaudit1/focal,now 1:2.8.5-2ubuntu6 amd64 [installed]
libbinutils/focal-updates,focal-security,now 2.34-6ubuntu1.3 amd64 [installed,automatic]
libblkid1/now 2.34-0.1ubuntu9.1 amd64 [installed,upgradable to: 2.34-0.1ubuntu9.3]
libbrotli1/focal-updates,focal-security,now 1.0.7-6ubuntu0.1 amd64 [installed,automatic]
libbsd0/focal,now 0.10.0-1 amd64 [installed,automatic]
libbz2-1.0/focal,now 1.0.8-2 amd64 [installed]
libc-bin/now 2.31-0ubuntu9.2 amd64 [installed,upgradable to: 2.31-0ubuntu9.7]
libc-dev-bin/focal-updates,focal-security,now 2.31-0ubuntu9.7 amd64 [installed,automatic]
libc6-dev/focal-updates,focal-security,now 2.31-0ubuntu9.7 amd64 [installed,automatic]
libc6/focal-updates,focal-security,now 2.31-0ubuntu9.7 amd64 [installed]
libcaf-openmpi-3/focal,now 2.8.0-1 amd64 [installed,automatic]
libcap-ng0/focal,now 0.7.9-2.1build1 amd64 [installed]
libcbor0.6/focal,now 0.6.0-0ubuntu1 amd64 [installed,automatic]
libcc1-0/focal-updates,focal-security,now 10.3.0-1ubuntu1~20.04 amd64 [installed,automatic]
libcoarrays-dev/focal,now 2.8.0-1 amd64 [installed,automatic]
libcoarrays-openmpi-dev/focal,now 2.8.0-1 amd64 [installed,automatic]
libcom-err2/focal,now 1.45.5-2ubuntu1 amd64 [installed]
libcrypt-dev/focal,now 1:4.4.10-10ubuntu4 amd64 [installed,automatic]
libcrypt1/focal,now 1:4.4.10-10ubuntu4 amd64 [installed]
libctf-nobfd0/focal-updates,focal-security,now 2.34-6ubuntu1.3 amd64 [installed,automatic]
libctf0/focal-updates,focal-security,now 2.34-6ubuntu1.3 amd64 [installed,automatic]
libcurl3-gnutls/focal-updates,focal-security,now 7.68.0-1ubuntu2.7 amd64 [installed,automatic]
libcurl4/focal-updates,focal-security,now 7.68.0-1ubuntu2.7 amd64 [installed,automatic]
libdb5.3/focal,now 5.3.28+dfsg1-0.6ubuntu2 amd64 [installed]
libdebconfclient0/focal,now 0.251ubuntu1 amd64 [installed]
libdpkg-perl/focal,now 1.19.7ubuntu3 all [installed,automatic]
libedit2/focal,now 3.1-20191231-1 amd64 [installed,automatic]
liberror-perl/focal,now 0.17029-1 all [installed,automatic]
libevent-2.1-7/focal,now 2.1.11-stable-1 amd64 [installed,automatic]
libevent-core-2.1-7/focal,now 2.1.11-stable-1 amd64 [installed,automatic]
libevent-dev/focal,now 2.1.11-stable-1 amd64 [installed,automatic]
libevent-extra-2.1-7/focal,now 2.1.11-stable-1 amd64 [installed,automatic]
libevent-openssl-2.1-7/focal,now 2.1.11-stable-1 amd64 [installed,automatic]
libevent-pthreads-2.1-7/focal,now 2.1.11-stable-1 amd64 [installed,automatic]
libexpat1-dev/focal-updates,focal-security,now 2.2.9-1ubuntu0.4 amd64 [installed,automatic]
libexpat1/focal-updates,focal-security,now 2.2.9-1ubuntu0.4 amd64 [installed,automatic]
libext2fs2/focal,now 1.45.5-2ubuntu1 amd64 [installed]
libfabric1/focal,now 1.6.2-3 amd64 [installed,automatic]
libfakeroot/focal,now 1.24-1 amd64 [installed,automatic]
libfdisk1/now 2.34-0.1ubuntu9.1 amd64 [installed,upgradable to: 2.34-0.1ubuntu9.3]
libffi7/focal,now 3.3-4 amd64 [installed]
libfido2-1/focal,now 1.3.1-1ubuntu2 amd64 [installed,automatic]
libfile-fcntllock-perl/focal,now 0.22-3build4 amd64 [installed,automatic]
libgcc-8-dev/focal,now 8.4.0-3ubuntu2 amd64 [installed,automatic]
libgcc-9-dev/focal-updates,focal-security,now 9.4.0-1ubuntu1~20.04.1 amd64 [installed,automatic]
libgcc-s1/focal-updates,focal-security,now 10.3.0-1ubuntu1~20.04 amd64 [installed]
libgcrypt20/focal-updates,focal-security,now 1.8.5-5ubuntu1.1 amd64 [installed]
libgdbm-compat4/focal,now 1.18.1-5 amd64 [installed,automatic]
libgdbm6/focal,now 1.18.1-5 amd64 [installed,automatic]
libgfortran-8-dev/focal,now 8.4.0-3ubuntu2 amd64 [installed,automatic]
libgfortran-9-dev/focal-updates,focal-security,now 9.4.0-1ubuntu1~20.04.1 amd64 [installed,automatic]
libgfortran5/focal-updates,focal-security,now 10.3.0-1ubuntu1~20.04 amd64 [installed,automatic]
libglib2.0-0/focal-updates,now 2.64.6-1~ubuntu20.04.4 amd64 [installed,automatic]
libglib2.0-data/focal-updates,now 2.64.6-1~ubuntu20.04.4 all [installed,automatic]
libgmp10/focal,now 2:6.2.0+dfsg-4 amd64 [installed]
libgnutls30/focal-updates,focal-security,now 3.6.13-2ubuntu1.6 amd64 [installed]
libgomp1/focal-updates,focal-security,now 10.3.0-1ubuntu1~20.04 amd64 [installed,automatic]
libgpg-error0/focal,now 1.37-1 amd64 [installed]
libgssapi-krb5-2/focal-updates,focal-security,now 1.17-6ubuntu4.1 amd64 [installed,automatic]
libgssapi3-heimdal/focal,now 7.7.0+dfsg-1ubuntu1 amd64 [installed,automatic]
libhcrypto4-heimdal/focal,now 7.7.0+dfsg-1ubuntu1 amd64 [installed,automatic]
libhdf5-openmpi-103/focal,now 1.10.4+repack-11ubuntu1 amd64 [installed,automatic]
libhdf5-openmpi-dev/focal,now 1.10.4+repack-11ubuntu1 amd64 [installed]
libheimbase1-heimdal/focal,now 7.7.0+dfsg-1ubuntu1 amd64 [installed,automatic]
libheimntlm0-heimdal/focal,now 7.7.0+dfsg-1ubuntu1 amd64 [installed,automatic]
libhogweed5/focal-updates,focal-security,now 3.5.1+really3.5.1-2ubuntu0.2 amd64 [installed]
libhwloc-dev/focal,now 2.1.0+dfsg-4 amd64 [installed,automatic]
libhwloc-plugins/focal,now 2.1.0+dfsg-4 amd64 [installed,automatic]
libhwloc15/focal,now 2.1.0+dfsg-4 amd64 [installed,automatic]
libhx509-5-heimdal/focal,now 7.7.0+dfsg-1ubuntu1 amd64 [installed,automatic]
libibverbs-dev/focal,now 28.0-1ubuntu1 amd64 [installed,automatic]
libibverbs1/focal,now 28.0-1ubuntu1 amd64 [installed,automatic]
libicu66/focal-updates,focal-security,now 66.1-2ubuntu2.1 amd64 [installed,automatic]
libidn2-0/focal,now 2.2.0-2 amd64 [installed]
libisl22/focal,now 0.22.1-1 amd64 [installed,automatic]
libitm1/focal-updates,focal-security,now 10.3.0-1ubuntu1~20.04 amd64 [installed,automatic]
libjpeg-dev/focal,now 8c-2ubuntu8 amd64 [installed,automatic]
libjpeg-turbo8-dev/focal-updates,focal-security,now 2.0.3-0ubuntu1.20.04.1 amd64 [installed,automatic]
libjpeg-turbo8/focal-updates,focal-security,now 2.0.3-0ubuntu1.20.04.1 amd64 [installed,automatic]
libjpeg8-dev/focal,now 8c-2ubuntu8 amd64 [installed,automatic]
libjpeg8/focal,now 8c-2ubuntu8 amd64 [installed,automatic]
libk5crypto3/focal-updates,focal-security,now 1.17-6ubuntu4.1 amd64 [installed,automatic]
libkeyutils1/focal,now 1.6-6ubuntu1 amd64 [installed,automatic]
libkrb5-26-heimdal/focal,now 7.7.0+dfsg-1ubuntu1 amd64 [installed,automatic]
libkrb5-3/focal-updates,focal-security,now 1.17-6ubuntu4.1 amd64 [installed,automatic]
libkrb5support0/focal-updates,focal-security,now 1.17-6ubuntu4.1 amd64 [installed,automatic]
libksba8/focal,now 1.3.5-2 amd64 [installed,automatic]
libldap-2.4-2/focal-updates,now 2.4.49+dfsg-2ubuntu1.8 amd64 [installed,automatic]
libldap-common/focal-updates,now 2.4.49+dfsg-2ubuntu1.8 all [installed,automatic]
liblocale-gettext-perl/focal,now 1.07-4 amd64 [installed,automatic]
liblsan0/focal-updates,focal-security,now 10.3.0-1ubuntu1~20.04 amd64 [installed,automatic]
libltdl-dev/focal,now 2.4.6-14 amd64 [installed,automatic]
libltdl7/focal,now 2.4.6-14 amd64 [installed,automatic]
liblz4-1/focal-updates,focal-security,now 1.9.2-2ubuntu0.20.04.1 amd64 [installed]
liblzma5/focal-updates,now 5.2.4-1ubuntu1 amd64 [installed]
libmagic-mgc/focal,now 1:5.38-4 amd64 [installed,automatic]
libmagic1/focal,now 1:5.38-4 amd64 [installed,automatic]
libmount1/now 2.34-0.1ubuntu9.1 amd64 [installed,upgradable to: 2.34-0.1ubuntu9.3]
libmpc3/focal,now 1.1.0-1 amd64 [installed,automatic]
libmpdec2/focal,now 2.4.2-3 amd64 [installed,automatic]
libmpfr6/focal,now 4.0.2-1 amd64 [installed,automatic]
libmpx2/focal,now 8.4.0-3ubuntu2 amd64 [installed,automatic]
libncurses6/focal,now 6.2-0ubuntu2 amd64 [installed]
libncursesw6/focal,now 6.2-0ubuntu2 amd64 [installed]
libnettle7/focal-updates,focal-security,now 3.5.1+really3.5.1-2ubuntu0.2 amd64 [installed]
libnghttp2-14/focal,now 1.40.0-1build1 amd64 [installed,automatic]
libnl-3-200/focal,now 3.4.0-1 amd64 [installed,automatic]
libnl-3-dev/focal,now 3.4.0-1 amd64 [installed,automatic]
libnl-route-3-200/focal,now 3.4.0-1 amd64 [installed,automatic]
libnl-route-3-dev/focal,now 3.4.0-1 amd64 [installed,automatic]
libnpth0/focal,now 1.6-1 amd64 [installed,automatic]
libnuma-dev/focal,now 2.0.12-1 amd64 [installed,automatic]
libnuma1/focal,now 2.0.12-1 amd64 [installed,automatic]
libopenmpi-dev/focal,now 4.0.3-0ubuntu1 amd64 [installed]
libopenmpi3/focal,now 4.0.3-0ubuntu1 amd64 [installed,automatic]
libp11-kit0/focal-updates,focal-security,now 0.23.20-1ubuntu0.1 amd64 [installed]
libpam-modules-bin/focal-updates,now 1.3.1-5ubuntu4.3 amd64 [installed]
libpam-modules/focal-updates,now 1.3.1-5ubuntu4.3 amd64 [installed]
libpam-runtime/focal-updates,now 1.3.1-5ubuntu4.3 all [installed]
libpam0g/focal-updates,now 1.3.1-5ubuntu4.3 amd64 [installed]
libpciaccess0/focal,now 0.16-0ubuntu1 amd64 [installed,automatic]
libpcre2-8-0/focal,now 10.34-7 amd64 [installed]
libpcre3/focal,now 2:8.39-12build1 amd64 [installed]
libperl5.30/focal-updates,focal-security,now 5.30.0-9ubuntu0.2 amd64 [installed,automatic]
libpmix2/focal,now 3.1.5-1 amd64 [installed,automatic]
libprocps8/focal-updates,now 2:3.3.16-1ubuntu2.3 amd64 [installed]
libpsl5/focal,now 0.21.0-1ubuntu1 amd64 [installed,automatic]
libpsm-infinipath1/focal,now 3.3+20.604758e7-6 amd64 [installed,automatic]
libpsm2-2/focal,now 11.2.86-1 amd64 [installed,automatic]
libpython3-dev/focal,now 3.8.2-0ubuntu2 amd64 [installed,automatic]
libpython3-stdlib/focal,now 3.8.2-0ubuntu2 amd64 [installed,automatic]
libpython3.8-dev/focal-updates,focal-security,now 3.8.10-0ubuntu1~20.04.2 amd64 [installed,automatic]
libpython3.8-minimal/focal-updates,focal-security,now 3.8.10-0ubuntu1~20.04.2 amd64 [installed,automatic]
libpython3.8-stdlib/focal-updates,focal-security,now 3.8.10-0ubuntu1~20.04.2 amd64 [installed,automatic]
libpython3.8/focal-updates,focal-security,now 3.8.10-0ubuntu1~20.04.2 amd64 [installed,automatic]
libquadmath0/focal-updates,focal-security,now 10.3.0-1ubuntu1~20.04 amd64 [installed,automatic]
librdmacm1/focal,now 28.0-1ubuntu1 amd64 [installed,automatic]
libreadline8/focal,now 8.0-4 amd64 [installed,automatic]
libroken18-heimdal/focal,now 7.7.0+dfsg-1ubuntu1 amd64 [installed,automatic]
librtmp1/focal,now 2.4+20151223.gitfa8646d.1-2build1 amd64 [installed,automatic]
libsasl2-2/focal-updates,focal-security,now 2.1.27+dfsg-2ubuntu0.1 amd64 [installed,automatic]
libsasl2-modules-db/focal-updates,focal-security,now 2.1.27+dfsg-2ubuntu0.1 amd64 [installed,automatic]
libsasl2-modules/focal-updates,focal-security,now 2.1.27+dfsg-2ubuntu0.1 amd64 [installed,automatic]
libseccomp2/now 2.5.1-1ubuntu1~20.04.1 amd64 [installed,upgradable to: 2.5.1-1ubuntu1~20.04.2]
libselinux1/focal,now 3.0-1build2 amd64 [installed]
libsemanage-common/focal,now 3.0-1build2 all [installed]
libsemanage1/focal,now 3.0-1build2 amd64 [installed]
libsepol1/focal,now 3.0-1 amd64 [installed]
libsigsegv2/focal,now 2.12-2 amd64 [installed,automatic]
libsmartcols1/now 2.34-0.1ubuntu9.1 amd64 [installed,upgradable to: 2.34-0.1ubuntu9.3]
libsqlite3-0/focal-updates,focal-security,now 3.31.1-4ubuntu0.2 amd64 [installed,automatic]
[…]
