### 2022-03-24 Cascella-Group-UiO/HyMD#160 @mortele
After all, there is nothing magical about this `1e-13` number I picked out of thin air.
