### 2022-03-24 Cascella-Group-UiO/HyMD#160 @mortele
> Running the mpi-enabled pytest script seems to pass but it skips several things:

This is expected as the pytest-mpi thing is setup to only run the specific tests tagged with
```python
@pytest.mark.mpi()
```
and skip everything else.
