### 2022-04-01 303bb7e
Remove deprecation warning
