+++ hymd/main.py
@@ -145,7 +145,7 @@ def main():
     dihedral_energy = 0.0
     kinetic_energy = 0.0
     field_q_energy = 0.0
-    plumed_energy = 0.0
+    plumed_bias = 0.0
 
     # Ignore numpy numpy.VisibleDeprecationWarning: Creating an ndarray from
     # ragged nested sequences until it is fixed in pmesh
@@ -284,6 +284,31 @@ def main():
             config, phi_q_fourier, elec_energy_field, field_q_energy,
             comm=comm,
         )
+
+    # configure PLUMED
+    if args.plumed:
+        plumed_obj = plumed.Plumed()
+
+        plumed_version = np.zeros(1, dtype=np.intc)
+        plumed_obj.cmd("getApiVersion", plumed_version)
+        assert plumed_version[0] > 3, "HyMD requires a PLUMED API > 3"
+
+        plumed_obj.cmd("setMDEngine", "HyMD")
+
+        plumed_obj.cmd("setMPIComm", comm)
+        plumed_obj.cmd("setPlumedDat", args.plumed)
+        plumed_obj.cmd("setLogFile", args.plumed_outfile)
+        plumed_obj.cmd("setTimestep", config.respa_inner * config.time_step)
+
+        plumed_obj.cmd("setNatoms", config.n_particles)
+        plumed_obj.cmd("setKbT", 
+            config.gas_constant * config.target_temperature
+        )
+        # plumed_obj.cmd("setNoVirial")
+        plumed_obj.cmd("init")
+
+        # set PLUMED initial forces
+        plumed_forces = np.zeros_like(positions)
 
     if molecules_flag:
         if not (args.disable_bonds
@@ -452,28 +477,6 @@ def main():
             comm=comm,
         )
 
-    # configure PLUMED
-    if args.plumed:
-        plumed_obj = plumed.Plumed()
-
-        plumed_version = np.zeros(1, dtype=np.intc)
-        plumed_obj.cmd("getApiVersion", plumed_version)
-        assert plumed_version[0] > 3, "HyMD requires a PLUMED API > 3"
-
-        plumed_obj.cmd("setMDEngine", "HyMD")
-
-        plumed_obj.cmd("setMPIComm", comm)
-        plumed_obj.cmd("setPlumedDat", args.plumed)
-        plumed_obj.cmd("setLogFile", args.plumed_outfile)
-        plumed_obj.cmd("setTimestep", config.respa_inner * config.time_step)
-
-        plumed_obj.cmd("setNatoms", config.n_particles)
-        plumed_obj.cmd("setKbT", 
-            config.gas_constant * config.target_temperature
-        )
-        plumed_obj.cmd("setNoVirial")
-        plumed_obj.cmd("init")
-
     if rank == 0:
         loop_start_time = datetime.datetime.now()
         last_step_time = datetime.datetime.now()
@@ -524,41 +527,12 @@ def main():
             velocities, field_forces / config.mass,
             config.respa_inner * config.time_step,
         )
-        # if args.plumed:
-        #     plumed_forces = np.zeros_like(positions)
-        #     if not charges_flag:
-        #         charges = np.zeros_like(indices, dtype=np.double)
-        #     bias = np.zeros(1, float)
-        #     masses = np.full_like(indices, config.mass, dtype=np.double)
-        #     box = np.diag(config.box_size)
 
-        #     plumed_obj.cmd("setAtomsNlocal", indices.shape[0]);
-        #     plumed_obj.cmd("setAtomsGatindex", indices);
-        #     plumed_obj.cmd("setStep", step*config.respa_inner)
-        #     plumed_obj.cmd("setForces", plumed_forces)
[…]
