### 2022-05-02 dd2c1a3
Changes unpacking of domain_decomposition. Fixes #174.
