### 2022-05-02 96cccb5
Disabling passing the virial to PLUMED for now.
