### 2022-05-03 4c26b3a
Moved PLUMED wrapper to a class.
