+++ hymd/main.py
@@ -19,6 +19,7 @@ from .thermostat import (csvr_thermostat, cancel_com_momentum,
                          generate_initial_velocities)
 from .force import dipole_forces_redistribution, prepare_bonds
 from .integrator import integrate_velocity, integrate_position
+from .plumed import PlumedBias
 
 # these look valuable to learn about MPI + PLUMED
 # https://groups.google.com/g/plumed-users/c/YTFZj3r1DHQ/m/HfRb9eeaCQAJ
@@ -30,6 +31,10 @@ from .integrator import integrate_velocity, integrate_position
 # http://openpathsampling.org/latest/_modules/openpathsampling/collectivevariables/plumed_wrapper.html
 # https://github.com/plumed/plumed2/tree/master/python/test
 # https://github.com/lab-cosmo/i-pi-dev_archive/blob/master/ipi/engine/forcefields.py
+
+# about energy as CV
+# https://github.com/plumed/plumed2/pull/486 # if energy is needed, scales forces
+
 
 def main():
     """Main simulation driver
@@ -68,9 +73,6 @@ def main():
         from .force import (
             compute_dihedral_forces__fortran as compute_dihedral_forces
         )
-
-    if args.plumed:
-        import plumed
 
     driver = "mpio" if not args.disable_mpio else None
     _kwargs = {"driver": driver, "comm": comm} if not args.disable_mpio else {}
@@ -285,28 +287,6 @@ def main():
             comm=comm,
         )
 
-    # configure PLUMED
-    if args.plumed:
-        plumed_obj = plumed.Plumed()
-
-        plumed_version = np.zeros(1, dtype=np.intc)
-        plumed_obj.cmd("getApiVersion", plumed_version)
-        assert plumed_version[0] > 3, "HyMD requires a PLUMED API > 3"
-
-        plumed_obj.cmd("setMDEngine", "HyMD")
-
-        plumed_obj.cmd("setMPIComm", comm)
-        plumed_obj.cmd("setPlumedDat", args.plumed)
-        plumed_obj.cmd("setLogFile", args.plumed_outfile)
-        plumed_obj.cmd("setTimestep", config.respa_inner * config.time_step)
-
-        plumed_obj.cmd("setNatoms", config.n_particles)
-        plumed_obj.cmd("setKbT", 
-            config.gas_constant * config.target_temperature
-        )
-        plumed_obj.cmd("setNoVirial")
-        plumed_obj.cmd("init")
-
     if molecules_flag:
         if not (args.disable_bonds
                 and args.disable_angle_bonds
@@ -474,6 +454,39 @@ def main():
             comm=comm,
         )
 
+    # initialize PLUMED and get initial forces and bias
+    if args.plumed:
+        plumed = PlumedBias(
+            config, 
+            args.plumed, 
+            args.plumed_outfile, 
+            comm
+        )
+
+        current_forces = (bond_forces + angle_forces + dihedral_forces
+                        + field_forces + elec_forces
+                        + reconstructed_forces)
+
+        # pass info and check if PLUMED also needs the energy
+        needs_energy = plumed.prepare(
+            0,
+            current_forces,
+            positions,
+            indices,
+            config,
+            (
+                np.zeros_like(indices, dtype=np.double) 
+                if not charges_flag else charges
+            )
+        )
+
+        # PLUMED likes the energy per rank
+        poteng = (field_energy + bond_energy + angle_energy
+                 + dihedral_energy + field_q_energy) / size
+
+        # get the forces and bias from PLUMED
+        plumed_forces, plumed_bias = plumed.calc(current_forces, poteng)
+
     if rank == 0:
         loop_start_time = datetime.datetime.now()
         last_step_time = datetime.datetime.now()
[…]
