### 2022-05-04 50eda3f
Fixed #177 and improves error message for old PLUMED versions.
