+++ hymd/file_io.py
@@ -109,7 +109,7 @@ def setup_time_dependent_element(
 def store_static(
     h5md, rank_range, names, types, indices, config, bonds_2_atom1,
     bonds_2_atom2, molecules=None, velocity_out=False, force_out=False,
-    charges=False, comm=MPI.COMM_WORLD,
+    charges=False, plumed_out=False, comm=MPI.COMM_WORLD,
 ):
     """Outputs all static time-independent quantities to the HDF5 output file
 
@@ -143,6 +143,8 @@ def store_static(
         If :code:`True`, forces are written to output HDF5 file.
     charges : (N,) numpy.ndarray
         Array of particle charge values for :code:`N` particles.
+    plumed_out : bool, optional
+        If :code:`True`, PLUMED bias is written to output HDF5 file.
     comm : mpi4py.Comm
         MPI communicator to use for rank commuication.
 
@@ -331,6 +333,15 @@ def store_static(
         ) = setup_time_dependent_element(
             "field_q_energy", h5md.observables, n_frames, (1,), dtype, units="kJ mol-1"  # noqa: E501
         )  # <-------- xinmeng
+    if plumed_out is not False:
+        (
+            _,
+            h5md.plumed_bias_step,
+            h5md.plumed_bias_time,
+            h5md.plumed_bias,
+        ) = setup_time_dependent_element(
+            "plumed_bias", h5md.observables, n_frames, (1,), dtype, units="kJ mol-1"  # noqa: E501
+        )
     (
         _,
         h5md.total_momentum_step,
@@ -440,9 +451,9 @@ def store_static(
 def store_data(
     h5md, step, frame, indices, positions, velocities, forces, box_size,
     temperature, kinetic_energy, bond2_energy, bond3_energy, bond4_energy,
-    field_energy, field_q_energy, time_step, config, velocity_out=False,
-    force_out=False, charge_out=False, dump_per_particle=False,
-    comm=MPI.COMM_WORLD,
+    field_energy, field_q_energy, plumed_bias, time_step, config, 
+    velocity_out=False, force_out=False, charge_out=False, plumed_out=False,
+    dump_per_particle=False, comm=MPI.COMM_WORLD,
 ):
     """Writes time-step data to HDF5 output file
 
@@ -479,8 +490,10 @@ def store_data(
         Calculated instantaneous dihedral four-particle torsion energy.
     field_energy : float
         Calculated instantaneous particle-field energy.
-    field_energy_q : float
+    field_q_energy : float
         Calculated instantaneous electrostatic energy.
+    plumed_bias : float
+        PLUMED instantaneous bias energy.
     time_step : float
         Value of the time step.
     config : Config
@@ -492,6 +505,8 @@ def store_data(
     charge_out : bool, optional
         If :code:`True`, electrostatic energies are written to the output
         HDF5 file.
+    plumed_out : bool, optional
+        If :code:`True`, PLUMED bias is written to the output HDF5 file.
     dump_per_particle : bool, optional
         If :code:`True`, all quantities are written **per particle**.
     comm : mpi4py.Comm
@@ -545,6 +560,9 @@ def store_data(
     if charge_out:
         h5md.field_q_energy_step[frame] = step
         h5md.field_q_energy_time[frame] = step * time_step
+    if plumed_out:
+        h5md.plumed_bias_step[frame] = step
+        h5md.plumed_bias_time[frame] = step * time_step
 
     ind_sort = np.argsort(indices)
     h5md.positions[frame, indices[ind_sort]] = positions[ind_sort]
@@ -555,6 +573,8 @@ def store_data(
[…]
