### 2022-05-09 b3b72ec
Fixed bug getting positions in step==0.
