### 2022-06-08 aabf39e
Revert "Fixes pytest warnings when passing None to pytest.warns()."
This reverts commit 1860943c6e895a48476b8460a0ab721988975fc4.
