+++ test/test_input_parser.py
@@ -1,5 +1,4 @@
 import pytest
-import warnings
 import re
 import io
 import logging
@@ -47,32 +46,24 @@ def test_input_parser_file_check_n_particles(config_toml, caplog):
     indices_take = indices_slices[MPI.COMM_WORLD.Get_rank()]
 
     if MPI.COMM_WORLD.Get_rank() == 0:
-        with pytest.warns(Warning) as recorded_warning:
-            config_ = check_n_particles(config, indices_take)
-            assert config_.n_particles == 10001
-            if MPI.COMM_WORLD.Get_rank() == 0:
-                message = recorded_warning[0].message.args[0]
-                log = caplog.text
-                assert all([(s in message) for s in ("10000", "not", "10001")])
-                assert all([(s in log) for s in ("10000", "not", "10001")])
+        warning = Warning
     else:
-        with warnings.catch_warnings() as recorded_warning:
-            warnings.simplefilter("error")
-            config_ = check_n_particles(config, indices_take)
-            assert config_.n_particles == 10001
-            if MPI.COMM_WORLD.Get_rank() == 0:
-                message = recorded_warning[0].message.args[0]
-                log = caplog.text
-                assert all([(s in message) for s in ("10000", "not", "10001")])
-                assert all([(s in log) for s in ("10000", "not", "10001")])
+        warning = None
+    with pytest.warns(warning) as recorded_warning:
+        config_ = check_n_particles(config, indices_take)
+        assert config_.n_particles == 10001
+        if MPI.COMM_WORLD.Get_rank() == 0:
+            message = recorded_warning[0].message.args[0]
+            log = caplog.text
+            assert all([(s in message) for s in ("10000", "not", "10001")])
+            assert all([(s in log) for s in ("10000", "not", "10001")])
 
     caplog.clear()
     indices = np.empty((n_particles_config,))
     indices_slices = np.array_split(indices, MPI.COMM_WORLD.Get_size())
     indices_take = indices_slices[MPI.COMM_WORLD.Get_rank()]
 
-    with warnings.catch_warnings() as recorded_warning:
-        warnings.simplefilter("error")
+    with pytest.warns(None) as recorded_warning:
         check_n_particles(config, indices_take)
         assert not recorded_warning
         assert not caplog.text
@@ -114,27 +105,18 @@ def test_input_parser_check_optionals(config_toml, caplog):
     )
 
     if MPI.COMM_WORLD.Get_rank() == 0:
-        with pytest.warns(Warning) as recorded_warning:
-            config = parse_config_toml(config_toml_wrong_max_molecule_size)
-            config = check_max_molecule_size(config)
-            assert config.max_molecule_size == 201
-            if MPI.COMM_WORLD.Get_rank() == 0:
-                message = recorded_warning[0].message.args[0]
-                log = caplog.text
-                assert all([(s in message) for s in ("must be", "integer", "201")])
-                assert all([(s in log) for s in ("must be", "integer", "201")])
+        warning = Warning
     else:
-        with warnings.catch_warnings() as recorded_warning:
-            warnings.simplefilter("error")
-            config = parse_config_toml(config_toml_wrong_max_molecule_size)
-            config = check_max_molecule_size(config)
-            assert config.max_molecule_size == 201
-            if MPI.COMM_WORLD.Get_rank() == 0:
-                message = recorded_warning[0].message.args[0]
-                log = caplog.text
-                assert all([(s in message) for s in ("must be", "integer", "201")])
-                assert all([(s in log) for s in ("must be", "integer", "201")])
-
+        warning = None
+    with pytest.warns(warning) as recorded_warning:
[…]
