### 2022-06-08 10b31c9
Fixes pytest warnings when passing None to pytest.warns().
