### 2022-06-08 d8ebff9
Cleaned up useless ifs.
