### 2022-06-08 60b258d
Added tests to PlumedBias.
