### 2022-06-08 e4611fa
Stores files in tmp_path.
