+++ .github/workflows/ci.yml
@@ -23,6 +23,12 @@ jobs:
         restore-keys: |
           ${{ runner.os }}-pip-
           ${{ runner.os }}-
+    - name: Cache PLUMED
+      uses: actions/cache@v2
+      with:
+        path: ~/.ccache
+        key: ccache-reset1-linux-hash-${{ github.sha }}
+        restore-keys: ccache-reset1-linux-hash-
     - name: Install dependencies
       run: |
         sudo apt-get update
@@ -32,14 +38,17 @@ jobs:
       run: |
         mkdir -p $HOME/opt/
         sudo apt-get install -y build-essential
+        sudo apt-get install -y ccache
         sudo apt-get install -y openmpi-bin libopenmpi-dev git
         git clone https://github.com/plumed/plumed2.git
         cd plumed2
         echo "CC=mpicc" >> $GITHUB_ENV
         echo "CXX=mpic++" >> $GITHUB_ENV
-        ./configure --prefix=$HOME/opt
+        ccache -s -M 100M
+        ./configure CXX="ccache $CXX" --prefix=$HOME/opt
         make -j 4
         make install
+        ccache -s -M 100M
         cd ../
     - name: Install python dependencies
       run: |
