### 2022-06-08 e2da4fc
Fix ccache for PLUMED in ci.
