+++ .github/workflows/ci.yml
@@ -32,6 +32,7 @@ jobs:
     - name: Install dependencies
       run: |
         sudo apt-get update
+        sudo apt-get install -y ccache
         sudo apt-get install -y python3-mpi4py python3-numpy
         sudo apt-get install -y libhdf5-openmpi-dev libfftw3-3 libfftw3-dev
     - name: Install PLUMED
@@ -39,7 +40,6 @@ jobs:
         mkdir -p $HOME/opt/
         sudo apt-get install -y build-essential
         sudo apt-get install -y openmpi-bin libopenmpi-dev git
-        sudo apt-get install -y ccache
         git clone https://github.com/plumed/plumed2.git
         cd plumed2
         echo "CC=mpicc" >> $GITHUB_ENV
