### 2022-06-08 8ef40e6
Symbolic link as in plumed-nest.
