+++ .github/workflows/ci.yml
@@ -32,7 +32,6 @@ jobs:
     - name: Install dependencies
       run: |
         sudo apt-get update
-        sudo apt-get install -y ccache
         sudo apt-get install -y python3-mpi4py python3-numpy
         sudo apt-get install -y libhdf5-openmpi-dev libfftw3-3 libfftw3-dev
     - name: Install PLUMED
@@ -40,12 +39,15 @@ jobs:
         mkdir -p $HOME/opt/
         sudo apt-get install -y build-essential
         sudo apt-get install -y openmpi-bin libopenmpi-dev git
+        sudo apt-get install -y ccache
         git clone https://github.com/plumed/plumed2.git
+        sudo ln -s ccache /usr/local/bin/mpic++
+        export PATH=/usr/lib/ccache:${PATH}
         cd plumed2
         echo "CC=mpicc" >> $GITHUB_ENV
         echo "CXX=mpic++" >> $GITHUB_ENV
         ccache -s -M 100M
-        ./configure CXX="ccache $CXX" --disable-dependency-tracking --prefix=$HOME/opt
+        ./configure --disable-dependency-tracking --prefix=$HOME/opt
         make -j 4
         make install
         ccache -s -M 100M
