+++ .github/workflows/ci.yml
@@ -26,9 +26,11 @@ jobs:
     - name: Cache PLUMED
       uses: actions/cache@v2
       with:
-        path: ~/.ccache
-        key: ccache-reset1-linux-hash-${{ github.sha }}
-        restore-keys: ccache-reset1-linux-hash-
+        path: |
+          ~/opt
+          ~/.ccache
+        key: ccache-${{ runner.os }}-${{ github.sha }}
+        restore-keys: ccache-${{ runner.os }}-
     - name: Install dependencies
       run: |
         sudo apt-get update
@@ -40,17 +42,13 @@ jobs:
         sudo apt-get install -y build-essential
         sudo apt-get install -y openmpi-bin libopenmpi-dev git
         sudo apt-get install -y ccache
-        git clone https://github.com/plumed/plumed2.git
         sudo ln -s ccache /usr/local/bin/mpic++
         export PATH=/usr/lib/ccache:${PATH}
+        git clone https://github.com/plumed/plumed2.git
         cd plumed2
-        echo "CC=mpicc" >> $GITHUB_ENV
-        echo "CXX=mpic++" >> $GITHUB_ENV
-        ccache -s -M 100M
-        ./configure --disable-dependency-tracking --prefix=$HOME/opt
-        make -j 4
-        make install
-        ccache -s -M 100M
+        ccache -s
+        CXX="mpic++" .ci/install.plumed version=master
+        ccache -s
         cd ../
     - name: Install python dependencies
       run: |
