### 2022-06-08 55c9170
No need to ignore plumed2 in pytest now.
