### 2022-06-08 4b43a81
Print pytest output.
