### 2022-06-08 ccc5fd2
Add PLUMED_KERNEL to GITHUB_ENV.
