+++ .github/workflows/ci.yml
@@ -60,14 +60,13 @@ jobs:
         make pip
         python3 -m pip install .
         cd $current_path
-        ls -R $HOME/opt/
-        PLUMED_KERNEL=$HOME/opt/lib/libplumedKernel.so
     - name: Install package
       run: |
         python3 -m pip install .
     - name: Test with pytest
       run: |
         python3 -m pip install pytest pytest-cov pytest-mpi coverage
+        PLUMED_KERNEL=$HOME/opt/lib/libplumedKernel.so
         pytest --cov=./
         chmod +x pytest-mpi
         ./pytest-mpi -c -ca -n 2 -ns -oo -s
