### 2022-06-09 1862a23
Add log-cli option.
