### 2022-06-09 20a10a1
Try to capture more events and show the log.
