### 2022-06-09 1ebec49
Rebuild cache and catch pytest-mpi errors.
