+++ .github/workflows/ci.yml
@@ -29,8 +29,8 @@ jobs:
         path: |
           ~/opt
           ~/.ccache
-        key: ccache-${{ runner.os }}-${{ github.sha }}
-        restore-keys: ccache-${{ runner.os }}-
+        key: ccache-plumed-${{ runner.os }}-${{ github.sha }}
+        restore-keys: ccache-plumed-${{ runner.os }}-
     - name: Install dependencies
       run: |
         sudo apt-get update
@@ -41,6 +41,7 @@ jobs:
         mkdir -p $HOME/opt/
         sudo apt-get install -y build-essential
         sudo apt-get install -y openmpi-bin libopenmpi-dev git
+        sudo apt-get install gsl-bin libgsl0-dev libboost-serialization-dev
         sudo apt-get install -y ccache
         sudo ln -s ccache /usr/local/bin/mpic++
         export PATH=/usr/lib/ccache:${PATH}
