### 2022-06-09 0c6e245
Check to see if with cython the PLUMED compilation will already have python.
