+++ .github/workflows/ci.yml
@@ -29,13 +29,18 @@ jobs:
         path: |
           ~/opt
           ~/.ccache
-        key: ccache-plumed-${{ runner.os }}-${{ github.sha }}
-        restore-keys: ccache-plumed-${{ runner.os }}-
+        key: ccache-${{ runner.os }}-${{ github.sha }}
+        restore-keys: ccache-${{ runner.os }}-
     - name: Install dependencies
       run: |
         sudo apt-get update
         sudo apt-get install -y python3-mpi4py python3-numpy
         sudo apt-get install -y libhdf5-openmpi-dev libfftw3-3 libfftw3-dev
+    - name: Install python dependencies
+      run: |
+        python3 -m pip install --upgrade pip
+        python3 -m pip install --upgrade wheel numpy mpi4py cython
+        python3 -m pip install -r requirements.txt
     - name: Install PLUMED
       run: |
         mkdir -p $HOME/opt/
@@ -47,13 +52,8 @@ jobs:
         export PATH=/usr/lib/ccache:${PATH}
         git clone https://github.com/plumed/plumed2.git $HOME/plumed2
         ccache -s
-        CXX="mpic++" .ci/install.plumed version=master program_path=$HOME/plumed2
+        CXX="mpic++" PYTHON_BIN=python3 .ci/install.plumed version=master program_path=$HOME/plumed2
         ccache -s
-    - name: Install python dependencies
-      run: |
-        python3 -m pip install --upgrade pip
-        python3 -m pip install --upgrade wheel numpy mpi4py cython
-        python3 -m pip install -r requirements.txt
     - name: Install PLUMED python wrapper
       run: |
         current_path=$PWD
