### 2022-06-09 f972cd1
Add --log-cli option to pytest-mpi.
