### 2022-06-09 839592c
Improvements to error message.
