### 2022-06-09 a8e187a
Moved info about PLUMED_KERNEL to INFO.
