### 2022-06-09 767354e
Add exception when the PLUMED_KERNEL is not set.
