### 2022-07-28 a966fa0
Init port of bonded forces routines in C
