### 2022-08-04 e65b4b5
Move store data so first log has the right bias E.
