### 2022-08-05 d88a1b9
Update utils/center_group.py
Co-authored-by: Manuel Carrer <manuelca@kjemi.uio.no>
