### 2022-08-05 2da3ffd
Update utils/center_group.py
Co-authored-by: Manuel Carrer <manuelca@kjemi.uio.no>
