### 2022-08-05 5fa85c3
Update utils/center_group.py
Co-authored-by: Manuel Carrer <manuelca@kjemi.uio.no>
