### 2022-08-09 1fe46ec
Added check for PLUMED returned forces.
Also adds an option for extra verbosity when using PLUMED.
