### 2022-08-10 3c0f6e0
PLUMED will not return np.inf.
