+++ hymd/barostat_bussi.py
@@ -1,4 +1,22 @@
-# Bussy Barostat
+"""Implements the stochastic cell rescaling (SCR) barostat.
+This is a first-order barostat that samples the correct
+volume fluctuations through a suitable noise term.
+Scales the box and particle positions by a factor (see functions
+`isotropic` and `semiisotropic` for details) during simulation to
+simulate coupling to an external pressure bath set at a
+target pressure.
+
+The box and particle positions are scaled in the L and N directions according
+to the nature of the barostat (see functions `isotropic` and `semiisotropic`
+below by an amount :math:`\\alpha`.
+
+The updated system information is passed on to the pmesh objects.
+
+References
+----------
+Mattia Bernetti and Giovanni Bussi , "Pressure control using stochastic
+cell rescaling", J. Chem. Phys. 153, 114107 (2020)
+"""
 import numpy as np
 from mpi4py import MPI
 from dataclasses import dataclass
@@ -25,14 +43,29 @@ def isotropic(
         phi_transfer,
         phi_grad_lap_fourier,
         phi_grad_lap,
-        bond_forces,
-        angle_forces,
-        args,
         bond_pr,
         angle_pr,
         step,
         comm=MPI.COMM_WORLD
     ):
+    """
+    It calculates the scaling factor according to
+
+    .. math::
+    
+        \\log \\alpha' = - \\frac{n_b\,dt\,\\beta}{\\tau_p}(P^t - P) + \\sqrt{\\frac{2n_b^2k_BT\\beta\,dt}{\mathcal{V}\\tau_p}}dW
+
+    .. math::
+    
+        \\alpha = \\exp{\\frac{1}{3}\\log \\alpha'}
+    
+    where :math:`dt` is the outer rRESPA time-step, :math:`n_b` is the frequency
+    of barostat calls, :math:`\\tau_p` is the pressure coupling time constant,
+    :math:`\\beta` is the isothermal compressibility, :math:`P_{L,N}^t` and 
+    :math:`P_{L,N}` is the target and instantaneous internal pressure in the 
+    lateral (L) and normal (N) directions respectively. Convention: Cartesian
+    z-direction is considered normal.
+    """
     rank = comm.Get_rank()
     beta = 7.6 * 10**(-4) #bar^(-1) #isothermal compressibility of water
     change = False
@@ -56,9 +89,6 @@ def isotropic(
                 phi_transfer,
                 phi_grad_lap_fourier,
                 phi_grad_lap,
-                args,
-                bond_forces,
-                angle_forces,
                 positions,
                 bond_pr,
                 angle_pr,
@@ -104,14 +134,37 @@ def semiisotropic(
         phi_transfer,
         phi_grad_lap_fourier,
         phi_grad_lap,
-        bond_forces,
-        angle_forces,
-        args,
         bond_pr,
         angle_pr,
         step,
         comm=MPI.COMM_WORLD
     ):
+    """
+    It calculates the scaling factor according to
+
+    .. math::
+    
+        \\log \\alpha'_{L} = - \\frac{2n_b\,dt\,\\beta}{3\\tau_p}(P_L^t - P_L - \\frac{\\gamma}{L_z}) + \\sqrt{\\frac{4n_b^2k_BT\\beta\,dt}{3\mathcal{V}\\tau_p}}dW_L
+
+    .. math::
+    
+        \\alpha_L = \\exp{\\frac{1}{2}\\log \\alpha'_L}
+    
+    .. math::
+    
+        \\log \\alpha'_{N} = - \\frac{n_b\,dt\,\\beta}{3\\tau_p}(P_N^t - P_N) + \\sqrt{\\frac{2n_b^2k_BT\\beta\,dt}{3\mathcal{V}\\tau_p}}dW_N
+
+    .. math::
+    
[…]
