+++ hymd/pressure.py
@@ -16,14 +16,79 @@ def comp_pressure(
         phi_transfer,
         phi_grad_lap_fourier,
         phi_grad_lap,
-        args,
-        bond_forces,
-        angle_forces,
         positions,
         bond_pr,
         angle_pr,
         comm=MPI.COMM_WORLD
 ):
+    """
+    Computes total internal pressure of the system.
+    Kinetic pressure is trivially calculated from the kinetic energy.
+    Already computed bond and angle pressure terms are inserted into the
+    total internal pressure.
+    The field pressure equation is implemented.
+
+    Parameters
+    ----------
+    phi : list[pmesh.pm.RealField], (M,)
+        Pmesh :code:`RealField` objects containing discretized particle number
+        density values on the computational grid; one for each particle type
+        :code:`M`. Pre-allocated, but empty; any values in this field are discarded.
+        Changed in-place. Local for each MPI rank--the full computaional grid
+        is represented by the collective fields of all MPI ranks.
+    phi_gradient : list[pmesh.pm.RealField], (M, 3)
+        Like phi, but containing the gradient of particle number densities.
+        Needed only for vestigial squaregradient term.
+    hamiltonian : Hamiltonian
+        Particle-field interaction energy handler object. Defines the
+        grid-independent filtering function, :math:`H`.
+    velocities : (N, D) numpy.ndarray
+        Array of velocities of N particles in D dimensions.
+    config : Config
+        Configuration dataclass containing simulation metadata and parameters.
+    phi_fourier : list[pmesh.pm.ComplexField], (M,)
+        Pmesh :code:`ComplexField` objects containing discretized particle
+        number density values in reciprocal space on the computational grid;
+        one for each particle type. Pre-allocated, but empty; any values in
+        this field are discarded Changed in-place. Local for each MPI rank--the
+        full computaional grid is represented by the collective fields of all
+        MPI ranks.
+    phi_laplacian : list[pmesh.pm.RealField], (M, 3)
+        Like phi, but containing the laplacian of particle number densities.
+    phi_transfer : list[pmesh.pm.ComplexField], (3,)
+        Like phi_fourier, used as an intermediary to perform FFT operations
+        to obtain the gradient or laplacian of particle number densities.
+    phi_grad_lap_fourier : list[pmesh.pm.ComplexField], (3,)
+        Like phi_fourier, used as a second intermediary after phi_transfer
+        to perform FFT operations to obtain gradient of laplacian of particle
+        number densities. Needed only for vestigial squaregradient term.
+    phi_grad_lap : list[pmesh.pm.RealField], (M, 3, 3)
+        Like phi, to obtain the gradient of laplacian in all 3x3 directions.
+        Needed only for vestigial squaregradient term.
+    positions : (N,D) numpy.ndarray
+        Array of positions for :code:`N` particles in :code:`D` dimensions.
+        Local for each MPI rank.
+    bond_pr : (3,) numpy.ndarray 
+        Total bond pressure due all two-particle bonds.
+    angle_pr : (3,) numpy.ndarray
+        Total angle pressure due all three-particle bonds
+    comm : MPI.Intracomm, optional
+        MPI communicator to use for rank commuication. Defaults to
+        MPI.COMM_WORLD.
+
+    Returns
+    -------
+    pressure : (25,) numpy.ndarray
+        Pressure contributions from various energy terms.
+        0: due to kinetic energy
+        1-5: due to field energy
+        6-12: due to vestigial squaregradient (defaults to 0)
+        13-15: due to two-particle bonded terms
[…]
