### 2022-08-10 8e2f48b
Add test to version logging.
