### 2022-08-11 cada0e4
Added tests for Logger.
