### 2022-08-11 e65c1fb
Fix redundant assertions.
