### 2022-08-11 876510d
Prepare configure_runtime to unit tests.
