+++ hymd/configure_runtime.py
@@ -12,11 +12,13 @@ from .logger import Logger
 from .input_parser import read_config_toml, parse_config_toml
 
 
-def configure_runtime(comm):
+def configure_runtime(args_in, comm):
     """Parse command line arguments and configuration file
 
     Parameters
     ----------
+    args_in : list
+        List with arguments to be processed.
     comm : mpi4py.Comm
         MPI communicator to use for rank commuication.
 
@@ -98,7 +100,7 @@ def configure_runtime(comm):
         "config", help="Config .py or .toml input configuration script"
     )
     ap.add_argument("input", help="input.hdf5")
-    args = ap.parse_args()
+    args = ap.parse_args(args_in)
 
     if comm.rank == 0:
         os.makedirs(args.destdir, exist_ok=True)
