### 2022-08-11 ae064b9
Add unit tests for configure_runtime.
