### 2022-08-11 08aed2d
Fixes and tests to check_n_print.
