### 2022-08-12 361b91c
Refactor and test get_hamiltonian.
