### 2022-08-12 63b2f3b
Fixes to checks and more unit tests.
