+++ hymd/input_parser.py
@@ -783,6 +783,8 @@ def check_integrator(config, comm=MPI.COMM_WORLD):
 
 
 def check_hamiltonian(config, comm=MPI.COMM_WORLD):
+    valid_hamiltonians = ["defaultnochi", "defaultwithchi", "squaredphi"]
+
     if config.hamiltonian is None:
         if len(config.chi) == 0:
             warn_str = (
@@ -803,21 +805,29 @@ def check_hamiltonian(config, comm=MPI.COMM_WORLD):
             if comm.Get_rank() == 0:
                 warnings.warn(warn_str)
             config.hamiltonian = "DefaultWithChi"
+    elif config.hamiltonian.lower() not in valid_hamiltonians:
+        err_str = (
+            f"The specified Hamiltonian {config.hamiltonian} was not "
+            f"recognized as a valid Hamiltonian."
+        )
+        Logger.rank0.log(logging.ERROR, err_str)
+        raise NotImplementedError(err_str)
+
     return config
 
 
 def check_n_print(config, comm=MPI.COMM_WORLD):
     if (not isinstance(config.n_print, int) and
-        not isinstance(config.n_print, float)):
+        not isinstance(config.n_print, float) and
+        config.n_print is not None):
         err_str = (
             f"invalid value for n_print ({config.n_print})"
         )
         Logger.rank0.log(logging.ERROR, err_str)
         raise RuntimeError(err_str)        
     
-    if config.n_print is None or config.n_print < 0:
+    if config.n_print is None or config.n_print <= 0:
         config.n_print = False
-        return config
     elif not isinstance(config.n_print, int):
         if isinstance(config.n_print, float):
             warn_str = (
@@ -896,38 +906,33 @@ def check_start_and_target_temperature(config, comm=MPI.COMM_WORLD):
 
 
 def check_mass(config, comm=MPI.COMM_WORLD):
-    if config.mass is not None:
-        try:
-            config.mass = float(config.mass)
-            return config
-        except ValueError:
-            pass
-
     if config.mass is None:
         info_str = "no mass specified, defaulting to 72.0"
         config.mass = 72.0
         Logger.rank0.log(logging.INFO, info_str)
-    elif isinstance(config.mass, int) or isinstance(config.mass, float):
+    elif not isinstance(config.mass, int) and not isinstance(config.mass, float):
         err_str = (
             f"specified mass is invalid type {config.mass}, "
             f"({type(config.mass)})"
         )
         Logger.rank0.log(logging.ERROR, err_str)
-        if comm.Get_rank() == 0:
-            raise TypeError(err_str)
+        raise TypeError(err_str)
     elif config.mass < 0:
         err_str = f"invalid mass specified, {config.mass}"
         Logger.rank0.log(logging.ERROR, err_str)
-        if comm.Get_rank() == 0:
-            raise TypeError(err_str)
+        raise ValueError(err_str)
+    else:
+        config.mass = float(config.mass)
+
     return config
 
 
 def check_domain_decomposition(config, comm=MPI.COMM_WORLD):
-    if config.domain_decomposition is None:
-        config.domain_decomposition = False
     dd = config.domain_decomposition
-    assert isinstance(dd, int) or isinstance(dd, float) or (dd is None)
+    if dd is None or dd is False:
+        config.domain_decomposition = False
+        return config
+
     if isinstance(dd, int):
         if dd < 0:
             warn_str = "negative domain_decomposition specified, using False"
@@ -935,16 +940,25 @@ def check_domain_decomposition(config, comm=MPI.COMM_WORLD):
             Logger.rank0.log(logging.WARNING, warn_str)
             if comm.Get_rank() == 0:
                 warnings.warn(warn_str)
-    if isinstance(dd, float):
[…]
