+++ test/conftest.py
@@ -254,6 +254,7 @@ def config_toml(mpi_file_name):
     integrator = "respa"
     respa_inner = 5
     domain_decomposition = false
+    hamiltonian = "DefaultWithChi"
     start_temperature = false
     target_temperature = 323
     tau = 0.7
