+++ docs/doc_pages/installation.rst
@@ -69,6 +69,12 @@ Dependencies
            brew install pkg-config  # Install pkg-config, required for h5py install
            brew install curl wget
 
+.. warning::
+    There might be memory leaks when using HyMD with OpenMPI 4.1.1.
+    Therefore, using a newer version of OpenMPI is recommended.
+    See `Issue #186 <https://github.com/Cascella-Group-UiO/HyMD/issues/186>` for more details.
+
+
 **Installing Python dependencies** may be done by
 
 .. tabs::
