+++ test/test_file_io.py
@@ -1,0 +1,180 @@
+import pytest
+from mpi4py import MPI
+import numpy as np
+import collections
+import itertools
+import h5py
+from hymd.file_io import (
+    OutDataset, setup_time_dependent_element, store_static, store_data,
+    distribute_input
+)
+from hymd.input_parser import Config, parse_config_toml, _setup_type_to_name_map
+from hymd.force import prepare_bonds
+
+def test_OutDataset(config_toml, tmp_path):
+    _, config_toml_str = config_toml
+    config = parse_config_toml(config_toml_str)
+
+    out = OutDataset(tmp_path, config)
+    assert hasattr(out, "config")
+    assert hasattr(out, "disable_mpio")
+    assert not out.disable_mpio
+    assert hasattr(out, "float_dtype")
+    assert out.float_dtype == "float32"
+    assert hasattr(out, "file")
+    assert isinstance(out.file, h5py.File)
+    assert out.file # check if open
+
+    out.flush()
+    out.close_file()
+    assert not out.file # check if closed
+
+    out = OutDataset(tmp_path, config, double_out=True)
+    assert out.float_dtype == "float64"
+    out.close_file()
+
+    out = OutDataset(tmp_path, config, disable_mpio=True)
+    assert out.disable_mpio
+    assert isinstance(out.file, h5py.File)
+    assert out.file # check if open
+    out.close_file()
+    assert not out.file
+
+
+def test_setup_time_dependent_element(config_toml, tmp_path):
+    _, config_toml_str = config_toml
+    config = parse_config_toml(config_toml_str)
+
+    out = OutDataset(tmp_path, config)
+    test_group = out.file.create_group("/test")
+
+    (
+        group, step, time, value
+    ) = setup_time_dependent_element(
+        "position",
+        test_group,
+        1,
+        (config.n_particles, 3),
+        "float32",
+        units="nm"
+    )
+
+    assert isinstance(group, h5py.Group)
+    assert group.name == "/test/position"
+    assert group.attrs["units"] == "nm"
+
+    assert isinstance(step, h5py.Dataset)
+    assert isinstance(time, h5py.Dataset)
+    assert isinstance(value, h5py.Dataset)
+
+    out.close_file()
+
+
+@pytest.mark.mpi()
+def test_distribute_input():
+    indices = np.arange(0, 30)
+    molecules = np.array([
+        0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1,
+        1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 3, 4, 5, 6
+    ], dtype=int)
+    comm = MPI.COMM_WORLD
+    rank = comm.Get_rank()
+    size = comm.Get_size()
+
+    in_file = {'indices': indices}
+    rank_range, molecules_flag = distribute_input(
+        in_file, rank, size,  None, comm=comm
+    )
+    assert not molecules_flag
+    gather_range = comm.allgather(rank_range)
+    gather_range = list(itertools.chain.from_iterable(gather_range))
+    assert np.array_equal(gather_range, indices)
+
+    in_file = {'indices': indices, 'molecules': molecules}
+    rank_range, molecules_flag = distribute_input(
+        in_file, rank, size, 30, max_molecule_size=100, comm=comm
+    )
+    assert molecules_flag
+    gather_range = comm.allgather(rank_range)
+    gather_range = list(itertools.chain.from_iterable(gather_range))
+    assert np.array_equal(gather_range, indices)
+    if rank == 0:
+        assert len(rank_range) == 25
+
+
+@pytest.mark.mpi()
[…]
