+++ requirements.txt
@@ -8,3 +8,4 @@ pfft-python
 pmesh
 sympy
 tomli
+setuptools < 60.0
