+++ .github/workflows/ci.yml
@@ -26,7 +26,7 @@ jobs:
     - name: Install dependencies
       run: |
         sudo apt-get update
-        sudo apt-get install -y python3-mpi4py python3-numpy
+        sudo apt-get install -y python3-mpi4py python3-numpy python3-distutils
         sudo apt-get install -y libhdf5-openmpi-dev libfftw3-3 libfftw3-dev
     - name: Install python dependencies
       run: |
