### 2022-08-15 3d6d554
Force setuptools<60 during build.
