+++ requirements.txt
@@ -8,4 +8,3 @@ pfft-python
 pmesh
 sympy
 tomli
-setuptools < 60.0
