### 2022-08-15 29b1ed8
Add MPI enabled h5py to CI.
