+++ .github/workflows/ci.yml
@@ -32,6 +32,7 @@ jobs:
       run: |
         python3 -m pip install --upgrade pip
         python3 -m pip install --upgrade wheel numpy mpi4py cython
+        CC="mpicc" HDF5_MPI="ON" python3 -m pip install --no-binary=h5py h5py
         python3 -m pip install -r requirements.txt
     - name: Install package
       run: |
