### 2022-08-15 dfe8471
Fixes potential energy logging and store_data unit tests.
