+++ test/test_file_io.py
@@ -178,3 +178,123 @@ def test_store_static(molecules_with_solvent, tmp_path):
 
     out.close_file()
 
+
+@pytest.mark.mpi()
+def test_store_data(molecules_with_solvent, tmp_path):
+    indices, positions, molecules, velocities, bonds, names, types = molecules_with_solvent
+    box_size = np.array([10, 10, 10], dtype=np.float64)
+    config = Config(n_steps=100, n_print=1, time_step=0.03, box_size=box_size,
+                    mesh_size=[5, 5, 5], sigma=0.5, kappa=0.05,
+                    n_particles=len(indices), mass=72.)
+    Bond = collections.namedtuple(
+        "Bond", ["atom_1", "atom_2", "equilibrium", "strength"]
+    )
+    cbonds = (
+        Bond("A", "A", 0.27, 10000),
+        Bond("A", "B", 0.27, 10000),
+        Bond("A", "C", 0.27, 10000),
+        Bond("B", "B", 0.27, 10000),
+        Bond("B", "C", 0.27, 10000),
+    )
+    config.bonds = cbonds
+    config = _setup_type_to_name_map(config, names, types)
+
+    comm = MPI.COMM_WORLD
+    rank = comm.Get_rank()
+    size = comm.Get_size()
+
+    test_dir = comm.bcast(tmp_path, root=0)
+    out = OutDataset(test_dir, config, comm=comm)
+
+    # Test stub acting like a hdf5 file for distribute_input
+    in_file = {'molecules': molecules, 'indices': indices}
+    rank_range, molecules_flag = distribute_input(
+        in_file, rank, size,  config.n_particles, 6, comm=comm
+    )
+    positions_ = positions[rank_range]
+    molecules_ = molecules[rank_range]
+    indices_ = indices[rank_range]
+    velocities_ = velocities[rank_range]
+    bonds_ = bonds[rank_range]
+    types_ = types[rank_range]
+    names_ = names[rank_range]
+
+    bonds_prep = prepare_bonds(
+        molecules_, names_, bonds_, indices_, config
+    )
+    (
+        # two-particle bonds
+        bonds_2_atom1, bonds_2_atom2, bonds_2_equilibrium,
+        bonds_2_stength,
+        # three-particle bonds
+        bonds_3_atom1, bonds_3_atom2, bonds_3_atom3,
+        bonds_3_equilibrium, bonds_3_stength,
+        # four-particle bondsl
+        bonds_4_atom1, bonds_4_atom2, bonds_4_atom3, bonds_4_atom4,
+        bonds_4_coeff, bonds_4_type, bonds_4_last,
+    ) = bonds_prep
+
+
+    # call it to prepare OutDataset
+    store_static(
+        out, rank_range, names_, types_, indices_, config, bonds_2_atom1,
+        bonds_2_atom2, molecules=molecules_
+    )
+
+    forces = np.copy(positions_)
+
+    store_data(
+        out, 0, 0, indices_, positions_, velocities_, forces,
+        box_size, 300., 1., 2., 3., 4., 5., 6., 0.02, config
+    )
+
+    outdataset_step = [
+        out.positions_step,
+        out.total_energy_step,
+        out.potential_energy_step,
+        out.kinetc_energy_step,
+        out.bond_energy_step,
+        out.angle_energy_step,
+        out.dihedral_energy_step,
+        out.field_energy_step,
+        out.total_momentum_step,
+        out.angular_momentum_step,
+        out.torque_step,
+        out.temperature_step,
+        out.thermostat_work_step,
+    ]
+
+    outdataset_time = [
+        out.positions_time,
+        out.total_energy_time,
+        out.potential_energy_time,
+        out.kinetc_energy_time,
[…]
