### 2022-08-16 46e5322
Add check to total charge.
