+++ hymd/input_parser.py
@@ -6,6 +6,7 @@ import datetime
 import logging
 import warnings
 import numpy as np
+import math
 from mpi4py import MPI
 from dataclasses import dataclass, field
 from typing import List, Union, ClassVar
@@ -1048,3 +1049,25 @@ def check_config(config, indices, names, types, comm=MPI.COMM_WORLD):
     config = check_thermostat_coupling_groups(config, comm=comm)
     config = check_cancel_com_momentum(config, comm=comm)
     return config
+
+
+def check_charges(charges, comm=MPI.COMM_WORLD):
+    """Check if charges across ranks sum to zero.
+
+    Parameters
+    ----------
+    charges : (N,) numpy.ndarray
+        Array of floats with charges for :code:`N` particles.
+    comm : mpi4py.Comm, optional
+        MPI communicator, defaults to :code:`mpi4py.COMM_WORLD`.
+    """  
+    total_charge = comm.allreduce(np.sum(charges), MPI.SUM)
+
+    if not math.isclose(total_charge, 0.):
+        err_str = (
+            f"Charges in the input file do not sum to zero. "
+            f"Total charge is {total_charge}."
+        )
+        Logger.rank0.log(logging.ERROR, err_str)
+        raise AssertionError(err_str)
+
