+++ hymd/thermostat.py
@@ -15,10 +15,10 @@ def cancel_com_momentum(velocities, config, comm=MPI.COMM_WORLD):
     return velocities
 
 
-def generate_initial_velocities(velocities, config, comm=MPI.COMM_WORLD):
+def generate_initial_velocities(velocities, config, prng, comm=MPI.COMM_WORLD):
     kT_start = config.gas_constant * config.start_temperature
     n_particles_ = velocities.shape[0]
-    velocities[...] = np.random.normal(
+    velocities[...] = prng.normal(
         loc=0, scale=kT_start / config.mass, size=(n_particles_, 3)
     )
     com_velocity = comm.allreduce(np.sum(velocities[...], axis=0), MPI.SUM)
@@ -46,20 +46,25 @@ def generate_initial_velocities(velocities, config, comm=MPI.COMM_WORLD):
     return velocities
 
 
-def _random_gaussian() -> float:
+def _random_gaussian(prng: np.random.Generator) -> float:
     """Draw a single random number from the standard normal distribution
     Generate a number from the Gaussian distribution centered at zero with unit
     standard deviation, :math:`N(0, 1)`.
+
+    Parameters
+    ----------
+    prng : np.random.Generator
+        Numpy object that provides a stream of random bits
 
     Returns
     -------
     float
         A random number drawn from :math:`N(0, 1)`.
     """
-    return np.random.normal()
+    return prng.normal()
 
 
-def _random_chi_squared(M: int) -> float:
+def _random_chi_squared(prng: np.random.Generator, M: int) -> float:
     """Draw the sum of :code:`M` squared normally distributed values
 
     The value is generated by the Gamma distribution, in lieu of generating
@@ -67,6 +72,8 @@ def _random_chi_squared(M: int) -> float:
 
     Parameters
     ----------
+    prng : np.random.Generator
+        Numpy object that provides a stream of random bits
     M : int
         Number of standard normally distributed numbers in the sum.
 
@@ -100,14 +107,15 @@ def _random_chi_squared(M: int) -> float:
     Computer Programming (Reading, MA: Addison-Wesley), pp. 120ff.
     J. H. Ahrens and U. Dieter, Computing 12 (1974), 223-246.
     """
-    return np.random.chisquare(M)
+    return prng.chisquare(M)
 
 
 def csvr_thermostat(
     velocity: np.ndarray, names: np.ndarray, config: Config,
+    prng: np.random.Generator,
     comm: MPI.Intracomm = MPI.COMM_WORLD,
-    random_gaussian: Callable[[], float] = _random_gaussian,
-    random_chi_squared: Callable[[int, float], float] = _random_chi_squared,
+    random_gaussian: Callable[[np.random.Generator], float] = _random_gaussian,
+    random_chi_squared: Callable[[np.random.Generator, int, float], float] = _random_chi_squared,
     remove_center_of_mass_momentum: bool = True,
 ) -> np.ndarray:
     """Canonical sampling through velocity rescaling thermostat
@@ -190,8 +198,8 @@ def csvr_thermostat(
         # MPI ranks
         R = SNf = None
         if comm.Get_rank() == 0:
-            R = random_gaussian()
-            SNf = random_chi_squared(N_f - 1)
+            R = random_gaussian(prng)
+            SNf = random_chi_squared(prng, N_f - 1)
         R = comm.bcast(R, root=0)
         SNf = comm.bcast(SNf, root=0)
 
