### 2022-08-16 41a6e95
Refactor check for n_flush.
