### 2022-08-16 54c9ba8
Remove duplicate code.
