### 2022-08-17 c31fe5b
Fix configure_runtime tests.
