### 2022-08-18 9f3b386
Fix thermostat test and pytest-mpi to fail properly before, even if the MPI tests failed, the CI wouldn't know.
