### 2022-08-18 c37fa74
Change to numpy implementation of isclose.
