### 2022-08-23 5890a80
Additional tests to configure_runtime.
