### 2022-08-23 126d1c2
Fix file_io tests.
