### 2022-08-24 a877407
Improve configure_runtime test.
