### 2022-08-24 071ef20
Allows to use PlumedBias in a with context and adds more testing.
