### 2022-08-24 ec7d308
Fix plumed test.
