### 2022-08-25 65787ce
Improve plumed tests.
