### 2022-08-25 cccc0c6
Remove check for plumed in configure_runtime this is already checked when a PlumedBias object is created.
