### 2022-09-23 64ef0ff
Initial implementation of self energy correction.
