+++ hymd/main.py
@@ -15,7 +15,8 @@ from .logger import Logger, format_timedelta
 from .file_io import distribute_input, OutDataset, store_static, store_data
 from .field import (compute_field_force, update_field,
                     compute_field_and_kinetic_energy, domain_decomposition,
-                    update_field_force_q, compute_field_energy_q,)
+                    update_field_force_q, compute_field_energy_q,
+                    compute_self_energy_q)
 from .thermostat import (csvr_thermostat, cancel_com_momentum,
                          generate_initial_velocities)
 from .force import dipole_forces_redistribution, prepare_bonds
@@ -258,6 +259,7 @@ def main():
         )
 
     if charges_flag and config.coulombtype == "PIC_Spectral":
+        field_q_self_energy = compute_self_energy_q(config, charges, comm=comm)
         layout_q = pm.decompose(positions)
         update_field_force_q(
             charges, phi_q, phi_q_fourier, elec_field_fourier, elec_field,
@@ -266,7 +268,7 @@ def main():
 
         field_q_energy = compute_field_energy_q(
             config, phi_q_fourier, elec_energy_field, field_q_energy,
-            comm=comm,
+            field_q_self_energy, comm=comm,
         )
 
     if molecules_flag:
@@ -567,7 +569,7 @@ def main():
                 )
                 field_q_energy = compute_field_energy_q(
                     config, phi_q_fourier, elec_energy_field, field_q_energy,
-                    comm=comm,
+                    field_q_self_energy, comm=comm,
                 )
 
             if protein_flag and not args.disable_dipole:
@@ -765,7 +767,7 @@ def main():
                     if charges_flag and config.coulombtype == "PIC_Spectral":
                         field_q_energy = compute_field_energy_q(
                             config, phi_q_fourier, elec_energy_field,
-                            field_q_energy, comm=comm,
+                            field_q_energy, field_q_self_energy, comm=comm,
                         )
                 else:
                     kinetic_energy = comm.allreduce(
@@ -833,7 +835,7 @@ def main():
 
                 field_q_energy = compute_field_energy_q(
                     config, phi_q_fourier, elec_energy_field, field_q_energy,
-                    comm=comm,
+                    field_q_self_energy, comm=comm,
                 )
 
         else:
