### 2022-09-26 964be4c
Fix dataset size for n_print == 1.
