### 2022-09-28 5782e25
Fix electrostatic energy also makes the smearing according to the hamiltonian defined function.
