+++ hymd/field.py
@@ -85,14 +85,16 @@ def compute_self_energy_q(config, charges, comm=MPI.COMM_WORLD):
     field_q_self_energy : float
         Electrostatic self energy.
     """
-    prefac = np.sqrt(1./(2.*np.pi*config.sigma*config.sigma))
+    elec_conversion_factor = config.coulomb_constant / config.dielectric_const
+
+    prefac = elec_conversion_factor * np.sqrt(1./(2.*np.pi*config.sigma*config.sigma))
     _squared_charges = charges * charges
     squared_charges_sum = comm.allreduce(np.sum(_squared_charges))
     return prefac * squared_charges_sum
 
 
 def compute_field_energy_q(
-    config, phi_q_fourier, elec_energy_field, field_q_energy,
+    config, phi_q, phi_q_fourier, elec_potential, elec_potential_fourier, 
     field_q_self_energy, comm=MPI.COMM_WORLD,
 ):
     """Calculate the electrostatic energy from a field configuration
@@ -102,68 +104,76 @@ def compute_field_energy_q(
 
     .. math::
 
-        E = \\frac{1}{2}\\int\\mathrm{d}\\mathbf{r}\\,
+        E = \\frac{1}{2\\varepsilon_0 \\varepsilon_r}\\int\\mathrm{d}\\mathbf{r}\\,
             \\rho(\\mathbf{r}) \\Psi(\\mathbf{r}),
 
     where :math:`\\rho(\\mathbf{r})` denotes the charge density at position
-    :math:`\\mathbf{r}`. Through application of Gauss' law and writing
-    :math:`\\Psi` in terms of the electric field :math:`\\mathbf{E}`, this
-    becomes
-
-    .. math::
-
-        E = \\frac{\\varepsilon}{2}\\int\\mathrm{d}\\mathbf{r}\\,
-            \\mathbf{E}\\cdot \\mathbf{E},
-
-    where :math:`\\varepsilon` is the relative dielectric of the simulation
+    :math:`\\mathbf{r}`, :math:`\\varepsilon_0` is the vacuum permittivity 
+    and :math:`\\varepsilon_r` is the relative dielectric of the simulation
     medium.
 
     Parameters
     ----------
     config : Config
         Configuration object.
+    phi_q : pmesh.pm.RealField
+        Pmesh :code:`RealField` object containing the discretized
+        electrostatic potential values in real space on the
+        computational grid. Local for each MPI rank--the full computayional grid
+        is represented by the collective fields of all MPI ranks.
     phi_q_fourier : pmesh.pm.ComplexField
         Pmesh :code:`ComplexField` object containing the discretized
         electrostatic potential values in reciprocal space on the
-        computational grid. Local for each MPI rank--the full computaional grid
+        computational grid. Local for each MPI rank--the full computayional grid
         is represented by the collective fields of all MPI ranks.
-    elec_energy_field : pmesh.pm.ComplexField
-        Pmesh :code:`ComplexField` object for storing calculated discretized
-        electrostatic energy density values in reciprocal space on the
+    elec_potential : pmesh.pm.RealField
+        Pmesh :code:`RealField` object for storing calculated discretized
+        electrostatic potential values in real space on the
         computational grid. Pre-allocated, but empty; any values in this field
         are discarded. Changed in-place. Local for each MPI rank--the full
         computaional grid is represented by the collective fields of all MPI
         ranks.
-    field_q_energy : float
-        Total elecrostatic energy.
+    elec_potential_fourier : pmesh.pm.ComplexField
+        Pmesh :code:`ComplexField` object for storing calculated discretized
+        electrostatic potential values in reciprocal space on the
[…]
