### 2022-09-29 973f10c
Broadcast root's PRNG to avoid comm in thermostat.
