### 2022-09-29 868d6aa
Fixes to initialize_pm and received pmesh objects.
