### 2022-10-03 077be3d
Fix error in simulations with barostat when electrostatics was not used.
