+++ hymd/pressure.py
@@ -117,7 +117,8 @@ def comp_pressure(
     V_bar = [sum(list(V_bar_tuple[i])) for i in range(len(V_bar_tuple))]
     #print("shape before", np.shape(V_bar))
 
-    V_bar = [sum(n) for n in zip(V_bar,Vbar_elec)]
+    if Vbar_elec:
+        V_bar = [sum(n) for n in zip(V_bar,Vbar_elec)]
     #print("shape after", np.shape(V_bar))
 
     p1 = [
