### 2022-10-04 bcf5e02
Move barostat_type check.
