### 2022-10-21 15fa9cd
Changed to latest version of PLUMED in CI.
