+++ .github/workflows/ci.yml
@@ -53,7 +53,7 @@ jobs:
         export PATH=/usr/lib/ccache:${PATH}
         git clone https://github.com/plumed/plumed2.git $HOME/plumed2
         ccache -s
-        CXX="mpic++" PYTHON_BIN=python3 .ci/install.plumed version=master program_path=$HOME/plumed2
+        CXX="mpic++" PYTHON_BIN=python3 .ci/install.plumed program_path=$HOME/plumed2
         ccache -s
     - name: Install PLUMED python wrapper
       run: |
