+++ hymd/field.py
@@ -182,7 +182,7 @@ def compute_field_force(layouts, r, force_mesh, force, types, n_types):
         Pmesh :code:`RealField` objects containing discretized particle-field
         force density values on the computational grid; :code:`D` fields in D
         dimensions for each particle type. Local for each MPI rank--the full
-        computaional grid is represented by the collective fields of all MPI
+        computational grid is represented by the collective fields of all MPI
         ranks.
     force : (N,D) numpy.ndarray
         Array of forces for :code:`N` particles in :code:`D` dimensions. Local
@@ -235,14 +235,14 @@ def compute_field_energy_q(
     phi_q_fourier : pmesh.pm.ComplexField
         Pmesh :code:`ComplexField` object containing the discretized
         electrostatic potential values in reciprocal space on the
-        computational grid. Local for each MPI rank--the full computaional grid
+        computational grid. Local for each MPI rank--the full computational grid
         is represented by the collective fields of all MPI ranks.
     elec_energy_field : pmesh.pm.ComplexField
         Pmesh :code:`ComplexField` object for storing calculated discretized
         electrostatic energy density values in reciprocal space on the
         computational grid. Pre-allocated, but empty; any values in this field
         are discarded. Changed in-place. Local for each MPI rank--the full
-        computaional grid is represented by the collective fields of all MPI
+        computational grid is represented by the collective fields of all MPI
         ranks.
     field_q_energy : float
         Total elecrostatic energy.
@@ -331,26 +331,26 @@ def update_field_force_q(
         Pmesh :code:`RealField` object for storing calculated discretized
         charge density density values on the computational grid. Pre-allocated,
         but empty; any values in this field are discarded. Changed in-place.
-        Local for each MPI rank--the full computaional grid is represented by
+        Local for each MPI rank--the full computational grid is represented by
         the collective fields of all MPI ranks.
     phi_q_fourier : pmesh.pm.ComplexField
         Pmesh :code:`ComplexField` object for storing calculated discretized
         Fourier transformed charge density values in reciprocal space on the
         computational grid. Pre-allocated, but empty; any values in this field
         are discarded. Changed in-place. Local for each MPI rank--the full
-        computaional grid is represented by the collective fields of all MPI
+        computational grid is represented by the collective fields of all MPI
         ranks.
     elec_field_fourier : pmesh.pm.ComplexField
         Pmesh :code:`ComplexField` object for storing calculated discretized
         electric field values in reciprocal space on the computational grid.
         Pre-allocated, but empty; any values in this field are discarded.
-        Changed in-place. Local for each MPI rank--the full computaional grid
+        Changed in-place. Local for each MPI rank--the full computational grid
         is represented by the collective fields of all MPI ranks.
     elec_field : pmesh.pm.RealField
         Pmesh :code:`RealField` object for storing calculated discretized
         electric field values on the computational grid. Pre-allocated,
         but empty; any values in this field are discarded. Changed in-place.
-        Local for each MPI rank--the full computaional grid is represented by
+        Local for each MPI rank--the full computational grid is represented by
         the collective fields of all MPI ranks.
     elec_forces : (N,D) numpy.ndarray
         Array of electrostatic forces on :code:`N` particles in :code:`D`
@@ -401,8 +401,6 @@ def update_field_force_q(
             elec_field[_d].readout(positions, layout=layout_q)
         )
 
-    #print("PE F max", np.max(elec_forces))
-    #print("elec_field PE", elec_field[2][0,0,0])
 
 def update_field_force_energy_q(
[…]
