+++ hymd/main.py
@@ -110,17 +110,16 @@ def main():
     if config.coulombtype == 'PIC_Spectral_GPE':
         config_charges = get_charges_types_list(config, types, charges, comm = comm)
         dielectric_sorted = sort_dielectric_by_type_id(config,charges,types)
-        ###^----- dielectric not needed in new version of force-GPE
         dielectric_flag = True
 
         if config.convergence_type is not None:
             if config.convergence_type == 'csum':
                 def conv_fun(comm,diff_mesh):
-                    return diff_mesh.csum()
+                    return diff_mesh.csum() # check if allreduce needed
 
             elif config.convergence_type == 'euclidean_norm':
                 def conv_fun(comm,diff_mesh):
-                    return diff_mesh.cnorm()
+                    return diff_mesh.cnorm() # check if allreduce nedded
 
             elif config.convergence_type == 'max_diff':
                 def conv_fun(comm,diffmesh):
@@ -341,26 +340,6 @@ def main():
                 comm=comm
                 )
 
-            #rank = comm.Get_rank()
-            #print(charges)
-            #print("main",elec_forces[:,2])
-            #sum_elec = np.sum(elec_forces,axis = 0)
-            #if rank == 0:
-            #    sums = np.zeros_like(sum_elec)
-            #else:
-            #    sums = None
-            #comm.Barrier()
-            #comm.Reduce(sum_elec, sums,
-            #op=MPI.SUM, root=0)
-            #print("rank ", comm.Get_rank())
-            #print("sum elec_forces", np.sum(elec_forces,axis = 0))
-            #if rank == 0:
-            #    #    f = open("./sum_forces.txt", "w")
-            #    #    f.write("time step, sum forces x,y,z \n")
-            #    #    f.write("{:.2f}, {:.2f}, {:.2f}, {:.2f} \n".format(0, sums[0], sums[1], sums[2]))
-            #    print("here ",sums)
-            #    #    #print(sum_elec)
-            #    #    print("sum elec_forces", np.sum(elec_forces,axis = 0))
 
         if config.coulombtype == "PIC_Spectral":
             update_field_force_q(
@@ -372,24 +351,6 @@ def main():
                 config, phi_q_fourier, elec_energy_field, field_q_energy,
                 comm=comm,
             )
-
-            #sum_elec = np.sum(elec_forces,axis = 0)
-            #rank = comm.Get_rank()
-            #if rank == 0:
-            #    sums = np.zeros_like(sum_elec)
-            #else:
-            #    sums = None
-            #comm.Barrier()
-            #comm.Reduce(sum_elec, sums,
-            #op=MPI.SUM, root=0)
-            #print("rank ", comm.Get_rank())
-            #print("sum elec_forces", np.sum(elec_forces,axis = 0))
-            #print("here", sums)
-
-            #if rank == 0:
-            #    print("sum elec forces", sums)
-            #print(config.type_to_name_map)
-            #print(charges[types == 4])
 
     if molecules_flag:
         if not (args.disable_bonds
@@ -647,7 +608,6 @@ def main():
 
         if charges_flag and (config.coulombtype == "PIC_Spectral" \
                     or config.coulombtype == "PIC_Spectral_GPE"):
-            #print("here integrate vel 2")
             velocities = integrate_velocity(
                 velocities, elec_forces / config.mass,
                 config.respa_inner * config.time_step,
@@ -800,7 +760,6 @@ def main():
                 #print(charges)
                 #print(config.name_to_type_map)
                 if config.coulombtype == "PIC_Spectral_GPE": # dielectric_flag and
-                    #print("update elec forces  w rank", comm.Get_rank())
                     Vbar_elec, phi_eps, elec_dot = update_field_force_q_GPE(
[…]
