### 2022-11-03 0b770d7
Fixed function declarations for GPE.
