### 2022-11-04 926672a
Adds Vbar_elec for constant dielectric electrostatics.
