+++ hymd/compute_bond_forces.f90
@@ -27,7 +27,7 @@ subroutine cbf(f, r, box, i, j, r0, k, energy, bond_pr)
     !
     implicit none
 
-    real(4), dimension(:,:), intent(in out) :: f
+    real(4), dimension(:,:), intent(inout)  :: f
     real(4), dimension(:,:), intent(in)     :: r
     real(8), dimension(:),   intent(in)     :: box
     integer, dimension(:),   intent(in)     :: i
