### 2022-11-16 a37a7bb
Creates hamiltonian terms for elecrostatics and also implements the V_bar_elec properly for PIC_Spectral.
