+++ hymd/field.py
@@ -71,8 +71,8 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
 
     # Initialize charge density fields
     coulomb_list = []
-    elec_common_list = []
-    _SPACE_DIM = 3
+    elec_common_list = [None, None, None, None]
+    _SPACE_DIM = config.box_size.size
 
     if (config.coulombtype == 'PIC_Spectral_GPE'
         or config.coulombtype == "PIC_Spectral"):
@@ -82,32 +82,26 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         phi_q_fourier = pm.create(
             "complex", value=0.0
         )
+        psi = pm.create(
+            "real", value=0.0
+        )
         elec_field = [
             pm.create("real", value=0.0) for _ in range(_SPACE_DIM)
-        ] # for force calculation
+        ]
 
-        elec_common_list = [phi_q, phi_q_fourier, elec_field]
+        elec_common_list = [phi_q, phi_q_fourier, psi, elec_field]
 
     if config.coulombtype == "PIC_Spectral":
         elec_field_fourier = [
             pm.create("complex", value=0.0) for _ in range(_SPACE_DIM)
         ]  # for force calculation
-        elec_potential = pm.create(
-            "real", value=0.0
-        )
-        elec_potential_fourier = pm.create(
+        psi_fourier = pm.create(
             "complex", value=0.0
         )
-        # Can run elecPE with pressure by adding this
-        Vbar_elec = [
-            pm.create("real", value=0.0) for _ in range(config.n_types)
-        ]
 
         coulomb_list = [
             elec_field_fourier,
-            elec_potential,
-            elec_potential_fourier,
-            Vbar_elec
+            psi_fourier,
         ]
 
     if config.coulombtype == 'PIC_Spectral_GPE': ## initializing the density mesh #dielectric_flag
@@ -118,7 +112,6 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         phi_pol = pm.create("real", value = 0.0) ## real contrib of the polarization charge
         phi_pol_prev = pm.create("real", value = 0.0)
         elec_dot = pm.create("real", value = 0.0)
-        elec_potential = pm.create("real", value = 0.0)
         elec_field_contrib = pm.create("real", value = 0.0) # needed for pol energies later
 
         # External potential and force meshes
@@ -140,19 +133,15 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         coulomb_list = [
             phi_eps, phi_eps_fourier,
             phi_eta, phi_eta_fourier, phi_pol,
-            phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec, Vbar_elec_fourier,
+            phi_pol_prev, elec_dot, elec_field_contrib, Vbar_elec, Vbar_elec_fourier,
             force_mesh_elec, force_mesh_elec_fourier
         ]
-        #coulomb_list = [phi_q, phi_q_fourier, elec_field,
-        #        phi_eps, phi_eps_fourier,
-        #        phi_eta, phi_eta_fourier, phi_pol,
-        #        phi_pol_prev, elec_dot, elec_field_contrib, elec_potential, Vbar_elec
-        #        ]
 
-    field_list = [phi, phi_fourier, force_on_grid, v_ext_fourier, v_ext, phi_transfer,
-            phi_gradient, phi_laplacian, phi_lap_filtered_fourier, phi_lap_filtered,
-            phi_grad_lap_fourier, phi_grad_lap, v_ext1
-            ]
+    field_list = [
+        phi, phi_fourier, force_on_grid, v_ext_fourier, v_ext, phi_transfer,
[…]
