### 2022-11-17 0fbf28d
Changes to some utilities also adds a few more tools in utils.
