### 2022-12-12 25a38c7
Init index based topology (again)
