+++ hymd/force.py
@@ -456,7 +456,77 @@ def prepare_bonds_old(molecules, names, bonds, indices, config):
     return bonds_2, bonds_3, bonds_4, bb_index
 
 
-def prepare_bonds(molecules, names, bonds, indices, config):
+def prepare_index_based_bonds(molecules, topol):
+    bonds_2 = []
+    bonds_3 = []
+    bonds_4 = []
+    impr_dih = []
+
+    different_molecules = np.unique(molecules)
+    for mol in different_molecules:
+        resid = mol + 1
+        top_summary = topol["system"]["molecules"]
+        resname = None
+        test_mol_number = 0
+        for molname in top_summary:
+            test_mol_number += molname[1]
+            if resid <= test_mol_number:
+                resname = molname[0]
+                break
+
+        if resname is None:
+            break
+
+        if "bonds" in topol[resname]:
+            first_id = np.where(molecules == mol)[0][0]
+            for bond in topol[resname]["bonds"]:
+                index_i = bond[0] - 1 + first_id
+                index_j = bond[1] - 1 + first_id
+                # bond[2] is the bond type, inherited by the itp format,
+                # we don't use it
+                equilibrium = bond[3]
+                strength = bond[4]
+                bonds_2.append([index_i, index_j, equilibrium, strength])
+
+        if "angles" in topol[resname]:
+            first_id = np.where(molecules == mol)[0][0]
+            for angle in topol[resname]["angles"]:
+                index_i = angle[0] - 1 + first_id
+                index_j = angle[1] - 1 + first_id
+                index_k = angle[2] - 1 + first_id
+                # angle[3] is the angle type, inherited by the itp format
+                # we don't use it
+                equilibrium = np.radians(angle[4])
+                strength = angle[5]
+                bonds_3.append([index_i, index_j, index_k, equilibrium, strength])
+
+        if "dihedrals" in topol[resname]:
+            first_id = np.where(molecules == mol)[0][0]
+            for angle in topol[resname]["dihedrals"]:
+                index_i = angle[0] - 1 + first_id
+                index_j = angle[1] - 1 + first_id
+                index_k = angle[2] - 1 + first_id
+                index_l = angle[3] - 1 + first_id
+                dih_type = angle[4]
+                coeff = angle[5]
+                bonds_4.append([index_i, index_j, index_k, index_l, coeff, dih_type, 0])
+
+        if "impropers" in topol[resname]:
+            first_id = np.where(molecules == mol)[0][0]
+            for angle in topol[resname]["impropers"]:
+                index_i = angle[0] - 1 + first_id
+                index_j = angle[1] - 1 + first_id
+                index_k = angle[2] - 1 + first_id
+                index_l = angle[3] - 1 + first_id
+                equilibrium = np.radians(angle[4])
+                strength = angle[5]
+                impr_dih.append(
+                    [index_i, index_j, index_k, index_l, equilibrium, strength]
+                )
+    return bonds_2, bonds_3, bonds_4, impr_dih
+
+
+def prepare_bonds(molecules, names, bonds, indices, config, topol=None):
     """Rearrange the bond information for usage in compiled Fortran kernels
 
     Restructures the lists resulting from the execution of
@@ -538,9 +608,14 @@ def prepare_bonds(molecules, names, bonds, indices, config):
         connectivity information in the structure/topology input file and the
         bonded types specified in the configuration file.
     """
[…]
