### 2022-12-15 4a13871
Remove squaregradient and change comp_pressure to perform operations with numpy. Also, files were formatted with Black.
