+++ hymd/input_parser.py
@@ -9,7 +9,7 @@ import numpy as np
 from mpi4py import MPI
 from dataclasses import dataclass, field
 from typing import List, Union, ClassVar
-from .force import Bond, Angle, Dihedral, Chi, Dielectric_type, K_Coupl
+from .force import Bond, Angle, Dihedral, Chi, Dielectric_type
 from .barostat import Target_pressure
 from .logger import Logger
 
@@ -155,6 +155,7 @@ class Config:
     hymd.input_parser.Dihedral :
         Four-particle bond type dataclass.
     """
+
     gas_constant: ClassVar[float] = 0.0083144621  # kJ mol-1 K-1
     coulomb_constant: ClassVar[float] = 138.935458  # kJ nm mol-1 e-2
 
@@ -181,7 +182,6 @@ class Config:
     angle_bonds: List[Angle] = field(default_factory=list)
     dihedrals: List[Dihedral] = field(default_factory=list)
     chi: List[Chi] = field(default_factory=list)
-    K_coupl: List[K_Coupl] = field(default_factory=list)
     n_particles: int = None
     max_molecule_size: int = None
     n_flush: int = None
@@ -198,30 +198,27 @@ class Config:
     self_energy: float = None
     type_charges: Union[List[float], np.ndarray] = None
 
-    squaregradient: bool = False
-
-    #For NPT runs
+    # For NPT runs
     rho0: float = None
     a: float = None
     pressure: bool = False
     barostat: str = None
     barostat_type: str = None
     tau_p: float = None
-    target_pressure : List[Target_pressure] = field(default_factory=list)
+    target_pressure: List[Target_pressure] = field(default_factory=list)
     n_b: int = None
     m: List[float] = field(default_factory=list)
 
     def __str__(self):
         target_pressure_str = "\ttarget_pressure:\n" + "".join(
-                "\t\tP_L: " + f"{self.target_pressure.P_L}\n" +
-                "\t\tP_N: " + f"{self.target_pressure.P_N}\n"
+            "\t\tP_L: "
+            + f"{self.target_pressure.P_L}\n"
+            + "\t\tP_N: "
+            + f"{self.target_pressure.P_N}\n"
         )
         bonds_str = "\tbonds:\n" + "".join(
             [
-                (
-                    f"\t\t{k.atom_1} {k.atom_2}: "
-                    f"{k.equilibrium}, {k.strength}\n"
-                )
+                (f"\t\t{k.atom_1} {k.atom_2}: " f"{k.equilibrium}, {k.strength}\n")
                 for k in self.bonds
             ]
         )
@@ -242,9 +239,7 @@ class Config:
                     # there's an easier way
                     + (
                         "\n\t\t"
-                        + " " * len(
-                            f"{k.atom_1} {k.atom_2} {k.atom_3} {k.atom_4}: "
-                        )
+                        + " " * len(f"{k.atom_1} {k.atom_2} {k.atom_3} {k.atom_4}: ")
                     ).join(
                         map(
                             str,
@@ -258,9 +253,7 @@ class Config:
                     )
                     + (
                         "\n\t\t"
-                        + " " * len(
-                            f"{k.atom_1} {k.atom_2} {k.atom_3} {k.atom_4}: "
-                        )
+                        + " " * len(f"{k.atom_1} {k.atom_2} {k.atom_3} {k.atom_4}: ")
                     )
                     + f"dih_type = {k.dih_type}\n"
                 )
@@ -284,13 +277,6 @@ class Config:
 
         """
 
-        K_coupl_str = "\tK_coupl:\n" + "".join(
-            [
-                    (f"\t\t{k.atom_1} {k.atom_2}: " + f"{k.squaregradient_energy}\n")
[…]
