### 2023-01-09 d4cd5aa
Changes how the phi_q are updated. For some reason, when used as an argument, apparently it wasn't updated.
