+++ hymd/field.py
@@ -7,7 +7,7 @@ from .logger import Logger
 import warnings
 
 
-def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
+def initialize_pm(pmesh, config, dtype, comm=MPI.COMM_WORLD):
     """
     Creates the necessary pmesh objects for pfft operations.
 
@@ -29,6 +29,10 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         Additional list of pmesh objects required for electrostatics.
     """
 
+    if dtype == np.float64:
+        pmeshtype = "f8"
+    else:
+        pmeshtype = "f4"
     # Ignore numpy numpy.VisibleDeprecationWarning: Creating an ndarray from
     # ragged nested sequences until it is fixed in pmesh
     with warnings.catch_warnings():
@@ -39,7 +43,7 @@ def initialize_pm(pmesh, config, comm=MPI.COMM_WORLD):
         )
         # The first argument of ParticleMesh has to be a tuple
         pm = pmesh.ParticleMesh(
-            config.mesh_size, BoxSize=config.box_size, dtype="f4", comm=comm
+            config.mesh_size, BoxSize=config.box_size, dtype=pmeshtype, comm=comm
         )
     phi = [pm.create("real", value=0.0) for _ in range(config.n_types)]
     phi_fourier = [
@@ -358,18 +362,25 @@ def update_field_force_q(
     # charges to grid
     pm.paint(positions, layout=layout_q, mass=charges, out=phi_q)
     phi_q /= volume_per_cell
+    # print("phi_q", np.sum(phi_q))
     phi_q.r2c(out=phi_q_fourier)
 
     # smear charges with filter
-    phi_q_fourier.apply(hamiltonian.H, out=phi_q_fourier)
+    phi_q_fourier.apply(hamiltonian.H, out=Ellipsis)
     phi_q_fourier.c2r(out=phi_q)
+    # print("volume_per_cell", volume_per_cell)
+    # print("phi_q_fourier", np.sum(phi_q_fourier))
 
     # solve Poisson equation in Fourier space to get electrostatic potential
     def poisson_transfer_function(k, v):
         return 4.0 * np.pi * elec_conversion_factor * v / k.normp(p=2, zeromode=1)
 
     phi_q_fourier.apply(poisson_transfer_function, out=psi_fourier)
+    # print("psi_fourier", np.sum(psi_fourier))
     psi_fourier.c2r(out=psi)
+    # print("phi_q * psi update_field", np.sum(phi_q * psi))
+
+    # exit()
 
     # compute electric field directly from smeared charged densities in Fourier
     for _d in np.arange(n_dimensions):
@@ -394,147 +405,7 @@ def update_field_force_q(
             elec_field[_d].readout(positions, layout=layout_q)
         )
 
-
-def update_field_force_energy_q(
-    charges,
-    phi_q,
-    phi_q_fourier,
-    elec_field_fourier,
-    elec_field,
-    elec_forces,
-    elec_energy_field,
-    field_q_energy,
-    layout_q,
-    pm,
-    positions,
-    config,
-    compute_energy=False,
-    comm=MPI.COMM_WORLD,
-):
-    """Calculate the electrostatic particle-field energy and forces on the grid
-
-    .. deprecated:: 1.0.0
-        :code:`update_field_force_energy_q` was deprecated in favour of
-        :code:`update_field_force_q` and independent calls to
-        :code:`compute_field_energy_q` prior to 1.0.0 release.
-
-    Parameters
-    ----------
-    charges : (N,) numpy.ndarray
-        Array of particle charge values for :code:`N` particles. Local for each
-        MPI rank.
-    phi_q : pmesh.pm.RealField
[…]
