### 2023-01-18 3564f60
Return unfiltered charge density
