+++ hymd/field.py
@@ -350,7 +350,6 @@ def update_field_force_q(
 
     # smear charges with filter
     phi_q_fourier.apply(hamiltonian.H, out=phi_q_fourier)
-    phi_q_fourier.c2r(out=phi_q)
 
     # solve Poisson equation in Fourier space to get electrostatic potential
     def poisson_transfer_function(k, v):
