### 2023-01-26 e566c93
Remove non-UTF-8 characters in Fortran modules
