### 2023-01-31 9958d7a
Remove comments with greek letters
