+++ hymd/dipole_reconstruction.f90
@@ -1,221 +1,213 @@
 module dipole_reconstruction
-implicit none
+  implicit none
 
 contains
-function cross(vector1, vector2) result(vector3)
-  real(8), dimension(3), intent(in) :: vector1, vector2
-  real(8), dimension(3)             :: vector3
+  function cross(vector1, vector2) result(vector3)
+    real(8), dimension(3), intent(in) :: vector1, vector2
+    real(8), dimension(3)             :: vector3
 
-  vector3(1) = vector1(2) * vector2(3) - vector1(3) * vector2(2)
-  vector3(2) = vector1(3) * vector2(1) - vector1(1) * vector2(3)
-  vector3(3) = vector1(1) * vector2(2) - vector1(2) * vector2(1)
-end function
+    vector3(1) = vector1(2) * vector2(3) - vector1(3) * vector2(2)
+    vector3(2) = vector1(3) * vector2(1) - vector1(1) * vector2(3)
+    vector3(3) = vector1(1) * vector2(2) - vector1(2) * vector2(1)
+  end function
 
-function cross_matrix(matrix, vector) result(output)
-  ! The i-th row of the output matrix is the cross product
-  ! between the i-th row of the input matrix and the input vector.
-  real(8), dimension(3,3), intent(in) :: matrix
-  real(8), dimension(3),   intent(in) :: vector
-  real(8), dimension(3,3)             :: output
+  function cross_matrix(matrix, vector) result(output)
+    ! The i-th row of the output matrix is the cross product
+    ! between the i-th row of the input matrix and the input vector.
+    real(8), dimension(3, 3), intent(in) :: matrix
+    real(8), dimension(3), intent(in) :: vector
+    real(8), dimension(3, 3)             :: output
 
-  output(1, :) = cross(matrix(1, :), vector)
-  output(2, :) = cross(matrix(2, :), vector)
-  output(3, :) = cross(matrix(3, :), vector)
-end function
+    output(1, :) = cross(matrix(1, :), vector)
+    output(2, :) = cross(matrix(2, :), vector)
+    output(3, :) = cross(matrix(3, :), vector)
+  end function
 
-function outer_product(vector1, vector2) result(output)
-  ! The i-th row of the output matrix is vector2
-  ! multiplied by the i-th component of vector1.
-  real(8), dimension(3), intent(in) :: vector1, vector2
-  real(8), dimension(3,3)           :: output
+  function outer_product(vector1, vector2) result(output)
+    ! The i-th row of the output matrix is vector2
+    ! multiplied by the i-th component of vector1.
+    real(8), dimension(3), intent(in) :: vector1, vector2
+    real(8), dimension(3, 3)           :: output
 
-  output(1, :) = vector1(1) * vector2
-  output(2, :) = vector1(2) * vector2
-  output(3, :) = vector1(3) * vector2
-end function outer_product
+    output(1, :) = vector1(1) * vector2
+    output(2, :) = vector1(2) * vector2
+    output(3, :) = vector1(3) * vector2
+  end function outer_product
 
[…]
