### 2023-03-02 967442b
Read box from h5 or config.toml. Don't parse optimization options from diff_hymd tomls.
