+++ hymd/input_parser.py
@@ -153,11 +153,11 @@ class Config:
 
     n_steps: int
     time_step: float
-    box_size: Union[List[float], np.ndarray]
     mesh_size: Union[Union[List[int], np.ndarray], int]
     sigma: float
     kappa: float
 
+    box_size: Union[List[float], np.ndarray] = None
     n_print: int = None
     tau: float = None
     start_temperature: Union[float, bool] = None
@@ -275,8 +275,8 @@ class Config:
 
 
 def read_config_toml(file_path):
-    with open(file_path, "r") as in_file:
-        toml_content = in_file.read()
+    with open(file_path, "rb") as in_file:
+        toml_content = tomli.load(in_file)
     return toml_content
 
 
@@ -329,11 +329,11 @@ def propensity_potential_coeffs(x: float, comm):
 
 
 def parse_config_toml(toml_content, file_path=None, comm=MPI.COMM_WORLD):
-    parsed_toml = tomli.loads(toml_content)
     config_dict = {}
 
     # Defaults = None
     for n in (
+        "box_size",
         "n_print",
         "tau",
         "start_temperature",
@@ -357,8 +357,10 @@ def parse_config_toml(toml_content, file_path=None, comm=MPI.COMM_WORLD):
 
     # Flatten the .toml dictionary, ignoring the top level [tag] directives (if
     # any).
-    for k, v in parsed_toml.items():
+    for k, v in toml_content.items():
         if isinstance(v, dict):
+            if k == "nn": # Don't parse diff-hymd optimization options
+                continue
             for nested_k, nested_v in v.items():
                 config_dict[nested_k] = nested_v
         else:
@@ -437,9 +439,6 @@ def parse_config_toml(toml_content, file_path=None, comm=MPI.COMM_WORLD):
                     coeffs=coeff,
                     dih_type=dih_type,
                 )
-        # if k == "improper dihedrals":
-        #     config_dict["improper dihedrals"] = [None] * len(v)
-        # ...
         if k == "chi":
             config_dict["chi"] = [None] * len(v)
             for i, c in enumerate(v):
@@ -452,7 +451,7 @@ def parse_config_toml(toml_content, file_path=None, comm=MPI.COMM_WORLD):
         config_dict["file_name"] = file_path
 
     for n in (
-        "n_steps", "time_step", "box_size", "mesh_size", "sigma", "kappa"
+        "n_steps", "time_step", "mesh_size", "sigma", "kappa"
     ):
         if n not in config_dict:
             err_str = (
