+++ hymd/main.py
@@ -162,7 +162,6 @@ def main():
     else:
         dielectric_flag = False
 
-
     if config.start_temperature:
         velocities = generate_initial_velocities(velocities, config, prng, comm=comm)
     elif config.cancel_com_momentum:
@@ -414,13 +413,10 @@ def main():
             )
 
     if molecules_flag:
-        if not (args.disable_bonds
-                and args.disable_angle_bonds
-                and args.disable_dihedrals):
-
-            bonds_prep = prepare_bonds(
-                molecules, names, bonds, indices, config, topol 
-            )
+        if not (
+            args.disable_bonds and args.disable_angle_bonds and args.disable_dihedrals
+        ):
+            bonds_prep = prepare_bonds(molecules, names, bonds, indices, config, topol)
             (
                 # two-particle bonds
                 bonds_2_atom1,
@@ -479,6 +475,7 @@ def main():
                 phi_dipoles_fourier = pm.create("complex", value=0.0)
                 psi_dipoles = pm.create("real", value=0.0)
                 psi_dipoles_fourier = pm.create("complex", value=0.0)
+                _SPACE_DIM = 3
                 dipoles_field_fourier = [
                     pm.create("complex", value=0.0) for _ in range(_SPACE_DIM)
                 ]
@@ -707,7 +704,6 @@ def main():
     for step in range(1, config.n_steps + 1):
         current_step_time = datetime.datetime.now()
 
-        print(args.verbose)
         if step == 1 and args.verbose > 1:
             Logger.rank0.log(logging.INFO, f"MD step = {step:10d}")
         else:
