+++ hymd/compute_bond_forces.f90
@@ -1,4 +1,4 @@
-subroutine cbf(f, r, box, i, j, r0, k, energy, bond_pr)
+subroutine cbf(f, r, box, a, b, r0, k, energy, bond_pr)
     ! Compute two-particle bond forces and energy
     !
     ! Parameters
@@ -9,7 +9,7 @@ subroutine cbf(f, r, box, i, j, r0, k, energy, bond_pr)
     !     Positions for N particles in D dimensions.
     ! box : (D,) numpy.ndarray
     !     D-dimensional simulation box size.
-    ! a : (M,) numpys.ndarray
+    ! a : (M,) numpy.ndarray
     !     Index of particle 1 for M individual two-particle bonds.
     ! b : (M,) numpy.ndarray
     !     Index of particle 2 for M individual two-particle bonds.
@@ -22,65 +22,41 @@ subroutine cbf(f, r, box, i, j, r0, k, energy, bond_pr)
     ! -------
     ! energy : float
     !     Total energy of all two-particle bonds.
-    ! bond_pr : (3,) numpy.ndarray
-    !     Total bond pressure due all two-particle bonds.
     !
     implicit none
 
-    real(4), dimension(:,:), intent(inout)  :: f
+    real(4), dimension(:,:), intent(in out) :: f
     real(4), dimension(:,:), intent(in)     :: r
     real(8), dimension(:),   intent(in)     :: box
-    integer, dimension(:),   intent(in)     :: i
-    integer, dimension(:),   intent(in)     :: j
+    integer, dimension(:),   intent(in)     :: a
+    integer, dimension(:),   intent(in)     :: b
     real(8), dimension(:),   intent(in)     :: r0
     real(8), dimension(:),   intent(in)     :: k
     real(8),                 intent(out)    :: energy
     real(4), dimension(3),   intent(out)    :: bond_pr
 
-    integer :: ind, ii, jj
-    real(8) :: rij, rij_x, rij_y, rij_z
-    real(8) :: df
-    real(8) :: bx, by, bz
+    integer :: ind, aa, bb
+    real(8), dimension(3) :: rab, fa
+    real(8) :: df, rab_norm
 
     energy = 0.0d00
-    bond_pr = 0.0d00 !Set x, y, z components to 0
-    f = 0.0d00 ! Set all array elements
+    f = 0.0d00
 
-    bx = 1.0d00 / box(1)
-    by = 1.0d00 / box(2)
-    bz = 1.0d00 / box(3)
+    do ind = 1, size(a)
+      aa = a(ind) + 1
+      bb = b(ind) + 1
 
-    do ind = 1, size(i)
-      ii = i(ind) + 1
-      jj = j(ind) + 1
+      rab = r(bb, :) - r(aa, :)
+      rab = rab - box * nint(rab / box)
+      rab_norm = norm2(rab)
 
-      rij_x = r(jj, 1) - r(ii, 1)
-      rij_x = rij_x - box(1) * nint(rij_x * bx)
+      df = k(ind) * (rab_norm - r0(ind))
+      fa = -df * rab / rab_norm
 
-      rij_y = r(jj, 2) - r(ii, 2)
-      rij_y = rij_y - box(2) * nint(rij_y * by)
+      f(aa, :) = f(aa, :) - fa
+      f(bb, :) = f(bb, :) + fa
 
-      rij_z = r(jj, 3) - r(ii, 3)
-      rij_z = rij_z - box(3) * nint(rij_z * bz)
-
[…]
