+++ hymd/plumed.py
@@ -16,7 +16,7 @@ class PlumedBias:
     This wraps the :code:`Plumed()` class, see
     `https://github.com/plumed/plumed2/tree/master/python`.
 
-    The :code:`PlumedBias()` object is created with the arguments from 
+    The :code:`PlumedBias()` object is created with the arguments from
     :code:`__init__` and at everystep the methods :code:`prepare()` and
     :code:`calc()` should be called.
 
@@ -31,7 +31,7 @@ class PlumedBias:
     positions : (N, D) numpy.ndarray
         Array of positions for :code:`N` particles in :code:`D` dimensions.
         Local for each MPI rank.
-        Needed because we need C-contiguous array for passing to PLUMED.       
+        Needed because we need C-contiguous array for passing to PLUMED.
     plumed_bias : (1,) numpy.ndarray
         Used as a pointer to an :code:`double` to store the bias energy.
     plumed_version : (1,) numpy.ndarray
@@ -41,6 +41,7 @@ class PlumedBias:
     ready : bool
         Stores wether the :code:`calc()` method can be called or not.
     """
+
     plumed_obj = None
     plumed_forces = None
     positions = None
@@ -51,8 +52,7 @@ class PlumedBias:
     ready = False
     verbose = None
 
-    def __init__(self, config, plumeddat, logfile, comm=MPI.COMM_WORLD,
-                 verbose=1):
+    def __init__(self, config, plumeddat, logfile, comm=MPI.COMM_WORLD, verbose=1):
         """Constructor
 
         Parameters
@@ -76,8 +76,7 @@ class PlumedBias:
             import plumed
         except ImportError:
             err_str = (
-                "You are trying to use PLUMED "
-                "but HyMD could not import py-plumed."
+                "You are trying to use PLUMED " "but HyMD could not import py-plumed."
             )
             Logger.rank0.log(logging.ERROR, err_str)
             raise ImportError(err_str)
@@ -86,16 +85,11 @@ class PlumedBias:
         self.verbose = verbose
 
         try:
-            kernel_str = (
-                "Using PLUMED_KERNEL={}".format(os.environ["PLUMED_KERNEL"])
-            )
+            kernel_str = "Using PLUMED_KERNEL={}".format(os.environ["PLUMED_KERNEL"])
         except:
             kernel_str = "The PLUMED_KERNEL environment variable is not set."
 
-        Logger.rank0.log(
-            logging.INFO, 
-            kernel_str
-        )
+        Logger.rank0.log(logging.INFO, kernel_str)
 
         try:
             self.plumed_obj = plumed.Plumed()
@@ -109,10 +103,7 @@ class PlumedBias:
 
         self.plumed_obj.cmd("getApiVersion", self.plumed_version)
         if self.plumed_version[0] <= 3:
-            err_str = (
-                "HyMD requires a PLUMED API > 3. "
-                "Use a newer PLUMED kernel."
-            )
+            err_str = "HyMD requires a PLUMED API > 3. " "Use a newer PLUMED kernel."
             Logger.rank0.log(logging.ERROR, err_str)
             raise AssertionError(err_str)
 
@@ -120,72 +111,56 @@ class PlumedBias:
         self.plumed_obj.cmd("setMPIComm", comm)
 
         Logger.rank0.log(
-            logging.INFO, 
-            f"Attempting to read PLUMED input from {plumeddat}"
+            logging.INFO, f"Attempting to read PLUMED input from {plumeddat}"
         )
 
         self.plumed_obj.cmd("setPlumedDat", plumeddat)
         self.plumed_obj.cmd("setLogFile", logfile)
-        self.plumed_obj.cmd("setTimestep", 
[…]
