+++ examples/testconfig.toml
@@ -1,108 +1,0 @@
-[meta]
-# Name of the simulation. May be ommitted.
-name = "DPPC bilayer with ML interaction parameters"
-# Tags classifying the simulation. May be ommitted.
-tags = ["bilayer", "solvent", "DPPC"]
-
-[particles]
-# Number of total particles in the simulation. If an input .hdf5 format file is
-# specified, the number of particles will be inferred from this and *may* be
-# ommited.
-n_particles = 10000
-# Mass of the particles in [g/mol]. All masses are assumed equal.
-mass = 72.0
-# Maximum number of particles per molecules present in the system. A default of
-# 200 is assumed, and this keyword may be ommitted for any system with smaller
-# molecules.
-max_molecule_size = 15
-
-[simulation]
-# Number of total time steps in the simulation.
-n_steps = 5
-# Frequency of trajectory/energy file output in time steps.
-n_print = 1
-# Frequency of requesting that the HDF5 library flush the file output buffers
-# to disk after in number of n_print timesteps.
-n_flush = 1
-# Time step used in the simulation in [picoseconds].
-time_step = 0.03
-# Simulation box size in [nanometers].
-box_size = [10.0, 10.0, 10.0]
-# Time integrator used in the simulation. Either "velocity-verlet" or "respa".
-# If "respa", specify also the number of small rRESPA time steps per large
-# time_step with the 'respa_inner' keyword.
-integrator = "respa"
-respa_inner = 5
-# Perform MPI rank domain decomposition every x time steps to (hopefully)
-# reduce the amount of neccessary communication between ranks in the pmesh
-# procedures. Ommit or set to 'false' or '0' to not perform any domain
-# decomposition.
-domain_decomposition = false
-# Remove linear center of mass momentum from the system before integration
-# starts.
-cancel_com_momentum = true
-# Starting temperature to generate before simulation begins in [kelvin]. Ommit
-# or set to 'false' to not change the temperature before starting.
-start_temperature = false
-# Target temperature used in the velocity rescale thermostat in [kelvin]. Ommit
-# or set to 'false' to use no thermostat, i.e. a constant energy simulation.
-target_temperature = 323
-# Thermostat collision frequency in [1/picoseconds].
-tau = 0.7
-# The energy functional W[phi] to use. Options:
-#    "SquaredPhi":      φ² / 2κφ₀,
-#    "DefaultNoChi":   (φ - φ₀)² / 2κφ₀
-#    "DefaultWithChi": (φ - φ₀)² / 2κφ₀ + Σ χφφ' / φ₀
-# Subclass Hamiltonian to create a new energy functional.
-hamiltonian = 'DefaultNoChi'
-
-
-[field]
-# Particle-mesh grid size, either a single integer or an array of 3 integers
-# (number of grid points in each dimension). In order to guarantee consistency
-# and speed in the PFFT routines, the actual mesh grid will be changed to ensure
-# that each dimension of the 2d PFFT process grid divides each dimension of the
-# mesh grid size.
-mesh_size = [40, 40, 40]
-# Compressibility used in the relaxed incompressibility term of W(phi) in
-# [mol/kJ].
-kappa = 0.05
-# Standard deviation in the Gaussian filter (window function) in [nanometers].
-# This value is a characteristic length scale for the size of the particles in
-# the simulation.
[…]

[…]
