+++ hymd/main.py
@@ -126,7 +126,7 @@ def main():
     if config.barostat_type == "berendsen":
         from .barostat import isotropic, semiisotropic
     elif config.barostat_type == "scr":
-        from .barostat_bussi import isotropic, semiisotropic
+        from .barostat_scr import isotropic, semiisotropic
 
     # dielectric from toml
     if config.coulombtype == "PIC_Spectral_GPE":
