### 2023-03-17 d046464
Modify utils \ Remove scripts that were united into plots.py and hymd_optimise.py Modify outdated scripts
