### 2023-03-17 b561e0b
Remove outdated BO scripts \ These are useful scripts but need modification to work with the current version of HyMD
