### 2023-03-27 #4149 @blakeaw
Hi @rkurchin, it looks like there have been some updates to the package the review process was paused/restarted. For example, it looks like the authors just recently released version 2.0.0 in the github repo; I was working with version 1.0.5 previously. Should I consider these updates in my review? Also, should we see if the authors want to incorporate any of the updates in the software paper?
