### 2023-03-28 #4149 @samiransen23
Hi @blakeaw !
Since there has been a rather long delay in the review process, there have been new releases after 1.0.5, but the paper was written and submitted a while ago, with respect to the version released then. So if you were working with version 1.0.5, the authors think you should continue the process with that version.
I hope @rkurchin finds that reasonable.
