+++ utils/center_group.py
@@ -1,0 +1,141 @@
+import os
+import argparse
+import h5py
+import numpy as np
+import re
+
+
+# based on https://stackoverflow.com/a/6512463/3254658
+def parse_bead_list(string):
+    m = re.match(r"(\d+)(?:-(\d+))?$", string)
+    if not m:
+        raise argparse.ArgumentTypeError(
+            "'"
+            + string
+            + "' is not a range of number. Expected forms like '0-5' or '2'."
+        )
+    start = int(m.group(1), base=10)
+    end = int(m.group(2), base=10) or start
+    if end < start:
+        raise argparse.ArgumentTypeError(
+            f"Start value ({start}) should be larger than final value ({end})"
+        )
+    return list(range(start, end + 1))
+
+
+def get_centers(positions, box):
+    centers = np.empty((0, positions.shape[2]))
+    # based on the position of the first atom get minimal distances
+    for frame in range(positions.shape[0]):
+        deltas = positions[frame, 1:, :] - positions[frame, 0, :]
+        subtract = np.where(deltas > 0.5 * np.diag(box[frame]), True, False)
+        add = np.where(-deltas > 0.5 * np.diag(box[frame]), True, False)
+
+        newpos = np.where(
+            subtract,
+            positions[frame, 1:, :] - np.diag(box[frame]),
+            positions[frame, 1:, :],
+        )
+        newpos = np.where(
+            add, positions[frame, 1:, :] + np.diag(box[frame]), newpos[:, :]
+        )
+        newpos = np.insert(newpos, 0, positions[frame, 0, :], axis=0)
+        centers = np.append(centers, [newpos.mean(axis=0)], axis=0)  # get the centroid
+
+    return centers
+
+
+def recebi(x, y):
+    print("recebi", x, "e", y)
+
+
+def center_trajectory(h5md_file, bead_list, overwrite=False, out_path=None):
+    if out_path is None:
+        out_path = os.path.join(
+            os.path.abspath(os.path.dirname(h5md_file)),
+            os.path.splitext(os.path.split(h5md_file)[-1])[0]
+            + "_new"
+            + os.path.splitext(os.path.split(h5md_file)[-1])[1],
+        )
+    if os.path.exists(out_path) and not overwrite:
+        error_str = (
+            f"The specified output file {out_path} already exists. "
+            f'use overwrite=True ("-f" flag) to overwrite.'
+        )
+        raise FileExistsError(error_str)
+
+    f_in = h5py.File(h5md_file, "r")
+    f_out = h5py.File(out_path, "w")
+
+    for k, v in f_in.attrs.items():
+        f_out.attrs[k] = v
+
+    for k in f_in.keys():
+        f_in.copy(k, f_out)
+
+    box_size = f_in["particles/all/box/edges/value"][:]
+
+    beads_pos = f_in["particles/all/position/value"][:][:, bead_list, :]
+    centers = get_centers(beads_pos, box_size)
+
+    box_diag = np.diag(box_size[0])
+    for frame in range(1, box_size.shape[0]):
+        box_diag = np.vstack((box_diag, np.diag(box_size[frame])))
+    translate = (0.5 * box_diag) - centers
+
+    translations = np.repeat(
+        translate[:, np.newaxis, :],
+        f_in["particles/all/position/value"].shape[1],
+        axis=1,
+    )
+
+    tpos = f_in["particles/all/position/value"] + translations
+    f_out["particles/all/position/value"][:] = np.mod(
+        tpos, np.repeat(box_diag[:, np.newaxis, :], tpos.shape[1], axis=1)
+    )
+
+    f_in.close()
+    f_out.close()
+
+
+if __name__ == "__main__":
+    description = (
+        "Center geometric center of beads in the box for each frame in a .H5 trajectory"
[…]
