### 2023-03-28 Cascella-Group-UiO/HyMD#207 @blakeaw
JOSS Review - Additional community guidelines suggestions
Hi, this issue is part of the JOSS Review at https://github.com/openjournals/joss-reviews/issues/4149.

Regarding the community guidelines:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I see that this item was addressed in https://github.com/Cascella-Group-UiO/HyMD/pull/163 by adding guidelines in the repo's CONTRIBUTING.md. However, I have a couple of follow-up suggestions.

------
First, I'd suggest that you also link to the CONTRIBUTING.md in the docs. For example, a good place might be under https://cascella-group-uio.github.io/HyMD/index.html#development; you could add some text that notes you also welcome contributions from the broader community (including reporting issues/problems) and point folks to the CONTRIBUTING.md for the relevant guidelines.

It might also be helpful to add a `## Contributing` section to the repo's README that points folks to the CONTRIBUTING.md.

------
Second, regarding
> 3) Seek support

Should users also use the repo's issue tracker to request support?  Or, are there other ways to seek support, such as via email or chat groups (Gitter, Discord, Slack, etc.)?

I'd suggest clarifying this as well in the README and docs.
