### 2023-03-28 #4149 @rkurchin
This does sound reasonable. My only concern with this approach is whether documentation or other resources have been updated past this version such that it would be hard for our new reviewer @abb58 to restrict review to the same version. We definitely don't want the two reviewers reviewing different versions of the software, so my inclination is to review the latest version instead. But since this is a somewhat unusual circumstance, I'm going to tag in @openjournals/joss-eics and see what the EIC team thinks about this...
