### 2023-03-29 #4149 @kyleniemeyer
@rkurchin This has definitely happened before with software that is either changing rapidly or a review that takes a while. I believe in the past the JOSS paper focused on the version that had been submitted, or at least the one that was reviewed.

So, in this case, if one reviewer has finished on the original version, then we should probably continue the review of that version, if at all possible. If the documentation has changed, is that version-controlled?
