### 2023-03-29 #4149 @samiransen23
@kyleniemeyer @rkurchin @blakeaw The documentation ( [HylleraasMD documentation](https://cascella-group-uio.github.io/HyMD/index.html) ) is the same as the one for the version being reviewed. We have not changed it, so there should arise no difficulty in continuing the review process. Hope this answers the doubt.
