### 2023-04-04 a322a4d
Fixes error when running with variable dielectric
