### 2023-04-05 698d372
Fix docs build. See https://github.com/readthedocs/readthedocs.org/issues/9038
