+++ .github/workflows/docs_pages.yml
@@ -33,6 +33,7 @@ jobs:
         python3 -m pip install --upgrade pip
         python3 -m pip install --upgrade wheel numpy mpi4py cython
         python3 -m pip install -r requirements.txt
+        python3 -m pip install jinja<3.1
     - name: Install package
       run: |
         python3 -m pip install .
