+++ requirements.txt
@@ -3,7 +3,7 @@ h5py
 mpi4py
 mpsort
 networkx
-numpy
+numpy<1.24
 pfft-python
 pmesh
 sympy
