+++ .github/workflows/ci.yml
@@ -48,6 +48,6 @@ jobs:
         coverage report
         coverage xml
     - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v2
+      uses: codecov/codecov-action@v3
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
