### 2023-04-05 46bd704
Update setup.py and attempts to fix the release action.
