### 2023-04-07 #4149 @blakeaw
Hi @rkurchin, my review is almost complete. I opened an issue regarding the Community Guidelines at https://github.com/Cascella-Group-UiO/HyMD/issues/207 and I'm waiting for the authors to address it before completing that final checklist item.
