### 2023-04-11 b591237
Attempt to bui. version-controlled docs (#56)
* Attempt to version-controlled doc

* Add more versioning docs

* Change sphinx installation line

* Change buildDocs.sh

* Attempt to version docs

* Attempt to version docs

* Attempt to version docs

* Attempt to version docs

* Attempt to version docs

* Attempt to version docs

* Attempt to version docs

* Attempt to version docs
