### 2023-04-12 f5b83cc
Fetch tags before building with sphinx
