+++ docs/conf.py
@@ -115,3 +115,9 @@ html_sidebars = {
 }
 
 numpydoc_show_class_members = False
+
+# whitelists tags and branch to build the docs for
+smv_tag_whitelist = r'^v\d+\.\d+.\d+$' # vX.Y.Z
+
+smv_branch_whitelist = r'^(?!gh-pages).*$' # all branches except gh-pages
+
