### 2023-04-12 f2ac330
Community guidelines info (#214)
* Modify README

* Modify README

* Add commnunity guideline link to documentation.
