+++ README.md
@@ -61,7 +61,6 @@ docker run -it mortele/hymd
 /app$ # Run simulation
 /app$ python3 -m hymd ideal_chain.toml ideal_chain.HDF5 --verbose
 ```
-
 ## Run tests
 Clone the repository and run tests with [pytest](https://docs.pytest.org/en/latest)
 ```bash
@@ -70,11 +69,16 @@ cd hymd
 python3 -m pip install pytest pytest-mpi
 pytest
 ```
+
 Running MPI enabled pytest tests is simplified with a convenient script
 ```bash
 chmod +x pytest-mpi
 pytest-mpi -oo -n 2 -ns
 ```
+
+## Contributions and issues
+We welcome contributions to our code and provide a set of guidelines to follow in [CONTRIBUTING.md](CONTRIBUTING.md).  
+To seek support in case of any issues and bugs, we welcome you to post them using the [issue tracker](https://github.com/Cascella-Group-UiO/HyMD/issues). 
 
 ## Please cite our work
 If you use HyMD for your purposes, please cite the appropriate references from the section below.  
