### 2023-04-13 f9d789e
Resolve rst links in documentation (#215)
* Modify README

* Modify README

* Add commnunity guideline link to documentation.

* Modify index.

* Modify index
