### 2023-04-13 Cascella-Group-UiO/HyMD#207 @samiransen23
Hi @blakeaw !
Sorry for the late reply, and thanks for the suggestions.
We have now mentioned CONTRIBUTING.md in the documentation under section: development, as well as linking it in README.md of the repository.
For seeking support, we prefer the issue tracker, which we have mentioned and linked now in the documentation as well as CONTRIBUTING.md.

The issue should be resolved with commit f9d789e.
