### 2023-04-13 Cascella-Group-UiO/HyMD#207 @blakeaw
Ok, great! Looks good. I'll go ahead and close this issue. I'll let the editor know in the review thread (https://github.com/openjournals/joss-reviews/issues/4149) you have addressed it. Thanks!
