### 2023-04-17 #4149 @mortele
> Some minor editorial remarks:
>
> * line 15: hyphenate "soft-matter" (it's a compound adjective here)
> * 25: hyphenate "open-source" (same reason)
> * 46: should be "requires" (subject is "elucidating")
> * 65: don't hyphenate "translationally invariant" (translationally is already an adverb)
> * 67: pluralize "calculations"
> * 78-79: feels like there's a missing word at the end of this, maybe add "implementation" or similar?
> * 95: remove extra slash at the beginning of URL
>
> Also, one reference issue:
>
> * line 133: capitalization of final name in list

> @mortele see post-review checklist above for a few things to do to finalize this!

Thank you for pointing these errors out. Looking at the post-review checklist items now.
