### 2023-04-17 c68f772
Bump version number for final JOSS release
Also fetch the changes to docs from 2.0.1 to enable version-controlled
documentation.
