+++ docs/conf.py
@@ -48,6 +48,7 @@ extensions = [
     'sphinx.ext.intersphinx',
     'matplotlib.sphinxext.plot_directive',
     'numpydoc',
+    "sphinx_multiversion",
 ]
 
 # Bibtex configuration.
@@ -107,4 +108,16 @@ html_context = {
     ],
 }
 
+html_sidebars = {
+    '**': [
+        'versions.html',
+    ],
+}
+
 numpydoc_show_class_members = False
+
+# whitelists tags and branch to build the docs for
+smv_tag_whitelist = r'^v\d+\.\d+.\d+$' # vX.Y.Z
+
+smv_branch_whitelist = r'^(?!gh-pages).*$' # all branches except gh-pages
+
