+++ docs/docs_requirements.txt
@@ -3,6 +3,8 @@ sphinx-rtd-theme
 sphinx-tabs
 sphinx-panels
 sphinxcontrib.bibtex
+git+https://github.com/Holzhaus/sphinx-multiversion.git
 matplotlib
 numpy
 numpydoc
+jinja2<3.1
