+++ .github/workflows/docs_pages.yml
@@ -1,10 +1,6 @@
 name: docs
 
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
+on: [push]
 
 jobs:
   build_docs_job:
