### 2023-04-17 6c76710
Ignore joss-paper-final branch in versioned docs
Will still show the tagged release, just not this branch that will be
orphaned.
