### 2023-04-18 #4149 @mortele
Final version is *1.0.9*, https://github.com/Cascella-Group-UiO/HyMD/releases/tag/v1.0.9. This version is archived in Zenodo, DOI: `10.5281/zenodo.7839898`. Author list and ORCIDs have been double checked and should match those in the JOSS paper.

> ## Additional Author Tasks After Review is Complete
> * [x]  Double check authors and affiliations (including ORCIDs)
> * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x]  Make sure that the license listed for the archive is the same as the software license.

@rkurchin Apologies for the delay, we had to fix a small issue with the documentation build in the CI before we could release.
