### 2023-04-18 1eea79f
Cleanup center_group.py
