### 2023-04-18 #4149 @rkurchin
Actually, @openjournals/bcm-eics, how should we handle versioning here? Since the review was done on 1.0.5, should there instead be some sort of "special" version number (1.0.5j or something?) that includes only that version plus the review changes? Or is it fine to do as we have now and continue the patch release increment, even though that means there will be un-reviewed features in the version tied to the JOSS publication?
