### 2023-04-21 #4149 @Kevin-Mattheus-Moerman
@mortele I am the AEiC for this track and here to help process final steps. I have checked the archive and the paper and have the below minor comments that require your attention:

On the archive:
- [ ] Please manually edit the ZENODO archive title to match the JOSS paper title.

On the paper:
- [ ] One affiliation says `United States`, this is fine. However, most JOSS papers seem to feature: `United States of America`, so perhaps consider conforming to this too.
- [ ] Check if `backend` should be written instead as `back-end`
