### 2023-04-21 5e42e6e
Add "of America" to Sigbjørn's affiliation
