### 2023-04-21 #4149 @mortele
> @mortele I am the AEiC for this track and here to help process final steps. I have checked the archive and the paper and have the below minor comments that require your attention:
>
> On the archive:
>
> * [x]  Please manually edit the ZENODO archive title to match the JOSS paper title.
>
> On the paper:
>
> * [x]  One affiliation says `United States`, this is fine. However, most JOSS papers seem to feature: `United States of America`, so perhaps consider conforming to this too.
> * [ ]  Check if `backend` should be written instead as `back-end`

Hi @Kevin-Mattheus-Moerman, I have changed the Zenodo archive title to match the paper title and added *of America* to Sigbjørn's affiliation in the archive and in the paper.

As for the last point, I think that *back end* and *backend* are acceptable forms when used as a noun, with *back-end* reserved for usage as a compound adjective. However, I have absolutely no strong feelings about this and would not mind at all changing it if anyone with a stronger background in the English language disagrees.
