### 2023-04-24 2027790
Create CITATION.cff file for the JOSS paper
