### 2023-04-24 88980c8
Add JOSS paper citation to README
