### 2023-04-26 ecc73ec
Improve center_group.py.
