### 2023-04-26 c6a7b9a
Make center_group.py more flexible.
